### 2024-12-06 #7573 @editorialbot
[PRE REVIEW]: FTorch: a library for coupling PyTorch models to Fortran
**Submitting author:** @jatkinson1000 (<a href="http://orcid.org/0000-0001-5001-4812">Jack Atkinson</a>)
**Repository:** https://github.com/Cambridge-ICCS/FTorch
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v1.0.0
