### 2024-12-06 #7573 @matthewfeickert
Thanks @danielskatz.

@jatkinson1000 @danielskatz I would be happy to be the editor for this, but given some hard deadlines for professional things fast approaching I might not be able to start on the process in earnest until late next week. If you would greatly prefer for someone to start the process ASAP though I think I'll need to decline.
