### 2024-12-09 ced9724
docs: add consistent comments to windows CI yml
