### 2024-12-09 a6ec9c2
test: make run_integration_tests.bat similar to linux test
