### 2024-12-09 b176fef
test: make integration tests cross-platform
