### 2024-12-09 ea3272a
test: make integration tests cross-platform
