+++ examples/4_MultiIO/CMakeLists.txt
@@ -26,16 +26,16 @@ if(CMAKE_BUILD_TESTS)
   include(CTest)
 
   # 1. Check the PyTorch model runs and its outputs meet expectations
-  add_test(NAME multiionet COMMAND ${PYTHON_EXECUTABLE}
+  add_test(NAME multiionet COMMAND ${Python_EXECUTABLE}
                                    ${PROJECT_SOURCE_DIR}/multiionet.py)
 
   # 1. Check the model is saved to file in the expected location with the
   #   pt2ts.py script
   add_test(
     NAME pt2ts
-    COMMAND ${PYTHON_EXECUTABLE} ${PROJECT_SOURCE_DIR}/pt2ts.py
-            ${PROJECT_BINARY_DIR}
-    # Command line argument: filepath for saving the model
+    COMMAND ${Python_EXECUTABLE} ${PROJECT_SOURCE_DIR}/pt2ts.py
+            ${PROJECT_BINARY_DIR} # Command line argument: filepath for saving
+                                  # the model
     WORKING_DIRECTORY ${PROJECT_BINARY_DIR})
 
   # 1. Check the model can be loaded from file and run in Python and that its
@@ -43,9 +43,8 @@ if(CMAKE_BUILD_TESTS)
   add_test(
     NAME multiionet_infer_python
     COMMAND
-      ${PYTHON_EXECUTABLE} ${PROJECT_SOURCE_DIR}/multiionet_infer_python.py
-      ${PROJECT_BINARY_DIR}
-    # Command line argument: filepath to find the model
+      ${Python_EXECUTABLE} ${PROJECT_SOURCE_DIR}/multiionet_infer_python.py
+      ${PROJECT_BINARY_DIR} # Command line argument: filepath to find the model
     WORKING_DIRECTORY ${PROJECT_BINARY_DIR})
 
   # 1. Check the model can be loaded from file and run in Fortran and that its
@@ -53,8 +52,7 @@ if(CMAKE_BUILD_TESTS)
   add_test(
     NAME multiionet_infer_fortran
     COMMAND
-      multiionet_infer_fortran
-      ${PROJECT_BINARY_DIR}/saved_multiio_model_cpu.pt
+      multiionet_infer_fortran ${PROJECT_BINARY_DIR}/saved_multiio_model_cpu.pt
       # Command line argument: model file
     WORKING_DIRECTORY ${PROJECT_BINARY_DIR})
   set_tests_properties(
