+++ examples/2_ResNet18/pt2ts.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 """Load a PyTorch model and convert it to TorchScript."""
 
 import os
