+++ examples/1_SimpleNet/simplenet.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 """Module defining a simple PyTorch 'Net' for coupling to Fortran."""
 
 import torch
