+++ examples/6_Autograd/autograd.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 """Autograd demo taken from https://pytorch.org/tutorials/beginner/blitz/autograd_tutorial.html."""
 
 import torch
