### 2024-12-09 966fbd4
test: make integration tests cross-platform
