### 2024-12-09 fab637e
test: make integration tests cross-platform
