+++ examples/2_ResNet18/CMakeLists.txt
@@ -24,12 +24,12 @@ if(CMAKE_BUILD_TESTS)
 
   # 1. Check the PyTorch model runs and its outputs meet expectations
   add_test(NAME resnet18
-    COMMAND ${PYTHON_EXECUTABLE} ${PROJECT_SOURCE_DIR}/resnet18.py
+    COMMAND ${Python_EXECUTABLE} ${PROJECT_SOURCE_DIR}/resnet18.py
     WORKING_DIRECTORY ${PROJECT_SOURCE_DIR})
 
   # 2. Check the model is saved to file in the expected location with the pt2ts.py script
   add_test(NAME pt2ts
-    COMMAND ${PYTHON_EXECUTABLE} ${PROJECT_SOURCE_DIR}/pt2ts.py
+    COMMAND ${Python_EXECUTABLE} ${PROJECT_SOURCE_DIR}/pt2ts.py
     ${PROJECT_BINARY_DIR} # Command line argument: filepath for saving the model
     WORKING_DIRECTORY ${PROJECT_BINARY_DIR})
 
