+++ .github/workflows/test_suite_windows.yml
@@ -69,15 +69,23 @@ jobs:
         shell: cmd
         run: |
           cd src
-          cmake -Bbuild -G "NMake Makefiles" -DCMAKE_Fortran_FLAGS="/fpscomp:logicals" -DCMAKE_PREFIX_PATH="C:\hostedtoolcache\windows\Python\3.12.7\x64\Lib\site-packages" -DCMAKE_BUILD_TYPE=Release -DCMAKE_Fortran_COMPILER=ifx -DCMAKE_C_COMPILER=icx -DCMAKE_CXX_COMPILER=icx -DCMAKE_BUILD_TESTS=TRUE
+          cmake ^
+            -Bbuild ^
+            -G "NMake Makefiles" ^
+            -DCMAKE_Fortran_FLAGS="/fpscomp:logicals" ^
+            -DCMAKE_PREFIX_PATH="C:\hostedtoolcache\windows\Python\3.12.7\x64\Lib\site-packages" ^
+            -DCMAKE_BUILD_TYPE=Release ^
+            -DCMAKE_Fortran_COMPILER=ifx ^
+            -DCMAKE_C_COMPILER=icx ^
+            -DCMAKE_CXX_COMPILER=icx ^
+            -DCMAKE_BUILD_TESTS=TRUE
           cmake --build build
           cmake --install build
 
       - name: Integration tests
         shell: cmd
         run: |
-          cd src
           set PATH=C:\hostedtoolcache\windows\Python\3.12.7\x64\Lib\site-packages;%PATH%
           set PATH=C:\Program Files (x86)\FTorch\bin;%PATH%
           set PATH=C:\hostedtoolcache\windows\Python\3.12.7\x64\Lib\site-packages\torch\lib;%PATH%
-          ..\run_integration_tests.bat
+          run_integration_tests.bat
