### 2024-12-09 5c6e916
chore: fix typos in CI workflows
