+++ .github/workflows/test_suite_ubuntu.yml
@@ -1,5 +1,5 @@
 # Workflow to run the FTorch test suite
-name: TestSuite
+name: TestSuiteUbuntu
 
 # Controls when the workflow will run
 on:
@@ -10,7 +10,7 @@ on:
   # Triggers the workflow on pushes to open pull requests with code changes
   pull_request:
     paths:
-      - '.github/workflows/test_suite.yml'
+      - '.github/workflows/test_suite_ubuntu.yml'
       - '**.c'
       - '**.cpp'
       - '**.fypp'
@@ -33,8 +33,8 @@ concurrency:
 
 # Workflow run - one or more jobs that can run sequentially or in parallel
 jobs:
-  # This workflow contains a single job called "test-suite"
-  test-suite:
+  # This workflow contains a single job called "test-suite-ubuntu"
+  test-suite-ubuntu:
     # The type of runner that the job will run on
     runs-on: ubuntu-latest
     strategy:
