### 2024-12-09 084fe0e
docs: update comments and docs for running windows tests
