### 2024-12-09 a590955
Apply clang-format
