+++ src/ctorch.cpp
@@ -234,88 +234,80 @@ void torch_tensor_delete(torch_tensor_t tensor) {
   delete t;
 }
 
-torch_tensor_t torch_tensor_assign(const torch_tensor_t input)
-{
-    auto in = reinterpret_cast<torch::Tensor* const>(input);
-    torch::AutoGradMode enable_grad(in->requires_grad());
-    torch::Tensor* output = nullptr;
-    output = new torch::Tensor;
-    *output = in->detach().clone();
-    return output;
+torch_tensor_t torch_tensor_assign(const torch_tensor_t input) {
+  auto in = reinterpret_cast<torch::Tensor *const>(input);
+  torch::AutoGradMode enable_grad(in->requires_grad());
+  torch::Tensor *output = nullptr;
+  output = new torch::Tensor;
+  *output = in->detach().clone();
+  return output;
 }
 
 torch_tensor_t torch_tensor_add(const torch_tensor_t tensor1,
-                                const torch_tensor_t tensor2)
-{
-    auto t1 = reinterpret_cast<torch::Tensor* const>(tensor1);
-    auto t2 = reinterpret_cast<torch::Tensor* const>(tensor2);
-    torch::Tensor* output = nullptr;
-    output = new torch::Tensor;
-    *output = *t1 + *t2;
-    return output;
+                                const torch_tensor_t tensor2) {
+  auto t1 = reinterpret_cast<torch::Tensor *const>(tensor1);
+  auto t2 = reinterpret_cast<torch::Tensor *const>(tensor2);
+  torch::Tensor *output = nullptr;
+  output = new torch::Tensor;
+  *output = *t1 + *t2;
+  return output;
 }
 
 torch_tensor_t torch_tensor_subtract(const torch_tensor_t tensor1,
-                                     const torch_tensor_t tensor2)
-{
-    auto t1 = reinterpret_cast<torch::Tensor* const>(tensor1);
-    auto t2 = reinterpret_cast<torch::Tensor* const>(tensor2);
-    torch::Tensor* output = nullptr;
-    output = new torch::Tensor;
-    *output = *t1 - *t2;
-    return output;
+                                     const torch_tensor_t tensor2) {
+  auto t1 = reinterpret_cast<torch::Tensor *const>(tensor1);
+  auto t2 = reinterpret_cast<torch::Tensor *const>(tensor2);
+  torch::Tensor *output = nullptr;
+  output = new torch::Tensor;
+  *output = *t1 - *t2;
+  return output;
 }
 
 torch_tensor_t torch_tensor_multiply(const torch_tensor_t tensor1,
-                                     const torch_tensor_t tensor2)
-{
-    auto t1 = reinterpret_cast<torch::Tensor* const>(tensor1);
-    auto t2 = reinterpret_cast<torch::Tensor* const>(tensor2);
-    torch::Tensor* output = nullptr;
-    output = new torch::Tensor;
-    *output = *t1 * *t2;
-    return output;
+                                     const torch_tensor_t tensor2) {
+  auto t1 = reinterpret_cast<torch::Tensor *const>(tensor1);
+  auto t2 = reinterpret_cast<torch::Tensor *const>(tensor2);
+  torch::Tensor *output = nullptr;
+  output = new torch::Tensor;
+  *output = *t1 * *t2;
+  return output;
 }
 
 torch_tensor_t torch_tensor_premultiply(const torch_data_t scalar,
-                                        const torch_tensor_t tensor)
-{
-    auto t = reinterpret_cast<torch::Tensor* const>(tensor);
-    torch::Tensor* output = nullptr;
-    output = new torch::Tensor;
-    *output = scalar * *t;
-    return output;
+                                        const torch_tensor_t tensor) {
[…]
