### 2024-12-09 c12d5c3
Apply clang-format to header
