### 2024-12-09 5cdfe22
docs: add comment about fypp dependency
