### 2024-12-10 7620d51
Add CMake formatter to StaticAnalysis
