### 2024-12-10 9c7202d
Use cmake-lint rather than cmake-format
