### 2024-12-10 51bf8e6
Move CMake lint check to be first
