### 2024-12-10 8618325
Apply cmake-format
