+++ examples/6_Autograd/CMakeLists.txt
@@ -1,6 +1,7 @@
 cmake_minimum_required(VERSION 3.5 FATAL_ERROR)
-#policy CMP0076 - target_sources source files are relative to file where target_sources is run
-cmake_policy (SET CMP0076 NEW)
+# policy CMP0076 - target_sources source files are relative to file where
+# target_sources is run
+cmake_policy(SET CMP0076 NEW)
 
 set(PROJECT_NAME AutogradExample)
 
@@ -8,7 +9,9 @@ project(${PROJECT_NAME} LANGUAGES Fortran)
 
 # Build in Debug mode if not specified
 if(NOT CMAKE_BUILD_TYPE)
-  set(CMAKE_BUILD_TYPE Debug CACHE STRING "" FORCE)
+  set(CMAKE_BUILD_TYPE
+      Debug
+      CACHE STRING "" FORCE)
 endif()
 
 find_package(FTorch)
@@ -23,13 +26,14 @@ if(CMAKE_BUILD_TESTS)
   include(CTest)
 
   # 1. Check the Python Autograd script runs successfully
-  add_test(NAME pyautograd
-    COMMAND ${PYTHON_EXECUTABLE} ${PROJECT_SOURCE_DIR}/autograd.py)
+  add_test(NAME pyautograd COMMAND ${PYTHON_EXECUTABLE}
+                                   ${PROJECT_SOURCE_DIR}/autograd.py)
 
-  # 2. Check the Fortran Autograd script runs successfully
-  add_test(NAME fautograd
+  # 1. Check the Fortran Autograd script runs successfully
+  add_test(
+    NAME fautograd
     COMMAND autograd
     WORKING_DIRECTORY ${PROJECT_BINARY_DIR})
   set_tests_properties(fautograd PROPERTIES PASS_REGULAR_EXPRESSION
-    "Autograd example ran successfully")
+                                            "Autograd example ran successfully")
 endif()
