### 2024-12-10 53c334f
Manually fix lint errors and add exclusions
