+++ src/CMakeLists.txt
@@ -50,9 +50,11 @@ find_package(Torch REQUIRED)
 add_library(${LIB_NAME} SHARED ctorch.cpp ftorch.f90 ftorch_test_utils.f90)
 # Add an alias FTorch::ftorch for the library
 add_library(${PROJECT_NAME}::${LIB_NAME} ALIAS ${LIB_NAME})
+# cmake-format: off
 set_target_properties(
-  ${LIB_NAME} PROPERTIES PUBLIC_HEADER "ctorch.h" Fortran_MODULE_DIRECTORY
-                                                  "${CMAKE_BINARY_DIR}/modules")
+  ${LIB_NAME} PROPERTIES PUBLIC_HEADER "ctorch.h"
+  Fortran_MODULE_DIRECTORY "${CMAKE_BINARY_DIR}/modules")
+# cmake-format: on
 # Link TorchScript
 target_link_libraries(${LIB_NAME} PRIVATE ${TORCH_LIBRARIES})
 # Include the Fortran mod files in the library
