### 2024-12-10 4834763
Manual fixes for examples CMakeLists
