### 2024-12-10 e0b805c
chore: bump cmake minimum version
