### 2024-12-10 5af255f
chore: bump cmake minimum version
