+++ examples/2_ResNet18/CMakeLists.txt
@@ -1,4 +1,4 @@
-cmake_minimum_required(VERSION 3.5 FATAL_ERROR)
+cmake_minimum_required(VERSION 3.15 FATAL_ERROR)
 # policy CMP0076 - target_sources source files are relative to file where
 # target_sources is run
 cmake_policy(SET CMP0076 NEW)
