+++ requirements.txt
@@ -2,3 +2,4 @@ ruff==0.7.3
 fortitude-lint
 clang-format==19.1.3
 clang-tidy==19.1.0
+cmakelang
