+++ .github/workflows/static_analysis.yml
@@ -68,7 +68,6 @@ jobs:
         run: |
           cd ${{ github.workspace }}
           . ftorch_venv/bin/activate
-          pip install cmakelang
           cmake-lint $(find . -name CMakeLists.txt)
 
       # Apply Fortran linter, fortitude
