### 2024-12-10 d359b19
Apply CMake formatting (#204)
* Add CMake formatter to StaticAnalysis
* Apply cmake-format
* Manually fix lint errors and add exclusions
* Manual fixes for examples CMakeLists
* Add cmakelang to requirements
* Add cmake-lint to developer docs
