+++ src/CMakeLists.txt
@@ -3,7 +3,10 @@ set(PROJECT_NAME FTorch)
 set(LIB_NAME ftorch)
 set(PACKAGE_VERSION 0.1)
 
-project(${PROJECT_NAME} VERSION ${PACKAGE_VERSION} LANGUAGES C CXX Fortran)
+project(
+  ${PROJECT_NAME}
+  VERSION ${PACKAGE_VERSION}
+  LANGUAGES C CXX Fortran)
 
 include(FortranCInterface)
 FortranCInterface_VERIFY(CXX QUIET)
@@ -26,19 +29,18 @@ endif()
 set(CMAKE_SKIP_RPATH FALSE)
 set(CMAKE_SKIP_BUILD_RPATH FALSE)
 set(CMAKE_BUILD_WITH_INSTALL_RPATH FALSE)
-# Embed absolute paths to external libraries that are not part of
-# the project, (they are expected to be at the same location on all
-# machines the project will be deployed to
+# Embed absolute paths to external libraries that are not part of the project,
+# (they are expected to be at the same location on all machines the project will
+# be deployed to
 set(CMAKE_INSTALL_RPATH_USE_LINK_PATH TRUE)
 
 # Follow GNU conventions for installing directories
 include(GNUInstallDirs)
 
-# Define RPATH for executables via a relative expression to enable a
-# fully relocatable package
-file(RELATIVE_PATH relDir
- ${CMAKE_CURRENT_BINARY_DIR}/${CMAKE_INSTALL_BINDIR}
- ${CMAKE_CURRENT_BINARY_DIR}/${CMAKE_INSTALL_LIBDIR})
+# Define RPATH for executables via a relative expression to enable a fully
+# relocatable package
+file(RELATIVE_PATH relDir ${CMAKE_CURRENT_BINARY_DIR}/${CMAKE_INSTALL_BINDIR}
+     ${CMAKE_CURRENT_BINARY_DIR}/${CMAKE_INSTALL_LIBDIR})
 set(CMAKE_INSTALL_RPATH $ORIGIN/${relDir})
 
 # Dependencies
@@ -48,66 +50,58 @@ find_package(Torch REQUIRED)
 add_library(${LIB_NAME} SHARED ctorch.cpp ftorch.f90 ftorch_test_utils.f90)
 # Add an alias FTorch::ftorch for the library
 add_library(${PROJECT_NAME}::${LIB_NAME} ALIAS ${LIB_NAME})
-set_target_properties(${LIB_NAME} PROPERTIES
-  PUBLIC_HEADER "ctorch.h"
-  Fortran_MODULE_DIRECTORY "${CMAKE_BINARY_DIR}/modules"
-  )
+# cmake-format: off
+set_target_properties(
+  ${LIB_NAME} PROPERTIES PUBLIC_HEADER "ctorch.h"
+  Fortran_MODULE_DIRECTORY "${CMAKE_BINARY_DIR}/modules")
+# cmake-format: on
 # Link TorchScript
 target_link_libraries(${LIB_NAME} PRIVATE ${TORCH_LIBRARIES})
 # Include the Fortran mod files in the library
-target_include_directories(${LIB_NAME}
-  PUBLIC
-    $<BUILD_INTERFACE:${CMAKE_BINARY_DIR}/modules>
-#    $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>
-    )
[…]
