### 2024-12-10 db3c6f5
docs: add example of how to build and run in windows
