+++ pages/troubleshooting.md
@@ -6,8 +6,8 @@ If you are experiencing problems building or using FTorch please see below for g
 
 ## Windows
 
-If possible we recommend using the [Windows Subsystem for Linux](https://learn.microsoft.com/en-us/windows/wsl/) (WSL) to build the library.
-In this case the build process is the same as for a Linux environment.
+If possible we recommend using the [Windows Subsystem for Linux](https://learn.microsoft.com/en-us/windows/wsl/) (WSL) to build
+the library. In this case the build process is the same as for a Linux environment.
 
 If you need to build in native Windows please read the following information:
 
@@ -19,26 +19,84 @@ It is possible to build using Visual Studio and the Intel Fortran Compiler. In t
 * [Intel OneAPI Basetoolkit](https://www.intel.com/content/www/us/en/developer/tools/oneapi/base-toolkit-download.html)
 * [Intel OneAPI HPC toolkit](https://www.intel.com/content/www/us/en/developer/tools/oneapi/hpc-toolkit.html) ensuring that the Intel Fortran compiler and VS integration is selected.
 
-You will then need to load the intel Fortran compilers using `setvars.bat`
-which is found in the Intel compiler install directory (see the 
-[intel docs](https://www.intel.com/content/www/us/en/docs/oneapi/programming-guide/2023-2/use-the-setvars-script-with-windows.html))
+You will then need to load the intel Fortran compilers using `setvars.bat` which is found in the Intel compiler install
+directory (see the [intel
+docs](https://www.intel.com/content/www/us/en/docs/oneapi/programming-guide/2023-2/use-the-setvars-script-with-windows.html))
 for more details.<br>
-From CMD this can be done with:
+
+From `cmd` this can be done with:
 ```
-"C:\Program Files (x86)\Intel\oneAPI\setvars.bat"
+call "C:\Program Files (x86)\Intel\oneAPI\setvars.bat"
 ```
 
 Finally you will need to add `-G "NMake Makefiles"` to the `cmake` command in the
 [regular install instructions](doc/page/cmake.html).<br>
-So the basic command to build from CMD becomes:
+So the basic command to build from `cmd` becomes:
 ```
 cmake -G "NMake Makefiles" -DCMAKE_PREFIX_PATH="C:\Users\<path-to-libtorch-download>\libtorch" -DCMAKE_BUILD_TYPE=Release ..
 cmake --build .
 cmake --install .
 ```
 
-We would recommend Windows users to review the Windows CI workflow (`.github/workflows/test_suite_windows.yml`) for more
-information, as this provides an example of how to build and run FTorch and the tests from scratch.
+The following is an example `cmd` script that installs FTorch and runs the integration tests. It assumes you have already
+install `cmake`, `git`, the intel compilers and visual studio.
+
+```cmd
+rem disable output for now
+ECHO ON
+
+rem load intel compilers
+call "C:\Program Files (x86)\Intel\oneAPI\setvars.bat"
+
+rem download ftorch
+git clone https://github.com/Cambridge-ICCS/FTorch.git
+cd FTorch
+
+rem make venv
+python -m venv .ftorch
+
+rem activate the environment
+call .ftorch\Scripts\activate
+
+rem install torch
+pip install torch torchvision torchaudio
+
+rem enable output
+ECHO ON
+
+rem run cmake to generate build scripts
+rem (update CMAKE_PREFIX_PATH depending on location of ftorch venv)
+cd src
[…]
