### 2024-12-10 cf56bc3
Pin fortitude version
