+++ requirements.txt
@@ -1,5 +1,5 @@
 ruff==0.7.3
-fortitude-lint
+fortitude-lint==0.6.0
 clang-format==19.1.3
 clang-tidy==19.1.0
 cmakelang
