+++ src/ftorch.F90
@@ -14,7 +14,7 @@ module ftorch
 
   implicit none
 
-  integer, parameter :: ftorch_int = int32 ! set integer size for FTorch library
+  integer, parameter :: ftorch_int = int32  ! set integer size for FTorch library
 
   !> Type for holding a torch neural net (nn.Module).
   type torch_model
@@ -33,12 +33,12 @@ module ftorch
   !  From c_torch.h (torch_data_t)
   !  Note that 0 `torch_kUInt8` and 5 `torch_kFloat16` are not sypported in Fortran
   enum, bind(c)
-    enumerator :: torch_kUInt8 = 0 ! not supported in Fortran
+    enumerator :: torch_kUInt8 = 0  ! not supported in Fortran
     enumerator :: torch_kInt8 = 1
     enumerator :: torch_kInt16 = 2
     enumerator :: torch_kInt32 = 3
     enumerator :: torch_kInt64 = 4
-    enumerator :: torch_kFloat16 = 5 ! not supported in Fortran
+    enumerator :: torch_kFloat16 = 5  ! not supported in Fortran
     enumerator :: torch_kFloat32 = 6
     enumerator :: torch_kFloat64 = 7
   end enum
@@ -172,9 +172,9 @@ contains
     integer(c_int64_t), intent(in)  :: tensor_shape(*)   !! Shape of the tensor
     integer(c_int), intent(in)      :: dtype      !! Data type of the tensor
     integer(c_int), intent(in)      :: device_type  !! Device type the tensor will live on (`torch_kCPU` or `torch_kCUDA`)
-    integer(c_int), optional, intent(in) :: device_index     !! device index to use for `torch_kCUDA` case
+    integer(c_int), optional, intent(in) :: device_index    !! device index to use for `torch_kCUDA` case
     logical(c_bool), optional, intent(in) :: requires_grad  !! Whether gradients need to be computed for the created tensor
-    integer(c_int)                  :: device_index_value  !! device index used
+    integer(c_int)                  :: device_index_value   !! device index used
     logical(c_bool)                 :: requires_grad_value  !! Whether gradients need to be computed for the created tensor
 
     interface
@@ -221,12 +221,12 @@ contains
     type(torch_tensor), intent(out) :: tensor     !! Returned tensor
     integer(c_int), intent(in)      :: ndims      !! Number of dimensions of the tensor
     integer(c_int64_t), intent(in)  :: tensor_shape(*)   !! Shape of the tensor
-    integer(c_int), intent(in)      :: dtype      !! Data type of the tensor
+    integer(c_int), intent(in)      :: dtype        !! Data type of the tensor
     integer(c_int), intent(in)      :: device_type  !! Device type the tensor will live on (`torch_kCPU` or `torch_kCUDA`)
     integer(c_int), optional, intent(in) :: device_index     !! device index to use for `torch_kCUDA` case
-    logical(c_bool), optional, intent(in) :: requires_grad  !! Whether gradients need to be computed for the created tensor
-    integer(c_int)                  :: device_index_value  !! device index used
-    logical(c_bool)                 :: requires_grad_value  !! Whether gradients need to be computed for the created tensor
+    logical(c_bool), optional, intent(in) :: requires_grad   !! Whether gradients need to be computed for the created tensor
+    integer(c_int)                  :: device_index_value    !! device index used
+    logical(c_bool)                 :: requires_grad_value   !! Whether gradients need to be computed for the created tensor
 
     interface
       function torch_ones_c(ndims, tensor_shape, dtype, &
@@ -275,16 +275,16 @@ contains
     type(torch_tensor), intent(out) :: tensor     !! Returned tensor
     type(c_ptr), intent(in)         :: data       !! Pointer to data
[…]
