### 2024-12-10 5125881
Update test_utils to satisfy latest fortitude.
