### 2024-12-10 e82816e
chore: commit suggestions from code review
Co-authored-by: Jack Atkinson <109271713+jatkinson1000@users.noreply.github.com>
