### 2024-12-10 b8ccc90
docs: update cmake version in docs
