+++ src/ftorch.F90
@@ -1089,10 +1089,10 @@ contains
     logical, optional, intent(in)        :: requires_grad  !! Whether gradients need to be computed for the created tensor
 
     ! local data
-    integer(c_int64_t)        :: c_tensor_shape(1)            !! Shape of the tensor
-    integer(c_int), parameter :: c_dtype = torch_kInt8  !! Data type
-    integer(c_int64_t)        :: strides(1)                   !! Strides for accessing data
-    integer(c_int), parameter :: ndims = 1                    !! Number of dimension of input data
+    integer(c_int64_t)        :: tensor_shape(1)            !! Shape of the tensor
+    integer(c_int), parameter :: dtype = torch_kInt8  !! Data type
+    integer(c_int64_t)        :: strides(1)                 !! Strides for accessing data
+    integer(c_int), parameter :: ndims = 1                  !! Number of dimension of input data
     integer(ftorch_int)       :: i
     integer(c_int)            :: device_index_value
     logical :: requires_grad_value  !! Whether gradients need to be computed for the created tensor
@@ -1143,10 +1143,10 @@ contains
     logical, optional, intent(in)        :: requires_grad  !! Whether gradients need to be computed for the created tensor
 
     ! local data
-    integer(c_int64_t)        :: c_tensor_shape(2)            !! Shape of the tensor
-    integer(c_int), parameter :: c_dtype = torch_kInt8  !! Data type
-    integer(c_int64_t)        :: strides(2)                   !! Strides for accessing data
-    integer(c_int), parameter :: ndims = 2                    !! Number of dimension of input data
+    integer(c_int64_t)        :: tensor_shape(2)            !! Shape of the tensor
+    integer(c_int), parameter :: dtype = torch_kInt8  !! Data type
+    integer(c_int64_t)        :: strides(2)                 !! Strides for accessing data
+    integer(c_int), parameter :: ndims = 2                  !! Number of dimension of input data
     integer(ftorch_int)       :: i
     integer(c_int)            :: device_index_value
     logical :: requires_grad_value  !! Whether gradients need to be computed for the created tensor
@@ -1197,10 +1197,10 @@ contains
     logical, optional, intent(in)        :: requires_grad  !! Whether gradients need to be computed for the created tensor
 
     ! local data
-    integer(c_int64_t)        :: c_tensor_shape(3)            !! Shape of the tensor
-    integer(c_int), parameter :: c_dtype = torch_kInt8  !! Data type
-    integer(c_int64_t)        :: strides(3)                   !! Strides for accessing data
-    integer(c_int), parameter :: ndims = 3                    !! Number of dimension of input data
+    integer(c_int64_t)        :: tensor_shape(3)            !! Shape of the tensor
+    integer(c_int), parameter :: dtype = torch_kInt8  !! Data type
+    integer(c_int64_t)        :: strides(3)                 !! Strides for accessing data
+    integer(c_int), parameter :: ndims = 3                  !! Number of dimension of input data
     integer(ftorch_int)       :: i
     integer(c_int)            :: device_index_value
     logical :: requires_grad_value  !! Whether gradients need to be computed for the created tensor
@@ -1251,10 +1251,10 @@ contains
     logical, optional, intent(in)        :: requires_grad  !! Whether gradients need to be computed for the created tensor
 
     ! local data
-    integer(c_int64_t)        :: c_tensor_shape(4)            !! Shape of the tensor
-    integer(c_int), parameter :: c_dtype = torch_kInt8  !! Data type
-    integer(c_int64_t)        :: strides(4)                   !! Strides for accessing data
-    integer(c_int), parameter :: ndims = 4                    !! Number of dimension of input data
[…]
