+++ src/CMakeLists.txt
@@ -115,11 +115,11 @@ if(CMAKE_BUILD_TESTS)
   file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/../examples/1_SimpleNet
        DESTINATION ${CMAKE_CURRENT_SOURCE_DIR}/test/examples)
   file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/../examples/2_ResNet18
+       DESTINATION ${CMAKE_CURRENT_SOURCE_DIR}/test/examples)
   if(ENABLE_CUDA)
     file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/../examples/3_MultiGPU
          DESTINATION ${CMAKE_CURRENT_SOURCE_DIR}/test/examples)
   endif()
-       DESTINATION ${CMAKE_CURRENT_SOURCE_DIR}/test/examples)
   file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/../examples/4_MultiIO
        DESTINATION ${CMAKE_CURRENT_SOURCE_DIR}/test/examples)
   # file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/../examples/5_Looping
