### 2024-12-10 664be32
Tweaks to test scripts
