### 2024-12-10 1f41cd1
Tweaks to test scripts
