### 2024-12-16 c49da28
Accept command line inputs for run_test_suite.sh
