### 2024-12-16 db4de15
Accept command line inputs for run_test_suite.sh
