### 2024-12-16 00d82eb
Setup shell lint action; fix CMake lint
