+++ .github/workflows/static_analysis.yml
@@ -18,6 +18,8 @@ on:
       - '**.F90'
       - '**.pf'
       - '**.py'
+      - '**.sh'
+      - '**CMakeLists.txt'
 
   # Allows you to run this workflow manually from the Actions tab
   workflow_dispatch:
@@ -63,12 +65,22 @@ jobs:
           cmake .. -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=${BUILD_DIR} -DCMAKE_Fortran_FLAGS="-std=f2008" -DCMAKE_EXPORT_COMPILE_COMMANDS=ON
 
       # Apply CMake linter, cmake-lint
-      - name: cmake
+      - name: cmake-lint
         if: always()
         run: |
           cd ${{ github.workspace }}
           . ftorch_venv/bin/activate
           cmake-lint $(find . -name CMakeLists.txt)
+
+      # Apply Shell linter, shellcheck
+      - name: shellcheck
+        if: always()
+        run: |
+          cd ${{ github.workspace }}
+          sudo apt install shellcheck
+          for FILE in $(find . -name "*.sh"); do
+            shellcheck ${FILE}
+          done
 
       # Apply Fortran linter, fortitude
       # Configurable using the fortitude.toml file if present
