### 2024-12-16 0332b09
Apply shellcheck
