+++ run_integration_tests.sh
@@ -11,13 +11,12 @@
 
 set -eu
 
-CTEST_ARGS=$@
 EXAMPLES="1_SimpleNet 2_ResNet18 4_MultiIO 6_Autograd"
 BUILD_DIR=src/build
 
 for EXAMPLE in ${EXAMPLES}; do
-  pip -q install -r examples/${EXAMPLE}/requirements.txt
-  cd ${BUILD_DIR}/test/examples/${EXAMPLE}
-  ctest ${CTEST_ARGS}
+  pip -q install -r examples/"${EXAMPLE}"/requirements.txt
+  cd "${BUILD_DIR}"/test/examples/"${EXAMPLE}"
+  ctest "$@"
   cd -
 done
