### 2024-12-16 d8e7e61
Setup Actions linter
