+++ requirements.txt
@@ -3,3 +3,4 @@ fortitude-lint==0.6.0
 clang-format==19.1.3
 clang-tidy==19.1.0
 cmakelang
+zizmor
