### 2024-12-16 e128fe3
Apply lint fixes for static_analysis workflow
