### 2024-12-16 639cded
Run Windows workflow when .bat files are changed, not .sh files
