+++ .github/workflows/test_suite_windows.yml
@@ -11,6 +11,7 @@ on:
   pull_request:
     paths:
       - '.github/workflows/test_suite_windows.yml'
+      - '**.bat'
       - '**.c'
       - '**.cpp'
       - '**.fypp'
@@ -18,7 +19,6 @@ on:
       - '**.F90'
       - '**.pf'
       - '**.py'
-      - '**.sh'
       - '**CMakeLists.txt'
       - '**requirements.txt'
       - '**data/*'
