### 2024-12-16 c87fd97
test: make windows CI more robust
Previously I hard-coded the path to libtorch. This commit replaces the
hard-code path using `pip show` to get the updated location.
