### 2024-12-16 13e12e2
chore: rename variable to torch_path
