### 2024-12-18 f0ed978
Reformulate autograd example to test multiply and divide
