+++ src/ctorch.h
@@ -180,7 +180,7 @@ EXPORT_C torch_tensor_t torch_tensor_premultiply(const torch_data_t scalar,
                                                  const torch_tensor_t tensor);
 
 /**
- * Overloads the postmultiplication operator for a scalar and a Torch Tensor
+ * Overloads the postmultiplication operator for a Torch Tensor and a scalar
  * @param Tensor to be multiplied
  * @param scalar to multiply by
  * @return product of the Tensor and scalar
@@ -196,6 +196,15 @@ EXPORT_C torch_tensor_t torch_tensor_postmultiply(const torch_tensor_t tensor,
  */
 EXPORT_C torch_tensor_t torch_tensor_divide(const torch_tensor_t tensor1,
                                             const torch_tensor_t tensor2);
+
+/**
+ * Overloads the post-division operator for a Torch Tensor and a scalar
+ * @param Tensor to be divided
+ * @param scalar to divide by
+ * @return quotient of the Tensor and scalar
+ */
+EXPORT_C torch_tensor_t torch_tensor_postdivide(const torch_tensor_t tensor,
+                                                const torch_data_t scalar);
 
 /**
  * Overloads the exponentiation operator for two Torch Tensors
