### 2024-12-18 f1e9add
Don't build unit tests on Windows
