### 2024-12-18 1e852b4
Update docs on autograd; add reference to looping example
