### 2024-12-18 23d4a76
Fix test script name
