### 2024-12-18 4be904c
Fix test script name
