### 2024-12-18 2834064
Pin zizmor version; update dev docs for linters
