+++ pages/developer.md
@@ -106,9 +106,17 @@ The tools we use are as follows on a language-by-language basis:
 * C++: [clang-format](https://clang.llvm.org/docs/ClangFormat.html) and [clang-tidy](https://clang.llvm.org/extra/clang-tidy/)
 * C: [clang-format](https://clang.llvm.org/docs/ClangFormat.html) and [clang-tidy](https://clang.llvm.org/extra/clang-tidy/)
 * Python: [ruff](https://docs.astral.sh/ruff/)
+* Shell: [ShellCheck](https://github.com/koalaman/shellcheck)
 * CMake: [cmake-format](https://github.com/cheshirekow/cmake_format)
+* GitHub Actions workflows: [zizmor](https://woodruffw.github.io/zizmor)
 
 Instructions on installing these tools can be found in their respective documentations.
+Note that all but ShellCheck may be installed with pip. A shortcut for doing
+this is to run the following from the base FTorch directory:
+```
+pip install -r requirements.txt
+```
+
 Contributors should run them over their code and ensure that it conforms before submitting
 a pull request. If there is a good reason to ignore a particular rule this should be
 justified in the pull request and ideally documented in the code.
