### 2024-12-18 11dcd03
Linting for shell and GitHub Actions (#206)
* Setup shell lint action
* Fix CMake lint
* Apply shellcheck
* Setup Actions linter
* Apply lint fixes for workflows
* Run Windows workflow when .bat files are changed, not .sh files
* Pin zizmor version
* Update dev docs for linters
