### 2024-12-18 #7602 @jatkinson1000
Thanks all for agreeing to edit and review.
I look forward to your comments.

Just a note to say - I created the JOSS paper in a separate branch, but as an active piece of software there will likely be additions to main since this was created. I'm not sure how well the JOSS system handles this, but I'm happy to keep the JOSS branch up to date by rebasing if you wish.
