### 2024-12-18 #7602 @matthewfeickert
> I created the JOSS paper in a separate branch, but as an active piece of software there will likely be additions to main since this was created. I'm not sure how well the JOSS system handles this, but I'm happy to keep the JOSS branch up to date by rebasing if you wish.

That's totally fine, @jatkinson1000. Any edits can happen on the paper branch and then we'll have `editorialbot ` rerender as needed.
