### 2024-12-18 e50915c
Add macos test suite
