### 2024-12-18 #7602 @matthewfeickert
As mentioned in https://github.com/openjournals/joss-reviews/issues/7573#issuecomment-2551792389, we have had some additional potential reviewers respond to agree to also contribute reviews, and I like to accept 3 reviewers when possible. So @anand-me will be joining as a reviewer (thanks!) though has similar time constraint as @parikshitbajpai.

@anand-me, please check out https://github.com/openjournals/joss-reviews/issues/7602#issuecomment-2551875338 for additional information, and if you have any questions please let me know.
