### 2024-12-19 8dfa92c
Revert adding example 3 to build
