+++ .github/workflows/fypp_checks.yml
@@ -30,7 +30,7 @@ jobs:
         python-version: "3.11"
     - run: pip install fypp
 
-    - name: Check ftorch.fypp matches ftorch.f90
+    - name: Check ftorch.fypp matches ftorch.F90
       run: |
         fypp src/ftorch.fypp src/temp.F90_temp
         if ! diff -q src/ftorch.F90 src/temp.F90_temp; then
