### 2024-12-20 4c2fc90
Update ftorch.F90 for XPU support
