+++ src/ftorch.F90
@@ -49,6 +49,8 @@ module ftorch
   enum, bind(c)
     enumerator :: torch_kCPU = 0
     enumerator :: torch_kCUDA = 1
+    enumerator :: torch_kMPS = 2
+    enumerator :: torch_kXPU = 3
   end enum
 
   !> Interface for directing `torch_tensor_from_array` to possible input types and ranks
