### 2024-12-20 2a30f6a
Update ftorch.F90 for XPU support
