### 2024-12-20 34316e0
Add xpu python modifications to examples 1 and 2
