+++ examples/2_ResNet18/pt2ts.py
@@ -105,10 +105,10 @@ if __name__ == "__main__":
 
     # FPTLIB-TODO
     # Uncomment the following lines to save for inference on GPU (rather than CPU):
-    # device = torch.device('cuda')
-    # trained_model = trained_model.to(device)
-    # trained_model.eval()
-    # trained_model_dummy_input = trained_model_dummy_input.to(device)
+    device = torch.device('xpu')
+    trained_model = trained_model.to(device)
+    trained_model.eval()
+    trained_model_dummy_input = trained_model_dummy_input.to(device)
 
     # FPTLIB-TODO
     # Run model for dummy inputs
