### 2024-12-23 1376dcb
Overload elementary operations (#139)
* Overload =,+,*; consistent variable names
* Update existing examples
* Add autograd example
* Overload - and test
* Overload / and test
* Overload ** and test
* Overload scalar premultiply and test
* Overload scalar postmultiply and test
* Implement torch_to_blob
* Implement postdivide
* Point to Torch C++ API
* Update docs on autograd; add reference to looping example
* Write as LibTorch in pages/autograd.md

Co-authored-by: Jack Atkinson <109271713+jatkinson1000@users.noreply.github.com>

* Use better links for Torch C++ docs

---------

Co-authored-by
[…]
