+++ run_integration_tests.sh
@@ -11,7 +11,12 @@
 
 set -eu
 
-EXAMPLES="1_SimpleNet 2_ResNet18 4_MultiIO 6_Autograd"
+EXAMPLES="
+  1_SimpleNet
+  2_ResNet18
+  4_MultiIO
+  6_Autograd
+"
 BUILD_DIR=src/build
 
 for EXAMPLE in ${EXAMPLES}; do
