### 2025-01-02 Cambridge-ICCS/FTorch#212 @timothyas
[JOSS Review] JOSS Paper Comments
Hi there, really great work on this package and JOSS paper. The paper is very well written and well motivated, I just have two comments for you.

1. I know the focus of this work is on pytorch rather than Keras, but I think that given the overlap in content, it would be worth mentioning the Fortran Keras Bridge (here's the [github repo](https://github.com/scientific-computing/FKB) and [arxiv paper](https://arxiv.org/abs/2004.10652)).
2. Tiny typo: line 102 should say "trained" not "trainied".

Otherwise the paper looks great!

cc: openjournals/joss-reviews#7602 @matthewfeickert
