### 2025-01-03 Cambridge-ICCS/FTorch#214 @timothyas
[JOSS Review] Install + Example Docs Comments
Hi everyone, nice work on the documentation! I was able to install FTorch successfully, thanks to all the helpful cmake descriptors - those were really essential. I have only had one hiccup running through the examples, noted below, and I skipped the multi gpu example since I was testing this out on a machine with a single GPU. Otherwise, I was able to very easily fly through the docs, reproduce the results, and reproduce the (approximate) timings noted in example 5 (looping).

During the 2nd example (`python resnet18.py`) I get the following error:

```python
Traceback (most recent call last):
  File "/home/tsmith/work/FTorch/examples/2_ResNet18/resnet18.py", line 5, in <module>
    import torchvision
  File "/home/tsmith/miniconda3/envs/ftorch/lib/python3.11/site-packages/torchvision/__init__.py", line 10, in <module>
    from torchvision import _meta_registrations, datasets, io, models, ops, transforms, utils  # usort:skip
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/tsmith/miniconda3/envs/ftorch/lib/python3.11/site-packages/torchvision/datasets/__init__.py", line 1, in <module>
    from ._optical_flow import FlyingChairs, FlyingThings3D, HD1K, KittiFlow, Sintel
  File "/home/tsmith/miniconda3/envs/ftorch/lib/python3.11/site-packages/torchvision/datasets/_optical_flow.py", line 10, in <module>
    from PIL import Image
  File "/home/tsmith/miniconda3/envs/ftorch/lib/python3.11/site-packages/PIL/Image.py", line 97, in <module>
    from . import _imaging as core
ImportError: /lib64/libstdc++.so.6: version `GLIBCXX_3.4.29' not found (required by /home/tsmith/miniconda3/envs/ftorch/lib/python3.11/site-packages/PIL/../../.././libLerc.so.4)
```

Some environment details: I'm using the most up to date version of FTorch with the following package versions (I used conda).

```
(ftorch) [psl_linux build]$ conda list gxx
# packages in environment at /home/tsmith/miniconda3/envs/ftorch:
#
# Name                    Version                   Build  Channel
gxx                       11.4.0              h602e360_13    conda-forge
gxx_impl_linux-64         11.4.0              h634f3ee_13    conda-forge

(ftorch) [psl_linux 2_ResNet18]$ conda list torch
# packages in environment at /home/tsmith/miniconda3/envs/ftorch:
#
# Name                    Version                   Build  Channel
pytorch                   2.5.1           py3.11_cuda12.1_cudnn9.1.0_0    pytorch
pytorch-cuda              12.1                 ha16c6d3_6    pytorch
pytorch-mutex             1.0                        cuda    pytorch
torchaudio                2.5.1               py311_cu121    pytorch
torchtriton               3.1.0                     py311    pytorch
torchvision               0.20.1              py311_cu121    pytorch
```

Any ideas on this one? Thanks!

cc: https://github.com/openjournals/joss-reviews/issues/7602 @matthewfeickert
