### 2025-01-06 Cambridge-ICCS/FTorch#214 @jwallwork23
Hi @timothyas,

Thanks for raising this.

I'm not a Conda user myself, but I noticed someone reported what appears to be the same issue on Stackoverflow:
https://stackoverflow.com/questions/68205760/install-glibcxx-3-4-29-in-anaconda

Does the following solution from that thread work for you? It installs the latest version of GlibC:
```
conda install -c conda-forge gxx_linux-64==11.1.0
```
