### 2025-01-06 Cambridge-ICCS/FTorch#214 @jwallwork23
That thread also references https://stackoverflow.com/questions/65349875/where-can-i-find-glibcxx-3-4-29 which includes the linker error you posted. What does `strings /usr/lib/x86_64-linux-gnu/libstdc++.so.6 | grep GLIBCXX` give?
