### 2025-01-06 Cambridge-ICCS/FTorch#212 @jatkinson1000
Hi @timothyas thanks for your helpful feedback!

Thanks for raising this.
We discussed this during writing and decided that we would not include FKB as it feels that it has moved towards abandonware - no updates in over 4 years, and unmerged PRs asking to upgrade to be compatible with keras 2.
The final decision was made when we saw that [neural-fortran](https://github.com/modern-fortran/neural-fortran?tab=readme-ov-file#related-projects), which we do cite, states on the repo that
> "As of v0.9.0, neural-fortran implements the full feature set of FKB in pure Fortran, and in addition supports training and inference of convolutional networks."

However, if you still feel it should be included we can do so.
We could also perhaps include [`fortran-tf-lib`](https://github.com/Cambridge-ICCS/fortran-tf-lib) as a similar library that works with more recent tensorflow/keras.

Let me know your thoughts.

Typo I'll look at correcting shortly :)
