### 2025-01-06 be2ce81
Avoid unused variable warnings
