### 2025-01-06 4a47f3c
Avoid empty loops
