+++ src/ftorch.F90
@@ -340,9 +340,12 @@ contains
       requires_grad_value = requires_grad
     end if
 
-    strides(layout(1)) = 1
-    do i = 2, ndims
-      strides(layout(i)) = strides(layout(i - 1)) * tensor_shape(layout(i - 1))
+    do i = 1, ndims
+      if (i == 1) then
+        strides(layout(i)) = 1
+      else
+        strides(layout(i)) = strides(layout(i - 1)) * tensor_shape(layout(i - 1))
+      end if
     end do
 
     ! Process optional arguments
@@ -1247,9 +1250,12 @@ contains
 
     tensor_shape = shape(data_in)
 
-    strides(layout(1)) = 1
-    do i = 2, ndims
-      strides(layout(i)) = strides(layout(i - 1)) * tensor_shape(layout(i - 1))
+    do i = 1, ndims
+      if (i == 1) then
+        strides(layout(i)) = 1
+      else
+        strides(layout(i)) = strides(layout(i - 1)) * tensor_shape(layout(i - 1))
+      end if
     end do
 
     tensor%p = torch_from_blob_c(c_loc(data_in), ndims, tensor_shape,          &
@@ -1301,9 +1307,12 @@ contains
 
     tensor_shape = shape(data_in)
 
-    strides(layout(1)) = 1
-    do i = 2, ndims
-      strides(layout(i)) = strides(layout(i - 1)) * tensor_shape(layout(i - 1))
+    do i = 1, ndims
+      if (i == 1) then
+        strides(layout(i)) = 1
+      else
+        strides(layout(i)) = strides(layout(i - 1)) * tensor_shape(layout(i - 1))
+      end if
     end do
 
     tensor%p = torch_from_blob_c(c_loc(data_in), ndims, tensor_shape,          &
@@ -1355,9 +1364,12 @@ contains
 
     tensor_shape = shape(data_in)
 
-    strides(layout(1)) = 1
-    do i = 2, ndims
-      strides(layout(i)) = strides(layout(i - 1)) * tensor_shape(layout(i - 1))
+    do i = 1, ndims
+      if (i == 1) then
+        strides(layout(i)) = 1
+      else
+        strides(layout(i)) = strides(layout(i - 1)) * tensor_shape(layout(i - 1))
+      end if
     end do
 
     tensor%p = torch_from_blob_c(c_loc(data_in), ndims, tensor_shape,          &
@@ -1409,9 +1421,12 @@ contains
 
     tensor_shape = shape(data_in)
 
-    strides(layout(1)) = 1
-    do i = 2, ndims
-      strides(layout(i)) = strides(layout(i - 1)) * tensor_shape(layout(i - 1))
+    do i = 1, ndims
+      if (i == 1) then
+        strides(layout(i)) = 1
+      else
+        strides(layout(i)) = strides(layout(i - 1)) * tensor_shape(layout(i - 1))
+      end if
     end do
 
     tensor%p = torch_from_blob_c(c_loc(data_in), ndims, tensor_shape,          &
@@ -1463,9 +1478,12 @@ contains
 
     tensor_shape = shape(data_in)
 
-    strides(layout(1)) = 1
-    do i = 2, ndims
-      strides(layout(i)) = strides(layout(i - 1)) * tensor_shape(layout(i - 1))
+    do i = 1, ndims
+      if (i == 1) then
+        strides(layout(i)) = 1
+      else
+        strides(layout(i)) = strides(layout(i - 1)) * tensor_shape(layout(i - 1))
+      end if
     end do
 
     tensor%p = torch_from_blob_c(c_loc(data_in), ndims, tensor_shape,          &
[…]
