### 2025-01-06 f2906f7
Typo correction as spotted by @timothyas during JOSS review.
