### 2025-01-06 0484364
Update JOSS workflow to satisfy zizmor linting.
