+++ .github/workflows/JOSS_paper_pdf.yml
@@ -22,6 +22,8 @@ jobs:
       # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
       - name: Checkout
         uses: actions/checkout@v4
+        with:
+          persist-credentials: false
 
       # Builds/renders the paper using the openjournals action
       - name: Build draft PDF
