### 2025-01-06 d2e48b7
Lean on compiler (#215)
* Avoid unused variable warnings
* Avoid empty loops
* Use c_prec correctly in overloaded ops
* Align number of characters with command line input in example 2
