+++ src/ftorch.F90
@@ -340,9 +340,12 @@ contains
       requires_grad_value = requires_grad
     end if
 
-    strides(layout(1)) = 1
-    do i = 2, ndims
-      strides(layout(i)) = strides(layout(i - 1)) * tensor_shape(layout(i - 1))
+    do i = 1, ndims
+      if (i == 1) then
+        strides(layout(i)) = 1
+      else
+        strides(layout(i)) = strides(layout(i - 1)) * tensor_shape(layout(i - 1))
+      end if
     end do
 
     ! Process optional arguments
@@ -553,10 +556,9 @@ contains
     interface
       function torch_tensor_premultiply_c(scalar_c, tensor_c) result(output_c) &
           bind(c, name = 'torch_tensor_premultiply')
-        use, intrinsic :: iso_c_binding, only : c_ptr
-        use, intrinsic :: iso_fortran_env, only : int8
+        use, intrinsic :: iso_c_binding, only : c_ptr, c_int8_t
         implicit none
-        integer(kind=int8), value, intent(in) :: scalar_c
+        integer(kind=c_int8_t), value, intent(in) :: scalar_c
         type(c_ptr), value, intent(in) :: tensor_c
         type(c_ptr) :: output_c
       end function torch_tensor_premultiply_c
@@ -574,10 +576,9 @@ contains
     interface
       function torch_tensor_premultiply_c(scalar_c, tensor_c) result(output_c) &
           bind(c, name = 'torch_tensor_premultiply')
-        use, intrinsic :: iso_c_binding, only : c_ptr
-        use, intrinsic :: iso_fortran_env, only : int16
+        use, intrinsic :: iso_c_binding, only : c_ptr, c_int16_t
         implicit none
-        integer(kind=int16), value, intent(in) :: scalar_c
+        integer(kind=c_int16_t), value, intent(in) :: scalar_c
         type(c_ptr), value, intent(in) :: tensor_c
         type(c_ptr) :: output_c
       end function torch_tensor_premultiply_c
@@ -595,10 +596,9 @@ contains
     interface
       function torch_tensor_premultiply_c(scalar_c, tensor_c) result(output_c) &
           bind(c, name = 'torch_tensor_premultiply')
-        use, intrinsic :: iso_c_binding, only : c_ptr
-        use, intrinsic :: iso_fortran_env, only : int32
+        use, intrinsic :: iso_c_binding, only : c_ptr, c_int32_t
         implicit none
-        integer(kind=int32), value, intent(in) :: scalar_c
+        integer(kind=c_int32_t), value, intent(in) :: scalar_c
         type(c_ptr), value, intent(in) :: tensor_c
         type(c_ptr) :: output_c
       end function torch_tensor_premultiply_c
@@ -616,10 +616,9 @@ contains
     interface
       function torch_tensor_premultiply_c(scalar_c, tensor_c) result(output_c) &
           bind(c, name = 'torch_tensor_premultiply')
-        use, intrinsic :: iso_c_binding, only : c_ptr
-        use, intrinsic :: iso_fortran_env, only : int64
+        use, intrinsic :: iso_c_binding, only : c_ptr, c_int64_t
         implicit none
-        integer(kind=int64), value, intent(in) :: scalar_c
+        integer(kind=c_int64_t), value, intent(in) :: scalar_c
         type(c_ptr), value, intent(in) :: tensor_c
         type(c_ptr) :: output_c
       end function torch_tensor_premultiply_c
@@ -637,10 +636,9 @@ contains
     interface
       function torch_tensor_premultiply_c(scalar_c, tensor_c) result(output_c) &
           bind(c, name = 'torch_tensor_premultiply')
-        use, intrinsic :: iso_c_binding, only : c_ptr
[…]
