### 2025-01-06 0db6316
Call from_blob inside from_array
