+++ src/ftorch.F90
@@ -1249,46 +1249,19 @@ contains
     integer(ftorch_int), intent(in)      :: layout(1)  !! Control order of indices
     integer(c_int), intent(in)           :: device_type    !! Device type the tensor will live on (`torch_kCPU` or `torch_kCUDA`)
     integer(c_int), optional, intent(in) :: device_index   !! device index to use for `torch_kCUDA` case
-    logical, optional, intent(in)        :: requires_grad  !! Whether gradients need to be computed for the created tensor
+    logical(c_bool), optional, intent(in) :: requires_grad  !! Whether gradients need to be computed for the created tensor
 
     ! local data
     integer(c_int64_t)        :: tensor_shape(1)            !! Shape of the tensor
     integer(c_int), parameter :: dtype = torch_kInt8  !! Data type
     integer(c_int64_t)        :: strides(1)                 !! Strides for accessing data
     integer(c_int), parameter :: ndims = 1                  !! Number of dimension of input data
-    integer(ftorch_int)       :: i
-    integer(c_int)            :: device_index_value
-    logical :: requires_grad_value  !! Whether gradients need to be computed for the created tensor
-
-    ! Process optional arguments
-    if (present(device_index)) then
-      device_index_value = device_index
-    else if (device_type == torch_kCPU) then
-      device_index_value = -1
-    else
-      device_index_value = 0
-    endif
-
-    if (.not. present(requires_grad)) then
-      requires_grad_value = .false.
-    else
-      requires_grad_value = requires_grad
-    end if
 
     tensor_shape = shape(data_in)
 
-    do i = 1, ndims
-      if (i == 1) then
-        strides(layout(i)) = 1
-      else
-        strides(layout(i)) = strides(layout(i - 1)) * tensor_shape(layout(i - 1))
-      end if
-    end do
-
-    tensor%p = torch_from_blob_c(c_loc(data_in), ndims, tensor_shape,          &
-                                 strides, dtype, device_type,                  &
-                                 device_index_value,                           &
-                                 logical(requires_grad_value, c_bool))
+    call torch_tensor_from_blob(tensor, c_loc(data_in), ndims, tensor_shape, &
+                                layout, dtype, device_type, device_index, &
+                                requires_grad)
 
   end subroutine torch_tensor_from_array_int8_1d
 
@@ -1306,46 +1279,19 @@ contains
     integer(ftorch_int), intent(in)      :: layout(2)  !! Control order of indices
     integer(c_int), intent(in)           :: device_type    !! Device type the tensor will live on (`torch_kCPU` or `torch_kCUDA`)
     integer(c_int), optional, intent(in) :: device_index   !! device index to use for `torch_kCUDA` case
-    logical, optional, intent(in)        :: requires_grad  !! Whether gradients need to be computed for the created tensor
+    logical(c_bool), optional, intent(in) :: requires_grad  !! Whether gradients need to be computed for the created tensor
 
     ! local data
     integer(c_int64_t)        :: tensor_shape(2)            !! Shape of the tensor
     integer(c_int), parameter :: dtype = torch_kInt8  !! Data type
     integer(c_int64_t)        :: strides(2)                 !! Strides for accessing data
     integer(c_int), parameter :: ndims = 2                  !! Number of dimension of input data
-    integer(ftorch_int)       :: i
-    integer(c_int)            :: device_index_value
-    logical :: requires_grad_value  !! Whether gradients need to be computed for the created tensor
-
-    ! Process optional arguments
-    if (present(device_index)) then
-      device_index_value = device_index
-    else if (device_type == torch_kCPU) then
-      device_index_value = -1
-    else
-      device_index_value = 0
-    endif
-
[…]
