### 2025-01-06 2d7758c
Accept requires_grad as logical rather than c_bool; test this arg
