### 2025-01-06 f24db4d
Call from_blob inside from_array (#216)
* Call from_blob inside from_array
* Accept requires_grad as logical rather than c_bool; test this arg
