### 2025-01-06 2011ee6
Accept integer for device_index, rather than c_int
