### 2025-01-06 781c004
Reorganise main FTorch module, with comments
