+++ src/ftorch.F90
@@ -43,13 +43,14 @@ module ftorch
     enumerator :: torch_kFloat64 = 7
   end enum
 
-
   !| Enumerator for Torch devices
   !  From c_torch.h (torch_device_t)
   enum, bind(c)
     enumerator :: torch_kCPU = 0
     enumerator :: torch_kCUDA = 1
   end enum
+
+  ! --- Interfaces for core FTorch procedures
 
   !> Interface for directing `torch_tensor_from_array` to possible input types and ranks
   interface torch_tensor_from_array
@@ -148,6 +149,21 @@ module ftorch
     end function torch_from_blob_c
   end interface
 
+  interface
+    function torch_to_blob_c(tensor, dtype) result(data) &
+        bind(c, name = 'torch_to_blob')
+      use, intrinsic :: iso_c_binding, only : c_int, c_ptr
+
+      implicit none
+
+      type(c_ptr), value, intent(in)    :: tensor
+      integer(c_int), value, intent(in) :: dtype
+      type(c_ptr)                       :: data
+    end function torch_to_blob_c
+  end interface
+
+  ! --- Interfaces for overloaded operators acting on tensors
+
   interface assignment (=)
     module procedure torch_tensor_assign
   end interface
@@ -195,20 +211,9 @@ module ftorch
     module procedure torch_tensor_power_real64
   end interface
 
-  interface
-    function torch_to_blob_c(tensor, dtype) result(data) &
-        bind(c, name = 'torch_to_blob')
-      use, intrinsic :: iso_c_binding, only : c_int, c_ptr
-
-      implicit none
-
-      type(c_ptr), value, intent(in)    :: tensor
-      integer(c_int), value, intent(in) :: dtype
-      type(c_ptr)                       :: data
-    end function torch_to_blob_c
-  end interface
-
 contains
+
+  ! --- Procedures for constructing tensors
 
   !> Returns a tensor with uninitialised values.
   subroutine torch_tensor_empty(tensor, ndims, tensor_shape, dtype, &
@@ -363,7 +368,6 @@ contains
                             device_index_value, requires_grad_value)
   end subroutine torch_tensor_ones
 
-  ! Torch Tensor API
   !| Exposes the given data as a tensor without taking ownership of the original data.
   !  This routine will take an (i, j, k) array and return an (k, j, i) tensor.
   subroutine torch_tensor_from_blob(tensor, data, ndims, tensor_shape, layout, dtype, &
@@ -412,828 +416,6 @@ contains
                                  device_type, device_index_value,              &
                                  requires_grad_value)
   end subroutine torch_tensor_from_blob
-
-  !> Prints the contents of a tensor.
-  subroutine torch_tensor_print(tensor)
-    type(torch_tensor), intent(in) :: tensor  !! Input tensor
-
-    interface
-      subroutine torch_tensor_print_c(tensor) &
-          bind(c, name = 'torch_tensor_print')
-        use, intrinsic :: iso_c_binding, only : c_ptr
-        implicit none
-        type(c_ptr), value, intent(in) :: tensor
-      end subroutine torch_tensor_print_c
-    end interface
-
-    call torch_tensor_print_c(tensor%p)
-  end subroutine torch_tensor_print
-
-  !> Determines the device index of a tensor.
-  function torch_tensor_get_device_index(tensor) result(device_index)
-    use, intrinsic :: iso_c_binding, only : c_int
-    type(torch_tensor), intent(in) :: tensor  !! Input tensor
-    integer(c_int) :: device_index            !! Device index of tensor
-
-    interface
[…]
