### 2025-01-06 ea78a8c
Add section for derived types and enums
