### 2025-01-06 cbdefb3
Apply suggestions from @TomMelt code review
Co-authored-by: tommelt <tdm39@cam.ac.uk>
