+++ .github/workflows/test_suite_ubuntu.yml
@@ -77,7 +77,7 @@ jobs:
           mkdir pFUnit/build
           cd pFUnit/build
           cmake -DMPI=YES ..
-          make install
+          make -j 4 install
 
       - name: Build FTorch
         run: |
@@ -94,7 +94,7 @@ jobs:
             -DCMAKE_BUILD_TYPE=Release \
             -DCMAKE_INSTALL_PREFIX=${BUILD_DIR} \
             -DCMAKE_BUILD_TESTS=TRUE \
-            -DCMAKE_PREFIX_PATH="${PFUNIT_DIR}:${Torch_DIR}" \
+            -DCMAKE_PREFIX_PATH="${PFUNIT_DIR};${Torch_DIR}" \
             -DCMAKE_Fortran_FLAGS="-std=${{ matrix.std }}"
           cmake --build .
           cmake --install .
