### 2025-01-06 385d57b
Try dropping MPI=YES
