### 2025-01-06 80c1d18
Apply suggestions from @jatkinson1000 code review
Co-authored-by: Jack Atkinson <109271713+jatkinson1000@users.noreply.github.com>
