### 2025-01-06 ef4615e
Apply other suggestions from @jatkinson1000 code review
Co-authored-by: Jack Atkinson <109271713+jatkinson1000@users.noreply.github.com>
