+++ src/test/README.md
@@ -7,6 +7,8 @@ The `unit` subdirectory contains unit tests written using
 compile these, you will need to first install pFUnit and then pass its install
 directory to the `CMAKE_PREFIX_PATH` when building FTorch.
 
+Note that the unit tests are currently only available on Unix based systems, not Windows.
+
 
 ## Integration tests
 
@@ -37,8 +39,8 @@ installing into your virtual environment.
 Alternatively, run the helper script in the root FTorch directory. Depending on
 which operating system you are running, you will need:
 
-- `./run_test_suite.sh` for unix (mac and linux)
-- `run_test_suite.bat` for windows
+- `./run_test_suite.sh` for unix (Mac and Linux)
+- `run_test_suite.bat` for Windows (note this will only run the integration tests, not unit tests).
 
 This will automatically install any additional Python dependencies for the
 examples and run the tests.
