### 2025-01-06 Cambridge-ICCS/FTorch#212 @timothyas
Thanks @jatkinson1000, that is good to know (and TIL the term Abandonware... nice). Since FKB seems to be history, I don't think it's necessary to mention in the paper.

That said, I think it would be good to mention the fortran-tf-lib as a similar library just for people's awareness.
