### 2025-01-07 #7602 @matthewfeickert
As we're now in 2025 ( :tada: ) and most people have their reviewer checklist created, I'll have @editorialbot give us reminders in 3 weeks to follow up on the initial state of the review. The review is currently in different stages, though as we already covered in https://github.com/openjournals/joss-reviews/issues/7602#issuecomment-2551875338, this is totally fine and expected. :+1:

@anand-me can you please create your reviewer checklist (c.f. https://github.com/openjournals/joss-reviews/issues/7602#issue-2748283405)?
