### 2025-01-07 Cambridge-ICCS/FTorch#214 @matthewfeickert
> Does the following solution from that thread work for you? It installs the latest version of GlibC:
>
> ```
> conda install -c conda-forge gxx_linux-64==11.1.0
> ```

:wave: I don't normally butt into review discussion, so please forgive the interruption, but as a [`pixi`](https://pixi.sh/latest/) user and maintainer of conda-forge packages, the recommendation is not to try to manually set your platform specific `cxx` compilers by hand but do user the [`compiler` metapackages](https://github.com/conda-forge/compilers-feedstock) that will do this for you. So here the recommendation would be to install `cxx-compiler`. So with `conda`/`mamba` that's

```
conda install --channel conda-forge cxx-compiler
```

and with `pixi` it is

```
pixi add cxx-compiler
```

As you can see from `pixi list`/`conda list`, this gets all required tools and on environment activation additionally will set environmental variables like `CXX`, `CFLAGS`, `CPPFLAGS`, `CXXFLAGS`, and `LDFLAGS`.

<summary>pixi list:</summary>

```console
$ mkdir example && cd $_
$ pixi init
$ pixi add cxx-compiler
$ pixi list
Package                   Version  Build         Size       Kind   Source
_libgcc_mutex             0.1      conda_forge   2.5 KiB    conda  _libgcc_mutex
_openmp_mutex             4.5      2_gnu         23.1 KiB   conda  _openmp_mutex
binutils                  2.43     h4852527_2    33.1 KiB   conda  binutils
binutils_impl_linux-64    2.43     h4bf12b8_2    5.4 MiB    conda  binutils_impl_linux-64
binutils_linux-64         2.43     h4852527_2    34.1 KiB   conda  binutils_linux-64
c-compiler                1.8.0    h2b85faf_1    5.9 KiB    conda  c-compiler
cxx-compiler              1.8.0    h1a2810e_1    5.9 KiB    conda  cxx-compiler
gcc                       13.3.0   h9576a4e_1    52.6 KiB   conda  gcc
gcc_impl_linux-64         13.3.0   hfea6d02_1    64.3 MiB   conda  gcc_impl_linux-64
gcc_linux-64              13.3.0   hc28eda2_7    31.3 KiB   conda  gcc_linux-64
gxx                       13.3.0   h9576a4e_1    52.1 KiB   conda  gxx
gxx_impl_linux-64         13.3.0   hdbfa832_1    12.7 MiB   conda  gxx_impl_linux-64
gxx_linux-64              13.3.0   h6834431_7    29.6 KiB   conda  gxx_linux-64
kernel-headers_linux-64   3.10.0   he073ed8_18   921.4 KiB  conda  kernel-headers_linux-64
ld_impl_linux-64          2.43     h712a8e2_2    653.5 KiB  conda  ld_impl_linux-64
libgcc                    14.2.0   h77fa898_1    828.9 KiB  conda  libgcc
libgcc-devel_linux-64     13.3.0   h84ea5a7_101  2.5 MiB    conda  libgcc-devel_linux-64
libgomp                   14.2.0   h77fa898_1    450.2 KiB  conda  libgomp
libsanitizer              13.3.0   heb74ff8_1    3.9 MiB    conda  libsanitizer
libstdcxx                 14.2.0   hc0a3c3a_1    3.7 MiB    conda  libstdcxx
libstdcxx-devel_linux-64  13.3.0   h84ea5a7_101  13.4 MiB   conda  libstdcxx-devel_linux-64
sysroot_linux-64          2.17     h0157908_18   14.5 MiB   conda  sysroot_linux-64
tzdata                    2024b    hc8b5060_0    119.5 KiB  conda  tzdat
```
