### 2025-01-07 Cambridge-ICCS/FTorch#214 @jatkinson1000
Thanks @matthewfeickert

We are not regular conda users here, so that's really useful, thank you.
If this works for you @timothyas let us know and we can perhaps add something to the online docs to help conda users.
