### 2025-01-07 Cambridge-ICCS/FTorch#214 @matthewfeickert
I nerd sniped myself on this a bit, and to test CUDA capabilities @timothyas is going to have to do some manual installation given that I forgot that [conda-forge is not legally able to distribute CUDA headers given's NVIDIA's EULA](https://conda-forge.org/docs/user/faq/#how-can-i-compile-cuda-host-or-device-codes-in-my-environment).

Though a note that it seems like as of 6055a7dd35e394daff469673feac063f1bfaae01 the `ENABLE_CUDA` flag doesn't properly disable checking for CUDA during the build(?) given that the following valid `pixi` manifest

```toml
[project]
authors = ["Matthew Feickert <matthew.feickert@cern.ch>"]
channels = ["conda-forge"]
description = "Add a short description here"
name = "example"
platforms = ["linux-64"]
version = "0.1.0"

[tasks]

# Selecting 6055a7dd35e394daff469673feac063f1bfaae01 as reference
clone = """
rm -rf FTorch && \
git clone https://github.com/Cambridge-ICCS/FTorch.git && \
cd FTorch && \
git reset --hard 6055a7dd35e394daff469673feac063f1bfaae01
"""

# Note that libtorch is at $CONDA_PREFIX/lib/libtorch.so
build = """
rm -rf build && \
cmake \
    -DCMAKE_INSTALL_PREFIX=$CONDA_PREFIX \
    -DCMAKE_PREFIX_PATH=$CONDA_PREFIX \
    -DCMAKE_BUILD_TYPE=Release \
    -DENABLE_CUDA=FALSE \
    -S FTorch/src/ \
    -B build && \
cmake -LH build && \
cmake --build build --clean-first --parallel=10 && \
cmake --install build
"""

start = { depends-on = ["clone", "build"] }

[system-requirements]
cuda = "12.0"

[dependencies]
cxx-compiler = ">=1.8.0,<2"
pytorch = ">=2.5.1,<3"
fortran-compiler = ">=1.8.0,<2"
cmake = ">=3.31.2,<4"
git = ">=2.47.1,<3"
zlib = ">=1.3.1,<2"
```

```console
$ pixi list | grep torch
libtorch                  2.5.1         cuda126_hebb32c0_306       490.2 MiB  conda  libtorch
pytorch                   2.5.1         cuda126_py313hae2543e_306  26.1 MiB   conda  pytorch
```

fails the build with

```console
$ pixi run start
...
-- Found Threads: TRUE
-- Caffe2: Found protobuf with new-style protobuf targets.
-- Caffe2: Protobuf version 28.2.0
-- Could NOT find CUDA (missing: CUDA_CUDART_LIBRARY) (found version "11.5")
CMake Warning at /tmp/example/.pixi/envs/default/share/cmake/Caffe2/public/cuda.cmake:31 (message):
  Caffe2: CUDA cannot be found.  Depending on whether you are building Caffe2
  or a Caffe2 dependent library, the next warning / error will give you more
  info.
Call Stack (most recent call first):
  /tmp/example/.pixi/envs/default/share/cmake/Caffe2/Caffe2Config.cmake:86 (include)
  /tmp/example/.pixi/envs/default/share/cmake/Torch/TorchConfig.cmake:68 (find_package)
  CMakeLists.txt:54 (find_package)

CMake Error at /tmp/example/.pixi/envs/default/share/cmake/Caffe2/Caffe2Config.cmake:90 (message):
  Your installed Caffe2 version uses CUDA but I cannot find the CUDA
  libraries.  Please set the proper CUDA prefixes and / or install CUDA.
Call Stack (most recent call first):
  /tmp/example/.pixi/envs/default/share/cmake/Torch/TorchConfig.cmake:68 (find_package)
  CMakeLists.txt:54 (find_package)

-- Configuring incomplete, errors occurred!
```

Note that removing the CUDA install and using CPU only environments results in a succesful build,

```toml
[project]
authors = ["Matthew Feickert <matthew.feickert@cern.ch>"]
channels = ["conda-forge"]
description = "Add a short description here"
name = "example"
platforms = ["linux-64"]
version = "0.1.0"

[tasks]

# Selecting 6055a7dd35e394daff469673feac063f1bfaae01 as reference
clone = """
rm -rf FTorch && \
git clone https://github.com/Cambridge-ICCS/FTorch.git && \
cd FTorch && \
git reset --hard 6055a7dd35e394daff469673feac063f1bfaae01
"""

# Note that libtorch is at $CONDA_PREFIX/lib/libtorch.so
build = """
rm -rf build && \
cmake \
    -DCMAKE_INSTALL_PREFIX=$CONDA_PREFIX \
    -DCMAKE_PREFIX_PATH=$CONDA_PREFIX \
    -DCMAKE_BUILD_TYPE=Release \
    -DENABLE_CUDA=FALSE \
    -S FTorch/src/ \
    -B build && \
cmake -LH build && \
cmake --build build --clean-first --parallel=10 && \
cmake --install build
"""

start = { depends-on = ["clone", "build"] }

# [system-requirements]
# cuda = "12.0"

[dependencies]
cxx-compiler = ">=1.8.0,<2"
pytorch = ">=2.5.1,<3"
fortran-compiler = ">=1.8.0,<2"
cmake = ">=3.31.2,<4"
git = ">=2.47.1,<3"
zlib = ">=1.3.1,<2"
```

```console
$ pixi clean
$ rm pixi.lock
$ pixi install
$ pixi list | grep torch
libtorch                  2.5.1         cpu_mkl_he8ec5d7_108        50.9 MiB   conda  libtorch
pytorch                   2.5.1         cpu_mkl_py313_h90df46e_108  26.2 MiB   conda  pytorch
```

```console
$ pixi run start
...
[100%] Linking CXX shared library libftorch.so
[100%] Built target ftorch
-- Install configuration: "Release"
-- Installing: /tmp/example/.pixi/envs/default/lib/libftorch.so
-- Set non-toolchain portion of runtime path of "/tmp/example/.pixi/envs/default/lib/libftorch.so" to "$ORIGIN/../lib:/lib/intel64:/lib/intel64_win:/lib/win-x64"
-- Installing: /tmp/example/.pixi/envs/default/include/ctorch.h
-- Installing: /tmp/example/.pixi/envs/default/lib/cmake/FTorch/FTorchConfig.cmake
-- Installing: /tmp/example/.pixi/envs/default/lib/cmake/FTorch/FTorchConfig-release.cmake
-- Installing: /tmp/example/.pixi/envs/default/include/ftorch/ftorch.mod
-- Installing: /tmp/example/.pixi/envs/default/include/ftorch/ftorch_test_utils.mod
```

so it seems that the `ENABLE_CUDA` flag might be getting ignored in CUDA enabled environments?

Also it seems that `zlib` is a dependency that isn't defined anywhere(?) as without it you get

```
Could NOT find ZLIB (missing: ZLIB_LIBRARY)
```

though maybe it is an optional dependency (it falls back to searching outside of the path and then finds `zlib` elsewhere on my machine, so not sure)?

Also enabling

```
-DCMAKE_BUILD_TESTS=TRUE
```

results in it selecting the wrong Python interpreter (perhaps the first it finds on the system and not the one that under `CMAKE_PREFIX_PATH`). Maybe the test suite requires additional undocumented configuration?
