### 2025-01-07 8eaf19f
Revert ftorch_int change
