### 2025-01-07 Cambridge-ICCS/FTorch#214 @jwallwork23
@matthewfeickert Thank you for raising this regarding `ENABLE_CUDA`. I'll look into this as part of #208.
