### 2025-01-07 9d4aa86
Build example 3 if CUDA and MPI enabled
