### 2025-01-07 a763b0e
Build example 3 if CUDA and MPI enabled
