+++ examples/CMakeLists.txt
@@ -1,7 +1,9 @@
 if(CMAKE_BUILD_TESTS)
   add_subdirectory(1_SimpleNet)
   add_subdirectory(2_ResNet18)
-  # add_subdirectory(3_MultiGPU)
+  if(ENABLE_CUDA AND ENABLE_MPI)
+    add_subdirectory(3_MultiGPU)
+  endif()
   add_subdirectory(4_MultiIO)
   # NOTE: Example 5_Looping is built here but not run as part of the integration
   #       test suite. This is because it demonstrates 'good' versus 'bad'
