### 2025-01-07 Cambridge-ICCS/FTorch#214 @ElliottKasoar
> I'll look into this as part of #208.

@jwallwork23, for context, I think I tried setting various flags to allow us to actively disable cuda (`NO_CUDA`, `USE_CUDA` etc.), and my memory is that libtorch ignored these.

[Originally](https://github.com/Cambridge-ICCS/FTorch/pull/49/files#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R90) there was a footnote: `This is often overridden by PyTorch. When installing with pip.....` which I think may have been moved/dropped in reorganising the docs.

I think I felt that `ENABLE_CUDA` was slightly more ambiguous than something like `USE_CUDA`, in that *not* enabling isn't quite the same as disabling.

The main alternative at the time was not having the check at all, which is also perfectly viable, but perhaps you can figure out a way to actually disable cuda/a more appropriate name.
