### 2025-01-07 Cambridge-ICCS/FTorch#214 @jatkinson1000
Yes, to back up what @ElliottKasoar is saying, to run with CUDA we need 2 things:

- CUDA aware libtorch
- to link FTorch to CUDA at build time

The `ENABLE_CUDA` flag is to tell the FTorch CMake that we want to build with CUDA, and to therefore look for a CUDA installation during the FTorch build, but we rely on the end user having installed the appropriate version of libtorch for what they want to build (i.e. CUDA, or cpu).

If a user has installed CUDA aware libtorch but does not have CUDA then libtorch (Torch) will raise this as an issue which is what you see above.

As far as we could tell there is no way of checking if we have a CUDA or CPU version of libtorch, but if there is we would add a compatibility check.

@ElliottKasoar Thanks for spotting the docs misalignment - I'll add this to a docs patch that I am currently working on.
