### 2025-01-07 36e1019
Update missing info about installing CUDA vs CPU libtorch in README, as pointed out by @ElliottKasoar.
