### 2025-01-07 c25c6d3
Fix broken link to test script in docs.
