+++ src/ctorch.cpp
@@ -277,22 +277,24 @@ torch_tensor_t torch_tensor_multiply(const torch_tensor_t tensor1,
   return output;
 }
 
-torch_tensor_t torch_tensor_premultiply(const torch_data_t scalar,
+// FIXME: Don't cast to float. Now doesn't work for integers
+torch_tensor_t torch_tensor_premultiply(const float scalar,
                                         const torch_tensor_t tensor) {
   auto t = reinterpret_cast<torch::Tensor *const>(tensor);
   torch::Tensor *output = nullptr;
   output = new torch::Tensor;
-  std::cout << "DEBUG: scalar = " << scalar << std::endl;
+  std::cout << "DEBUG: pre-multiplier = " << scalar << std::endl;
   *output = scalar * *t;
   return output;
 }
 
+// FIXME: Don't cast to float. Now doesn't work for integers
 torch_tensor_t torch_tensor_postmultiply(const torch_tensor_t tensor,
-                                         const torch_data_t scalar) {
+                                         const float scalar) {
   auto t = reinterpret_cast<torch::Tensor *const>(tensor);
   torch::Tensor *output = nullptr;
   output = new torch::Tensor;
-  std::cout << "DEBUG: scalar = " << scalar << std::endl;
+  std::cout << "DEBUG: post-multiplier = " << scalar << std::endl;
   *output = *t * scalar;
   return output;
 }
@@ -307,18 +309,22 @@ torch_tensor_t torch_tensor_divide(const torch_tensor_t tensor1,
   return output;
 }
 
+// FIXME: Don't cast to float. Now doesn't work for integers
 torch_tensor_t torch_tensor_postdivide(const torch_tensor_t tensor,
-                                       const torch_data_t scalar) {
+                                       const float scalar) {
   auto t = reinterpret_cast<torch::Tensor *const>(tensor);
   torch::Tensor *output = nullptr;
   output = new torch::Tensor;
-  std::cout << "DEBUG: scalar = " << scalar << std::endl;
-  *output = *t / scalar;
+  torch::Scalar *alpha = nullptr;
+  alpha = new torch::Scalar;
+  *alpha = scalar;
+  std::cout << "DEBUG: divisor = " << scalar << std::endl;
+  *output = *t / *alpha;
   return output;
 }
 
-torch_tensor_t torch_tensor_power(const torch_tensor_t tensor,
-                                  const torch_data_t exponent) {
+// FIXME: Don't cast to float. Now doesn't work for integers
+torch_tensor_t torch_tensor_power(const torch_tensor_t tensor, const float exponent) {
   auto t = reinterpret_cast<torch::Tensor *const>(tensor);
   torch::Tensor *output = nullptr;
   output = new torch::Tensor;
