### 2025-01-07 e01cc83
Introduce torch_scalar_t
