+++ src/ftorch.F90
@@ -2428,8 +2428,8 @@ contains
 
   !> Overloads multiplication operator for a scalar of type int8 and a tensor.
   function torch_tensor_premultiply_int8(scalar, tensor) result(output)
-    use, intrinsic :: iso_fortran_env, only : int8
-    integer(int8), intent(in) :: scalar
+    use, intrinsic :: iso_c_binding, only : c_loc
+    integer(int8), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
     type(torch_tensor) :: output
 
@@ -2438,19 +2438,19 @@ contains
           bind(c, name = 'torch_tensor_premultiply')
         use, intrinsic :: iso_c_binding, only : c_ptr, c_int8_t
         implicit none
-        integer(kind=c_int8_t), value, intent(in) :: scalar_c
+        type(c_ptr), value, intent(in) :: scalar_c
         type(c_ptr), value, intent(in) :: tensor_c
         type(c_ptr) :: output_c
       end function torch_tensor_premultiply_c
     end interface
 
-    output%p = torch_tensor_premultiply_c(scalar, tensor%p)
+    output%p = torch_tensor_premultiply_c(c_loc(scalar), tensor%p)
   end function torch_tensor_premultiply_int8
 
   !> Overloads multiplication operator for a scalar of type int16 and a tensor.
   function torch_tensor_premultiply_int16(scalar, tensor) result(output)
-    use, intrinsic :: iso_fortran_env, only : int16
-    integer(int16), intent(in) :: scalar
+    use, intrinsic :: iso_c_binding, only : c_loc
+    integer(int16), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
     type(torch_tensor) :: output
 
@@ -2459,19 +2459,19 @@ contains
           bind(c, name = 'torch_tensor_premultiply')
         use, intrinsic :: iso_c_binding, only : c_ptr, c_int16_t
         implicit none
-        integer(kind=c_int16_t), value, intent(in) :: scalar_c
+        type(c_ptr), value, intent(in) :: scalar_c
         type(c_ptr), value, intent(in) :: tensor_c
         type(c_ptr) :: output_c
       end function torch_tensor_premultiply_c
     end interface
 
-    output%p = torch_tensor_premultiply_c(scalar, tensor%p)
+    output%p = torch_tensor_premultiply_c(c_loc(scalar), tensor%p)
   end function torch_tensor_premultiply_int16
 
   !> Overloads multiplication operator for a scalar of type int32 and a tensor.
   function torch_tensor_premultiply_int32(scalar, tensor) result(output)
-    use, intrinsic :: iso_fortran_env, only : int32
-    integer(int32), intent(in) :: scalar
+    use, intrinsic :: iso_c_binding, only : c_loc
+    integer(int32), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
     type(torch_tensor) :: output
 
@@ -2480,19 +2480,19 @@ contains
           bind(c, name = 'torch_tensor_premultiply')
         use, intrinsic :: iso_c_binding, only : c_ptr, c_int32_t
         implicit none
-        integer(kind=c_int32_t), value, intent(in) :: scalar_c
+        type(c_ptr), value, intent(in) :: scalar_c
         type(c_ptr), value, intent(in) :: tensor_c
         type(c_ptr) :: output_c
       end function torch_tensor_premultiply_c
     end interface
 
-    output%p = torch_tensor_premultiply_c(scalar, tensor%p)
+    output%p = torch_tensor_premultiply_c(c_loc(scalar), tensor%p)
   end function torch_tensor_premultiply_int32
 
[…]
