### 2025-01-07 985b591
Drop pre/post, just pass trivial tensors
