+++ src/ftorch.F90
@@ -33,6 +33,11 @@ module ftorch
     procedure :: get_shape
   end type torch_tensor
 
+  !> Type for holding a Torch scalar.
+  type torch_scalar
+    type(c_ptr) :: p = c_null_ptr  !! pointer to the scalar in memory
+  end type torch_scalar
+
   !| Enumerator for Torch data types
   !  From c_torch.h (torch_data_t)
   !  Note that 0 `torch_kUInt8` and 5 `torch_kFloat16` are not sypported in Fortran
@@ -2431,11 +2436,11 @@ contains
     use, intrinsic :: iso_c_binding, only : c_loc
     integer(int8), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
-    type(torch_tensor) :: output
+    type(torch_tensor) :: wrk, output
 
     interface
       function torch_tensor_premultiply_c(scalar_c, tensor_c) result(output_c) &
-          bind(c, name = 'torch_tensor_premultiply')
+          bind(c, name = 'torch_tensor_multiply')
         use, intrinsic :: iso_c_binding, only : c_ptr, c_int8_t
         implicit none
         type(c_ptr), value, intent(in) :: scalar_c
@@ -2444,7 +2449,8 @@ contains
       end function torch_tensor_premultiply_c
     end interface
 
-    output%p = torch_tensor_premultiply_c(c_loc(scalar), tensor%p)
+    call torch_tensor_from_array(wrk, [scalar], [1], torch_kCPU)
+    output%p = torch_tensor_premultiply_c(wrk%p, tensor%p)
   end function torch_tensor_premultiply_int8
 
   !> Overloads multiplication operator for a scalar of type int16 and a tensor.
@@ -2452,11 +2458,11 @@ contains
     use, intrinsic :: iso_c_binding, only : c_loc
     integer(int16), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
-    type(torch_tensor) :: output
+    type(torch_tensor) :: wrk, output
 
     interface
       function torch_tensor_premultiply_c(scalar_c, tensor_c) result(output_c) &
-          bind(c, name = 'torch_tensor_premultiply')
+          bind(c, name = 'torch_tensor_multiply')
         use, intrinsic :: iso_c_binding, only : c_ptr, c_int16_t
         implicit none
         type(c_ptr), value, intent(in) :: scalar_c
@@ -2465,7 +2471,8 @@ contains
       end function torch_tensor_premultiply_c
     end interface
 
-    output%p = torch_tensor_premultiply_c(c_loc(scalar), tensor%p)
+    call torch_tensor_from_array(wrk, [scalar], [1], torch_kCPU)
+    output%p = torch_tensor_premultiply_c(wrk%p, tensor%p)
   end function torch_tensor_premultiply_int16
 
   !> Overloads multiplication operator for a scalar of type int32 and a tensor.
@@ -2473,11 +2480,11 @@ contains
     use, intrinsic :: iso_c_binding, only : c_loc
     integer(int32), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
-    type(torch_tensor) :: output
+    type(torch_tensor) :: wrk, output
 
     interface
       function torch_tensor_premultiply_c(scalar_c, tensor_c) result(output_c) &
-          bind(c, name = 'torch_tensor_premultiply')
+          bind(c, name = 'torch_tensor_multiply')
         use, intrinsic :: iso_c_binding, only : c_ptr, c_int32_t
         implicit none
         type(c_ptr), value, intent(in) :: scalar_c
@@ -2486,7 +2493,8 @@ contains
       end function torch_tensor_premultiply_c
     end interface
 
[…]
