### 2025-01-07 735b78b
Restrict torch_tensor_power to integer exponents
