+++ src/ftorch.F90
@@ -220,8 +220,6 @@ module ftorch
     module procedure torch_tensor_power_int16
     module procedure torch_tensor_power_int32
     module procedure torch_tensor_power_int64
-    module procedure torch_tensor_power_real32
-    module procedure torch_tensor_power_real64
   end interface
 
 contains
@@ -2849,7 +2847,7 @@ contains
   end function torch_tensor_postdivide_real64
 
 
-  !> Overloads exponentiation operator for a tensor and a scalar of type `int8`
+  !> Overloads exponentiation operator for a tensor and an integer of type `int8`
   function torch_tensor_power_int8(tensor, power) result(output)
     use, intrinsic :: iso_c_binding, only : c_loc
     type(torch_tensor), intent(in) :: tensor
@@ -2870,7 +2868,7 @@ contains
     output%p = torch_tensor_power_c(tensor%p, c_loc(power))
   end function torch_tensor_power_int8
 
-  !> Overloads exponentiation operator for a tensor and a scalar of type `int16`
+  !> Overloads exponentiation operator for a tensor and an integer of type `int16`
   function torch_tensor_power_int16(tensor, power) result(output)
     use, intrinsic :: iso_c_binding, only : c_loc
     type(torch_tensor), intent(in) :: tensor
@@ -2891,7 +2889,7 @@ contains
     output%p = torch_tensor_power_c(tensor%p, c_loc(power))
   end function torch_tensor_power_int16
 
-  !> Overloads exponentiation operator for a tensor and a scalar of type `int32`
+  !> Overloads exponentiation operator for a tensor and an integer of type `int32`
   function torch_tensor_power_int32(tensor, power) result(output)
     use, intrinsic :: iso_c_binding, only : c_loc
     type(torch_tensor), intent(in) :: tensor
@@ -2912,7 +2910,7 @@ contains
     output%p = torch_tensor_power_c(tensor%p, c_loc(power))
   end function torch_tensor_power_int32
 
-  !> Overloads exponentiation operator for a tensor and a scalar of type `int64`
+  !> Overloads exponentiation operator for a tensor and an integer of type `int64`
   function torch_tensor_power_int64(tensor, power) result(output)
     use, intrinsic :: iso_c_binding, only : c_loc
     type(torch_tensor), intent(in) :: tensor
@@ -2932,48 +2930,6 @@ contains
 
     output%p = torch_tensor_power_c(tensor%p, c_loc(power))
   end function torch_tensor_power_int64
-
-  !> Overloads exponentiation operator for a tensor and a scalar of type `real32`
-  function torch_tensor_power_real32(tensor, power) result(output)
-    use, intrinsic :: iso_c_binding, only : c_loc
-    type(torch_tensor), intent(in) :: tensor
-    real(real32), target, intent(in) :: power
-    type(torch_tensor) :: output
-
-    interface
-      function torch_tensor_power_c(tensor_c, power_c) result(output_c)        &
-          bind(c, name = 'torch_tensor_power')
-        use, intrinsic :: iso_c_binding, only : c_ptr, c_float
-        implicit none
-        type(c_ptr), value, intent(in) :: tensor_c
-        type(c_ptr), value, intent(in) :: power_c
-        type(c_ptr) :: output_c
-      end function torch_tensor_power_c
-    end interface
-
-    output%p = torch_tensor_power_c(tensor%p, c_loc(power))
-  end function torch_tensor_power_real32
-
-  !> Overloads exponentiation operator for a tensor and a scalar of type `real64`
-  function torch_tensor_power_real64(tensor, power) result(output)
-    use, intrinsic :: iso_c_binding, only : c_loc
-    type(torch_tensor), intent(in) :: tensor
-    real(real64), target, intent(in) :: power
[…]
