### 2025-01-07 d0fb832
Reuse multiply and divide interfaces
