+++ src/ftorch.F90
@@ -205,6 +205,17 @@ module ftorch
     module procedure torch_tensor_postmultiply_real64
   end interface
 
+  interface
+    function torch_tensor_multiply_c(tensor1_c, tensor2_c) result(output_c)  &
+        bind(c, name = 'torch_tensor_multiply')
+      use, intrinsic :: iso_c_binding, only : c_ptr
+      implicit none
+      type(c_ptr), value, intent(in) :: tensor1_c
+      type(c_ptr), value, intent(in) :: tensor2_c
+      type(c_ptr) :: output_c
+    end function torch_tensor_multiply_c
+  end interface
+
   interface operator (/)
     module procedure torch_tensor_divide
     module procedure torch_tensor_postdivide_int8
@@ -213,6 +224,17 @@ module ftorch
     module procedure torch_tensor_postdivide_int64
     module procedure torch_tensor_postdivide_real32
     module procedure torch_tensor_postdivide_real64
+  end interface
+
+  interface
+    function torch_tensor_divide_c(tensor1_c, tensor2_c) result(output_c)  &
+        bind(c, name = 'torch_tensor_divide')
+      use, intrinsic :: iso_c_binding, only : c_ptr
+      implicit none
+      type(c_ptr), value, intent(in) :: tensor1_c
+      type(c_ptr), value, intent(in) :: tensor2_c
+      type(c_ptr) :: output_c
+    end function torch_tensor_divide_c
   end interface
 
   interface operator (**)
@@ -2415,17 +2437,6 @@ contains
     type(torch_tensor), intent(in) :: tensor2
     type(torch_tensor) :: output
 
-    interface
-      function torch_tensor_multiply_c(tensor1_c, tensor2_c) result(output_c)  &
-          bind(c, name = 'torch_tensor_multiply')
-        use, intrinsic :: iso_c_binding, only : c_ptr
-        implicit none
-        type(c_ptr), value, intent(in) :: tensor1_c
-        type(c_ptr), value, intent(in) :: tensor2_c
-        type(c_ptr) :: output_c
-      end function torch_tensor_multiply_c
-    end interface
-
     output%p = torch_tensor_multiply_c(tensor1%p, tensor2%p)
   end function torch_tensor_multiply
 
@@ -2436,19 +2447,9 @@ contains
     type(torch_tensor), intent(in) :: tensor
     type(torch_tensor) :: wrk, output
 
-    interface
-      function torch_tensor_premultiply_c(scalar_c, tensor_c) result(output_c) &
-          bind(c, name = 'torch_tensor_multiply')
-        use, intrinsic :: iso_c_binding, only : c_ptr, c_int8_t
-        implicit none
-        type(c_ptr), value, intent(in) :: scalar_c
-        type(c_ptr), value, intent(in) :: tensor_c
-        type(c_ptr) :: output_c
-      end function torch_tensor_premultiply_c
-    end interface
-
+    ! Create a tensor with a single entry, the scalar pre-multiplier
     call torch_tensor_from_array(wrk, [scalar], [1], torch_kCPU)
-    output%p = torch_tensor_premultiply_c(wrk%p, tensor%p)
+    output%p = torch_tensor_multiply_c(wrk%p, tensor%p)
   end function torch_tensor_premultiply_int8
 
   !> Overloads multiplication operator for a scalar of type int16 and a tensor.
@@ -2458,19 +2459,9 @@ contains
     type(torch_tensor), intent(in) :: tensor
     type(torch_tensor) :: wrk, output
 
-    interface
-      function torch_tensor_premultiply_c(scalar_c, tensor_c) result(output_c) &
-          bind(c, name = 'torch_tensor_multiply')
[…]
