### 2025-01-07 Cambridge-ICCS/FTorch#214 @timothyas
Hi everyone, looks like quite some lively discussion :)

I'll just say that I definitely don't want to force conda on you guys just because I like to use it. Part of the reason I was using conda is because I was running these examples on a linux machine at my institute which I think has some issues with the CUDA driver installation, at least as it relates to pytorch... And I think it's a really detailed issue related to environment variables, because I've not had problems running vanilla pytorch in my experience. Over the past couple of days I've made a lot of different attempts to create a working environment with CUDA, but haven't had any success.

So... I switched to a linux machine with no GPU/CUDA etc and was able to successfully run through the docs, including example 2 that was, for some reason, giving me trouble before.

With all that said I'd like to ask @matthewfeickert for a suggestion on how to proceed? Do you guys want to keep hacking through these environment issues? If so I can follow up with more info answering everyone's questions. My guess is that the installation trouble is truly related to getting CUDA+pytorch installed properly, rather than this package, given how smoothly everything ran on the CPU machine. But I can provide more details on what I tried, etc if you like. I'll just say though that I don't have a ton more time to dedicate toward getting it to work.
