### 2025-01-07 Cambridge-ICCS/FTorch#214 @matthewfeickert
> I switched to a linux machine with no GPU/CUDA etc and was able to successfully run through the docs, including example 2 that was, for some reason, giving me trouble before.
>
> With all that said I'd like to ask @matthewfeickert for a suggestion on how to proceed? Do you guys want to keep hacking through these environment issues? If so I can follow up with more info answering everyone's questions. My guess is that the installation trouble is truly related to getting CUDA+pytorch installed properly, rather than this package, given how smoothly everything ran on the CPU machine. But I can provide more details on what I tried, etc if you like. I'll just say though that I don't have a ton more time to dedicate toward getting it to work.

@timothyas Good question. The focus is on evaluating if the software does what it claims and is properly documented and then adding your own professional judgement to the quality and veracity of the software/performance. As FTorch doesn't make its whole experience and use pitch on hardware acceleration it is totally fine to evaluate it on CPU only if you are not able to get a working GPU enabled environment.

It also is not the reviewer's responsibility to fight with setting up the software environment. It ideally should be well documented on how to get things installed for all supported systems. That being said, while I think it would be great to make things as streamlined as possible for getting CUDA enabled environments, I think at some level this goes beyond the FTorch team and more to the point that if you need to be able to compile against CUDA and not just use its libraries, that is still a hard thing to do in today's world, even with all the advancements that conda-forge has made (it is a bit of a shame that NVIDIA has stopped supporting the `nvidia` channels for things like `cudatoolkit` as this used to be a really useful service).

So as to how to proceed, I would continue with your review in your working environment and then note the difficulties you had with CUDA in your review and give your professional suggestions for how things might be better improved in the future.
