+++ README.md
@@ -83,7 +83,7 @@ _For a similar approach to calling TensorFlow models from Fortran please see [Fo
 To install the library requires the following to be installed on the system:
 
 * CMake >= 3.15
-* [libtorch](https://pytorch.org/cppdocs/installing.html)<sup>*</sup> or [PyTorch](https://pytorch.org/)
+* [LibTorch](https://pytorch.org/cppdocs/installing.html)<sup>*</sup> or [PyTorch](https://pytorch.org/)
 * Fortran (2008 standard compliant), C++ (must fully support C++17), and C compilers
 
 <sup>*</sup> _The minimal example provided downloads the CPU-only Linux Nightly binary. [Alternative versions](https://pytorch.org/get-started/locally/) may be required._
@@ -95,13 +95,13 @@ If building in a Windows environment then you can either use
 or Visual Studio and the Intel Fortran Compiler.
 For full details on the process see the
 [online Windows documentation](https://cambridge-iccs.github.io/FTorch/page/troubleshooting.html#windows).\
-Note that libTorch is not supported for the GNU Fortran compiler with MinGW.
+Note that LibTorch is not supported for the GNU Fortran compiler with MinGW.
 
 #### Apple Silicon Support
 
-At the time of writing, libtorch is only officially available for x86 architectures
+At the time of writing, LibTorch is only officially available for x86 architectures
 (according to https://pytorch.org/). However, the version of PyTorch provided by
-`pip install torch` uses an ARM binary for libtorch which works on Apple Silicon.
+`pip install torch` uses an ARM binary for LibTorch which works on Apple Silicon.
 
 ### Library installation
 
@@ -137,7 +137,7 @@ To build and install the library:
     | [`CMAKE_Fortran_COMPILER`](https://cmake.org/cmake/help/latest/variable/CMAKE_LANG_COMPILER.html) | `ifort` / `gfortran`         | Specify a Fortran compiler to build the library with. This should match the Fortran compiler you're using to build the code you are calling this library from.<sup>1</sup>        |
     | [`CMAKE_C_COMPILER`](https://cmake.org/cmake/help/latest/variable/CMAKE_LANG_COMPILER.html)       | `icc` / `gcc`                | Specify a C compiler to build the library with.<sup>1</sup>                |
     | [`CMAKE_CXX_COMPILER`](https://cmake.org/cmake/help/latest/variable/CMAKE_LANG_COMPILER.html)     | `icpc` / `g++`               | Specify a C++ compiler to build the library with.<sup>1</sup>              |
-    | [`CMAKE_PREFIX_PATH`](https://cmake.org/cmake/help/latest/variable/CMAKE_PREFIX_PATH.html)        | `</path/to/libTorch/>`       | Location of Torch installation<sup>2</sup>                    |
+    | [`CMAKE_PREFIX_PATH`](https://cmake.org/cmake/help/latest/variable/CMAKE_PREFIX_PATH.html)        | `</path/to/LibTorch/>`       | Location of Torch installation<sup>2</sup>                    |
     | [`CMAKE_INSTALL_PREFIX`](https://cmake.org/cmake/help/latest/variable/CMAKE_INSTALL_PREFIX.html)  | `</path/to/install/lib/at/>` | Location at which the library files should be installed. By default this is `/usr/local` |
     | [`CMAKE_BUILD_TYPE`](https://cmake.org/cmake/help/latest/variable/CMAKE_BUILD_TYPE.html)          | `Release` / `Debug`          | Specifies build type. The default is `Debug`, use `Release` for production code|
     | `CMAKE_BUILD_TESTS`                                                                               | `TRUE` / `FALSE`             | Specifies whether to compile FTorch's [test suite](https://cambridge-iccs.github.io/FTorch/page/testing.html) as part of the build. |
[…]
