### 2025-01-07 8511c49
Update suitable compilers lists to include latest intel.
