### 2025-01-07 b776e55
Remove rogue code fence in docs.
