### 2025-01-07 Cambridge-ICCS/FTorch#214 @matthewfeickert
Ah, okay one more thing that is good news: Now that I realize that `cudatoolkit` is now [`cuda-toolkit`](https://anaconda.org/nvidia/cuda-toolkit) on the [`nvidia` conda channel](https://anaconda.org/nvidia/) I am able to get a working build environment with `ENABLE_CUDA=TRUE` with `pixi` with the following:

```toml
[project]
authors = ["Matthew Feickert <matthew.feickert@cern.ch>"]
channels = ["conda-forge", "nvidia/label/cuda-12.6.0"]
description = "Add a short description here"
name = "example"
platforms = ["linux-64"]
version = "0.1.0"

[tasks]

# Selecting 6055a7dd35e394daff469673feac063f1bfaae01 as reference
clone = """
rm -rf FTorch && \
git clone https://github.com/Cambridge-ICCS/FTorch.git && \
cd FTorch && \
git reset --hard 6055a7dd35e394daff469673feac063f1bfaae01
"""

# Note that libtorch is at $CONDA_PREFIX/lib/libtorch.so
build = """
rm -rf build && \
cmake \
    -DCMAKE_INSTALL_PREFIX=$CONDA_PREFIX \
    -DCMAKE_PREFIX_PATH=$CONDA_PREFIX \
    -DCMAKE_BUILD_TYPE=Release \
    -DENABLE_CUDA=TRUE \
    -S FTorch/src/ \
    -B build && \
cmake -LH build && \
cmake --build build --clean-first --parallel=10 && \
cmake --install build
"""

start = { depends-on = ["clone", "build"] }

[system-requirements]
cuda = "12.0"

[dependencies]
fortran-compiler = ">=1.8.0,<2"
cxx-compiler = ">=1.8.0,<2"
cmake = ">=3.31.2,<4"
pytorch = ">=2.5.1,<3"
zlib = ">=1.3.1,<2"
git = ">=2.47.1,<3"
cuda-toolkit = { version = "12.6.0.*", channel = "nvidia/label/cuda-12.6.0" }
```

Note that you're will need to massage the value of the `cuda-toolkit` and channel version a bit depending on what machine you're on, but this should wo

<summary>Full output:</summary>

```console
$ pixi clean && pixi run start
  removed /home/feickert/tmp/example/.pixi/envs                                                                                                                                                             ✨ Pixi task (clone): rm -rf FTorch && git clone https://github.com/Cambridge-ICCS/FTorch.git && cd FTorch && git reset --hard 6055a7dd35e394daff469673feac063f1bfaae01

Cloning into 'FTorch'...
remote: Enumerating objects: 5453, done.
remote: Total 5453 (delta 1347), reused 1115 (delta 1075), pack-reused 3796 (from 2)
Resolving deltas: 100% (4210/4210), done.
HEAD is now at 6055a7d Unit tests (using pFunit) (#76)

✨ Pixi task (build): rm -rf build && cmake -DCMAKE_INSTALL_PREFIX=$CONDA_PREFIX -DCMAKE_PREFIX_PATH=$CONDA_PREFIX -DCMAKE_BUILD_TYPE=Release -DENABLE_CUDA=TRUE -S FTorch/src/ -B build && cmake -LH build && cmake --build build --clean-first --parallel=10 && cmake --install build

-- The C compiler identification is GNU 13.3.0
-- The CXX compiler identification is GNU 13.3.0
-- The Fortran compiler identification is GNU 13.3.0
-- Detecting C compiler ABI info
-- Detecting C compiler ABI info - done
-- Check for working C compiler: /home/feickert/tmp/example/.pixi/envs/default/bin/x86_64-conda-linux-gnu-cc - skipped
-- Detecting C compile features
-- Detecting C compile features - done
-- Detecting CXX compiler ABI info
-- Check for working CXX compiler: /home/feickert/tmp/example/.pixi/envs/default/bin/x86_64-conda-linux-gnu-c++ - skipped
-- Detecting CXX compile features
-- Detecting Fortran compiler ABI info
-- Detecting Fortran compiler ABI info - done
-- Check for working Fortran compiler: /home/feickert/tmp/example/.pixi/envs/default/bin/x86_64-conda-linux-gnu-gfortran - skipped
-- Detecting Fortran/C Interface
-- Detecting Fortran/C Interface - Found GLOBAL and MODULE mangling
-- Verifying Fortran/CXX Compiler Compatibility
-- Verifying Fortran/CXX Compiler Compatibility - Success
-- Looking for a CUDA compiler
-- Looking for a CUDA compiler - /home/feickert/tmp/example/.pixi/envs/default/bin/nvcc
-- The CUDA compiler identification is NVIDIA 12.6.20 with host compiler GNU 13.3.0
-- Detecting CUDA compiler ABI info
-- Detecting CUDA compiler ABI info - done
-- Check for working CUDA compiler: /home/feickert/tmp/example/.pixi/envs/default/bin/nvcc - skipped
-- Detecting CUDA compile features
-- Detecting CUDA compile features - done
-- Found ZLIB: /home/feickert/tmp/example/.pixi/envs/default/lib/libz.so (found version "1.3.1")
-- Performing Test CMAKE_HAVE_LIBC_PTHREAD
-- Performing Test CMAKE_HAVE_LIBC_PTHREAD - Failed
-- Looking for pthread_create in pthreads
-- Looking for pthread_create in pthreads - not found
-- Looking for pthread_create in pthread
-- Looking for pthread_create in pthread - found
-- Found Threads: TRUE
-- Caffe2: Found protobuf with new-style protobuf targets.
-- Caffe2: Protobuf version 28.2.0
-- Found CUDA: /home/feickert/tmp/example/.pixi/envs/default (found version "12.6")
-- Found CUDAToolkit: /home/feickert/tmp/example/.pixi/envs/default/targets/x86_64-linux/include (found version "12.6.20")
-- Caffe2: CUDA detected: 12.6
-- Caffe2: CUDA nvcc is: /home/feickert/tmp/example/.pixi/envs/default/bin/nvcc
-- Caffe2: CUDA toolkit directory: /home/feickert/tmp/example/.pixi/envs/default
-- Caffe2: Header version is: 12.6
-- Found Python: /home/feickert/.pyenv/versions/3.11.7/envs/base/bin/python3.11 (found version "3.11.7") found components: Interpreter
CMake Warning at /home/feickert/tmp/example/.pixi/envs/default/share/cmake/Caffe2/public/cuda.cmake:140 (message):
  Failed to compute shorthash for libnvrtc.so
Call Stack (most recent call first):
  /home/feickert/tmp/example/.pixi/envs/default/share/cmake/Caffe2/Caffe2Config.cmake:86 (include)
  /home/feickert/tmp/example/.pixi/envs/default/share/cmake/Torch/TorchConfig.cmake:68 (find_package)
  CMakeLists.txt:54 (find_package)

CMake Warning (dev) at /home/feickert/tmp/example/.pixi/envs/default/share/cmake-3.31/Modules/FindPackageHandleStandardArgs.cmake:441 (message):
  The package name passed to `find_package_handle_standard_args` (nvtx3) does
  not match the name of the calling package (Caffe2).  This can lead to
  problems in calling code that expects `find_package` result variables
  (e.g., `_FOUND`) to follow a certain pattern.
Call Stack (most recent call first):
  /home/feickert/tmp/example/.pixi/envs/default/share/cmake/Caffe2/public/cuda.cmake:174 (find_package_handle_standard_args)
  /home/feickert/tmp/example/.pixi/envs/default/share/cmake/Caffe2/Caffe2Config.cmake:86 (include)
  /home/feickert/tmp/example/.pixi/envs/default/share/cmake/Torch/TorchConfig.cmake:68 (find_package)
  CMakeLists.txt:54 (find_package)
This warning is for project developers.  Use -Wno-dev to suppress it.

-- Found nvtx3: /usr/include
-- USE_CUDNN is set to 0. Compiling without cuDNN support
-- USE_CUSPARSELT is set to 0. Compiling without cuSPARSELt support
-- USE_CUDSS is set to 0. Compiling without cuDSS support
-- USE_CUFILE is set to 0. Compiling without cuFile support
-- Autodetected CUDA architecture(s):  8.9
-- Added CUDA NVCC flags for: -gencode;arch=compute_89,code=sm_89
-- Found Torch: /home/feickert/tmp/example/.pixi/envs/default/lib/libtorch.so
-- Configuring done (7.1s)
-- Generating done (0.0s)
-- Build files have been written to: /home/feickert/tmp/example/build
-- Caffe2: Found protobuf with new-style protobuf targets.
-- Caffe2: Protobuf version 28.2.0
-- Caffe2: CUDA detected: 12.6
-- Caffe2: CUDA nvcc is: /home/feickert/tmp/example/.pixi/envs/default/bin/nvcc
-- Caffe2: CUDA toolkit directory: /home/feickert/tmp/example/.pixi/envs/default
-- Caffe2: Header version is: 12.6
CMake Warning at /home/feickert/tmp/example/.pixi/envs/default/share/cmake/Caffe2/public/cuda.cmake:140 (message):
  Failed to compute shorthash for libnvrtc.so
Call Stack (most recent call first):
  /home/feickert/tmp/example/.pixi/envs/default/share/cmake/Caffe2/Caffe2Config.cmake:86 (include)
  /home/feickert/tmp/example/.pixi/envs/default/share/cmake/Torch/TorchConfig.cmake:68 (find_package)
  CMakeLists.txt:54 (find_package)

CMake Warning (dev) at /home/feickert/tmp/example/.pixi/envs/default/share/cmake-3.31/Modules/FindPackageHandleStandardArgs.cmake:441 (message):
  The package name passed to `find_package_handle_standard_args` (nvtx3) does
  not match the name of the calling package (Caffe2).  This can lead to
  problems in calling code that expects `find_package` result variables
  (e.g., `_FOUND`) to follow a certain pattern.
Call Stack (most recent call first):
  /home/feickert/tmp/example/.pixi/envs/default/share/cmake/Caffe2/public/cuda.cmake:174 (find_package_handle_standard_args)
  /home/feickert/tmp/example/.pixi/envs/default/share/cmake/Caffe2/Caffe2Config.cmake:86 (include)
  /home/feickert/tmp/example/.pixi/envs/default/share/cmake/Torch/TorchConfig.cmake:68 (find_package)
  CMakeLists.txt:54 (find_package)
This warning is for project developers.  Use -Wno-dev to suppress it.

-- USE_CUDNN is set to 0. Compiling without cuDNN support
-- USE_CUSPARSELT is set to 0. Compiling without cuSPARSELt support
-- USE_CUDSS is set to 0. Compiling without cuDSS support
-- USE_CUFILE is set to 0. Compiling without cuFile support
-- Autodetected CUDA architecture(s):  8.9
-- Added CUDA NVCC flags for: -gencode;arch=compute_89,code=sm_89
-- Configuring done (1.1s)
-- Generating done (0.0s)
-- Build files have been written to: /home/feickert/tmp/example/build
-- Cache values
// Path to a library.
C10_CUDA_LIBRARY:FILEPATH=/home/feickert/tmp/example/.pixi/envs/default/lib/libc10_cuda.so

// Choose the type of build, options are: None Debug Release RelWithDebInfo MinSizeRel ...
CMAKE_BUILD_TYPE:STRING=Release

// Install path prefix, prepended onto install directories.
CMAKE_INSTALL_PREFIX:PATH=/home/feickert/tmp/example/.pixi/envs/default

// Host side compiler used by NVCC
CUDA_HOST_COMPILER:FILEPATH=/home/feickert/tmp/example/.pixi/envs/default/bin/x86_64-conda-linux-gnu-cc

// Path to a program.
CUDA_NVCC_EXECUTABLE:FILEPATH=/home/feickert/tmp/example/.pixi/envs/default/bin/nvcc

//
CUDA_NVRTC_LIB:FILEPATH=/home/feickert/tmp/example/.pixi/envs/default/lib/libnvrtc.so

// Path to a file.
CUDA_SDK_ROOT_DIR:PATH=CUDA_SDK_ROOT_DIR-NOTFOUND

// Toolkit location.
CUDA_TOOLKIT_ROOT_DIR:PATH=/home/feickert/tmp/example/.pixi/envs/default

// The directory containing a CMake configuration file for Caffe2.
Caffe2_DIR:PATH=/home/feickert/tmp/example/.pixi/envs/default/share/cmake/Caffe2

// The directory containing a CMake configuration file for MKL.
MKL_DIR:PATH=MKL_DIR-NOTFOUND

// The directory containing a CMake configuration file for Protobuf.
Protobuf_DIR:PATH=/home/feickert/tmp/example/.pixi/envs/default/lib/cmake/protobuf

// Path to a library.
TORCH_LIBRARY:FILEPATH=/home/feickert/tmp/example/.pixi/envs/default/lib/libtorch.so

// The directory containing a CMake configuration file for Torch.
Torch_DIR:PATH=/home/feickert/tmp/example/.pixi/envs/default/share/cmake/Torch

// The directory containing a CMake configuration file for absl.
absl_DIR:PATH=/home/feickert/tmp/example/.pixi/envs/default/lib/cmake/absl

// Path to a library.
c10_LIBRARY:FILEPATH=/home/feickert/tmp/example/.pixi/envs/default/lib/libc10.so

// Path to a file.
nvtx3_dir:PATH=/usr/include

// The directory containing a CMake configuration file for utf8_range.
utf8_range_DIR:PATH=/home/feickert/tmp/example/.pixi/envs/default/lib/cmake/utf8_range

[ 25%] Building Fortran object CMakeFiles/ftorch.dir/ftorch.F90.o
[ 50%] Building Fortran object CMakeFiles/ftorch.dir/ftorch_test_utils.f90.o
[ 75%] Building CXX object CMakeFiles/ftorch.dir/ctorch.cpp.o
[100%] Linking CXX shared library libftorch.so
[100%] Built target ftorch
-- Install configuration: "Release"
-- Installing: /home/feickert/tmp/example/.pixi/envs/default/lib/libftorch.so
-- Set non-toolchain portion of runtime path of "/home/feickert/tmp/example/.pixi/envs/default/lib/libftorch.so" to "$ORIGIN/../lib:/lib/intel64:/lib/intel64_win:/lib/win-x64"
-- Installing: /home/feickert/tmp/example/.pixi/envs/default/include/ctorch.h
-- Installing: /home/feickert/tmp/example/.pixi/envs/default/lib/cmake/FTorch/FTorchConfig.cmake
-- Installing: /home/feickert/tmp/example/.pixi/envs/default/lib/cmake/FTorch/FTorchConfig-release.cmake
-- Installing: /home/feickert/tmp/example/.pixi/envs/default/include/ftorch/ftorch.mod
-- Installing: /home/feickert/tmp/example/.pixi/envs/default/include/ftorch/ftorch_test_utils.mod
```

Though note that the wrong Python interpreter is selected during the build

```
Found Python: /home/feickert/.pyenv/versions/3.11.7/envs/base/bin/python3.11 (found version "3.11.7") found components: Interpreter
```

instead of

```console
$ pixi list python
Package     Version  Build               Size      Kind   Source
cpython     3.12.8   py312hd8ed1ab_1     43.7 KiB  conda  cpython
python      3.12.8   h9e4cc4f_1_cpython  30.1 MiB  conda  python
python_abi  3.12     5_cp312             6.1 KiB   conda  python_abi
(base) feickert@feickert-XPS-15-9530:~/tmp/example$ pixi shell
(example) feickert@feickert-XPS-15-9530:~/tmp/example$ command -v python
/home/feickert/tmp/example/.pixi/envs/default/bin/python
(example) feickert@feickert-XPS-15-9530:~/tmp/example$ python --version --version
Python 3.12.8 | packaged by conda-forge | (main, Dec  5 2024, 14:24:40) [GCC 13.3.0]
```

@jwallwork23 @ElliottKasoar @jatkinson1000 if you add the ability to specify the Python interpreter as a CMake option (and/or have CMake check `CMAKE_PREFIX_PATH` for Python first) I think that would get people >95% of the way to a plug and play solution for getting CUDA enabled builds on Linux.

Again though, this is just suggestions and not requirements for a JOSS publicaiton.
