### 2025-01-07 c173d32
Minor docs typos and infelicities.
