### 2025-01-08 Cambridge-ICCS/FTorch#214 @matthewfeickert
@timothyas if you are still interested in a CUDA enabled setup, I have a working plug and play setup with `pixi` described in https://github.com/conda-forge/cuda-feedstock/issues/59#issuecomment-2578855942. Note that you'll need to set at the command line the CMake variables `Python_EXECUTABLE` and `nvtx3_dir`.

<summary>Example to show that setup actually has a CUDA enabled PyTorch:</summary>

After install

```console
$ curl -sLO https://raw.githubusercontent.com/matthewfeickert/nvidia-gpu-ml-library-test/refs/heads/main/torch_detect_GPU.py
$ pixi run python torch_detect_GPU.py
PyTorch build CUDA version: 12.6
PyTorch build cuDNN version: 90300
PyTorch build NCCL version: (2, 23, 4)

Number of GPUs found on system: 1

Active GPU index: 0
Active GPU name: NVIDIA GeForce RTX 4060 Laptop GPU
$ pixi shell
$ mkdir -p run && cd run
$ curl -sLO https://raw.githubusercontent.com/matthewfeickert/nvidia-gpu-ml-library-test/refs/heads/main/torch_MNIST.py
$ python torch_MNIST.py
...
Train Epoch: 14 [58880/60000 (98%)]	Loss: 0.003339
Train Epoch: 14 [59520/60000 (99%)]	Loss: 0.003421

Test set: Average loss: 0.0257, Accuracy: 9920/10000 (99%)

```

---

@jwallwork23 @ElliottKasoar @jatkinson1000 Two notes from my figuring things out on https://github.com/conda-forge/cuda-feedstock/issues/59#issuecomment-2578855942:

* It would be good unless you have a technical reason not to, to set the CMake versions to be an [allowed range](https://cmake.org/cmake/help/latest/command/cmake_minimum_required.html) rather than restricting to an old release. So replacing

https://github.com/Cambridge-ICCS/FTorch/blob/6055a7dd35e394daff469673feac063f1bfaae01/src/CMakeLists.txt#L1

with

```cmake
cmake_minimum_required(VERSION 3.15...3.31 FATAL_ERROR)
```

* Example 3 - MultiGPU's [`simplenet.py`](https://github.com/Cambridge-ICCS/FTorch/blob/6055a7dd35e394daff469673feac063f1bfaae01/examples/3_MultiGPU/simplenet.py) fails with

```pytb
RuntimeError: Expected all tensors to be on the same device, but found at least two devices, cuda:0 and cpu! (when checking argument for argument mat2 in method wrapper_CUDA_mm)
```

<summary>Reproducer and log:</summary>

```console
(base) feickert@feickert-XPS-15-9530:/tmp/example$ pixi shell
(example) feickert@feickert-XPS-15-9530:/tmp/example$ cd FTorch/examples/3_MultiGPU/
(example) feickert@feickert-XPS-15-9530:/tmp/example/FTorch/examples/3_MultiGPU$ python simplenet.py
SimpleNet forward pass on CUDA device 0
Traceback (most recent call last):
  File "/tmp/example/FTorch/examples/3_MultiGPU/simplenet.py", line 53, in <module>
    print(model(input_tensor_gpu))
          ^^^^^^^^^^^^^^^^^^^^^^^
  File "/tmp/example/.pixi/envs/default/lib/python3.12/site-packages/torch/nn/modules/module.py", line 1736, in _wrapped_call_impl
    return self._call_impl(*args, **kwargs)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/tmp/example/.pixi/envs/default/lib/python3.12/site-packages/torch/nn/modules/module.py", line 1747, in _call_impl
    return forward_call(*args, **kwargs)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/tmp/example/FTorch/examples/3_MultiGPU/simplenet.py", line 41, in forward
    return self._fwd_seq(batch)
           ^^^^^^^^^^^^^^^^^^^^
  File "/tmp/example/.pixi/envs/default/lib/python3.12/site-packages/torch/nn/modules/module.py", line 1736, in _wrapped_call_impl
    return self._call_impl(*args, **kwargs)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/tmp/example/.pixi/envs/default/lib/python3.12/site-packages/torch/nn/modules/module.py", line 1747, in _call_impl
    return forward_call(*args, **kwargs)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/tmp/example/.pixi/envs/default/lib/python3.12/site-packages/torch/nn/modules/container.py", line 250, in forward
    input = module(input)
            ^^^^^^^^^^^^^
  File "/tmp/example/.pixi/envs/default/lib/python3.12/site-packages/torch/nn/modules/module.py", line 1736, in _wrapped_call_impl
    return self._call_impl(*args, **kwargs)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/tmp/example/.pixi/envs/default/lib/python3.12/site-packages/torch/nn/modules/module.py", line 1747, in _call_impl
    return forward_call(*args, **kwargs)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/tmp/example/.pixi/envs/default/lib/python3.12/site-packages/torch/nn/modules/linear.py", line 125, in forward
    return F.linear(input, self.weight, self.bias)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
RuntimeError: Expected all tensors to be on the same device, but found at least two devices, cuda:0 and cpu! (when checking argument for argument mat2 in method wrapper_CUDA_mm)
```

Are you able to reproduce this with similar PyTorch installs?

```console
$ pixi list | grep 'python\|torch'
cpython                      3.12.8        py312hd8ed1ab_1                 43.7 KiB   conda  cpython
libtorch                     2.5.1         cuda126_mkl_h2a4acef_308        491.1 MiB  conda  libtorch
python                       3.12.8        h9e4cc4f_1_cpython              30.1 MiB   conda  python
python_abi                   3.12          5_cp312                         6.1 KiB    conda  python_abi
pytorch                      2.5.1         cuda126_mkl_py312_hdbe889e_308  35.6 MiB   conda  pytorch
torchvision                  0.20.1        cuda126_py312_h17ccbaa_4        2.7 MiB    conda  torchvision
```

---

I'm happy to give additional feedback or test things (given that I've already overstepped into reviewing a bit), but beyond that I'm going to step back and let the reviewers do their thing. :)
