### 2025-01-09 Cambridge-ICCS/FTorch#214 @ElliottKasoar
Thanks for the suggestions, @matthewfeickert!

> * It would be good, unless you have a technical reason not to, to set the CMake versions to be an [allowed range](https://cmake.org/cmake/help/latest/command/cmake_minimum_required.html) rather than restricting to an old release. So replacing
>
> https://github.com/Cambridge-ICCS/FTorch/blob/6055a7dd35e394daff469673feac063f1bfaae01/src/CMakeLists.txt#L1
>
> with
>
> ```cmake
> cmake_minimum_required(VERSION 3.15...3.31 FATAL_ERROR)
> ```

I'm probably missing something, but I'm not sure I understand your recommendation here. My understanding of `cmake_minimum_required` is that currently we specify 3.15 as the oldest compatible release, but it doesn't prevent users from using more modern releases, and don't recall encountering any issues along those lines myself.
