### 2025-01-09 Cambridge-ICCS/FTorch#214 @matthewfeickert
> My understanding of `cmake_minimum_required` is that currently we specify 3.15 as the oldest compatible release, but it doesn't prevent users from using more modern releases, and don't recall encountering any issues along those lines myself.

@ElliottKasoar No, `cmake_minimum_required(VERSION 3.15)` means "enforce the behavior policies of CMake v3.15 exactly". c.f https://cliutils.gitlab.io/modern-cmake/chapters/basics.html#id5

You want to avoid this as it both leaves bug fixes and performance improvements of modern CMake unused.
