### 2025-01-10 Cambridge-ICCS/FTorch#214 @ElliottKasoar
> > My understanding of `cmake_minimum_required` is that currently we specify 3.15 as the oldest compatible release, but it doesn't prevent users from using more modern releases, and don't recall encountering any issues along those lines myself.
>
> No, `cmake_minimum_required(VERSION 3.15)` means "enforce the behavior policies of CMake v3.15 exactly".
>
> You want to avoid this as it both leaves bug fixes and performance improvements of modern CMake unused.
>
> From the docs:
>
> > The `cmake_minimum_required(VERSION)` command implicitly invokes the [`cmake_policy(VERSION)`](https://cmake.org/cmake/help/latest/command/cmake_policy.html#version) command to specify that the current project code is written for the given range of CMake versions. All policies known to the running version of CMake and introduced in the `<min>` (or `<max>`, if specified) version or earlier will be set to use NEW behavior. All policies introduced in later versions will be unset.
>
> c.f https://cliutils.gitlab.io/modern-cmake/chapters/basics.html for more info.
>
> (Also you [don't need `FATAL_ERROR` anymore](https://cliutils.gitlab.io/modern-cmake/chapters/basics.html#id5), which I forgot.)

Ah, I was too focused on whether errors would be raised. That makes sense, thanks!
