### 2025-01-10 Cambridge-ICCS/FTorch#214 @jatkinson1000
To avoid this merging off-topic from the original point of discussion (conda install and environments) I have opened #226 to discuss CMake versioning syntax.  Please move discussion there and add anything else you feel may be required.
