### 2025-01-10 Cambridge-ICCS/FTorch#214 @jatkinson1000
Similarly I have moved the query around example 3 to a new issue #227 @matthewfeickert
