### 2025-01-10 a4378eb
Set upper CMake minimum version in CMakeLists across codebase.
