### 2025-01-10 5d19407
Set upper CMake minimum version in CMakeLists across codebase.
