### 2025-01-10 6ab8207
Add a note to the API updates documentation page to state the commit from which it applies and that new users do not need to worry about it.
