+++ pages/updates.md
@@ -1,11 +1,18 @@
 title: Recent API Changes
 
+If you use a version of FTorch from before commit 
+[e92ad9e](https://github.com/Cambridge-ICCS/FTorch/commit/e92ad9ec7c2198dbb2ca819854d604b984d293c4)
+(June 2024) you will notice that the latest API documentation is not suitable.
+This is because a number of breaking changes were made to the FTorch API in preparation
+for implementing new functionalities.
+
+This page describes how to migrate from code (pre-e92ad9e) to the most recent version.
+
+If you are already using a more recent version there is no need to read this page.
+
 [TOC]
 
-
 ## Why?
-
-Recently we made a number of breaking changes to the FTorch API.
 
 We realise that this forms an inconvenience to those of you who are actively
 using FTorch and is not something we did lightly.
