### 2025-01-13 1ad2087
#229: Drop n_c_and_cpp example for now
