### 2025-01-13 abf78a3
Drop debug print
