### 2025-01-13 e32b896
Support float exponent as well as int
