+++ src/ctorch.cpp
@@ -287,11 +287,22 @@ torch_tensor_t torch_tensor_divide(const torch_tensor_t tensor1,
   return output;
 }
 
-torch_tensor_t torch_tensor_power(const torch_tensor_t tensor,
-                                  const torch_scalar_t exponent) {
+torch_tensor_t torch_tensor_power_int(const torch_tensor_t tensor,
+                                      const torch_int_t exponent) {
   auto t = reinterpret_cast<torch::Tensor *const>(tensor);
   // NOTE: The following cast will only work for integer exponents
   auto exp = reinterpret_cast<int *const>(exponent);
+  torch::Tensor *output = nullptr;
+  output = new torch::Tensor;
+  *output = pow(*t, *exp);
+  return output;
+}
+
+torch_tensor_t torch_tensor_power_float(const torch_tensor_t tensor,
+                                        const torch_float_t exponent) {
+  auto t = reinterpret_cast<torch::Tensor *const>(tensor);
+  // NOTE: The following cast will only work for floating point exponents
+  auto exp = reinterpret_cast<float *const>(exponent);
   torch::Tensor *output = nullptr;
   output = new torch::Tensor;
   *output = pow(*t, *exp);
