### 2025-01-13 7497aff
Move cleanup subroutine and checks that input arrays aren't changed by arithmetic
