### 2025-01-13 8a81b22
Move cleanup subroutine and checks that input arrays aren't changed by arithmetic
