+++ examples/6_Autograd/autograd.f90
@@ -67,39 +67,13 @@ program example
     stop 999
   end if
 
-  ! Check first input array is unchanged by the arithmetic operations
-  expected(:,1) = [2.0_wp, 3.0_wp]
-  test_pass = assert_allclose(in_data1, expected, test_name="torch_tensor_to_array", rtol=1e-5)
-  if (.not. test_pass) then
-    call clean_up()
-    print *, "Error :: in_data1 was changed during arithmetic operations"
-    stop 999
-  end if
-
-  ! Check second input array is unchanged by the arithmetic operations
-  expected(:,1) = [6.0_wp, 4.0_wp]
-  test_pass = assert_allclose(in_data2, expected, test_name="torch_tensor_to_array", rtol=1e-5)
-  if (.not. test_pass) then
-    call clean_up()
-    print *, "Error :: in_data2 was changed during arithmetic operations"
-    stop 999
-  end if
-
   ! Back-propagation
   ! TODO: Requires API extension
 
-  ! Cleanup
-  call clean_up()
+  nullify(out_data)
+  call torch_tensor_delete(a)
+  call torch_tensor_delete(b)
+  call torch_tensor_delete(Q)
   write (*,*) "Autograd example ran successfully"
-
-  contains
-
-    ! Subroutine for freeing memory and nullifying pointers used in the example
-    subroutine clean_up()
-      nullify(out_data)
-      call torch_tensor_delete(a)
-      call torch_tensor_delete(b)
-      call torch_tensor_delete(Q)
-    end subroutine clean_up
 
 end program example
