### 2025-01-13 e03f75c
Post-rebase fixes
