+++ src/ftorch.F90
@@ -33,11 +33,6 @@ module ftorch
     procedure :: get_shape
   end type torch_tensor
 
-  !> Type for holding a Torch scalar.
-  type torch_scalar
-    type(c_ptr) :: p = c_null_ptr  !! pointer to the scalar in memory
-  end type torch_scalar
-
   !| Enumerator for Torch data types
   !  From c_torch.h (torch_data_t)
   !  Note that 0 `torch_kUInt8` and 5 `torch_kFloat16` are not sypported in Fortran
@@ -2445,6 +2440,7 @@ contains
   !> Overloads multiplication operator for a scalar of type int8 and a tensor.
   function torch_tensor_premultiply_int8(scalar, tensor) result(output)
     use, intrinsic :: iso_c_binding, only : c_loc
+    use, intrinsic :: iso_fortran_env, only : int8
     integer(int8), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
     type(torch_tensor) :: wrk, output
@@ -2457,6 +2453,7 @@ contains
   !> Overloads multiplication operator for a scalar of type int16 and a tensor.
   function torch_tensor_premultiply_int16(scalar, tensor) result(output)
     use, intrinsic :: iso_c_binding, only : c_loc
+    use, intrinsic :: iso_fortran_env, only : int16
     integer(int16), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
     type(torch_tensor) :: wrk, output
@@ -2469,6 +2466,7 @@ contains
   !> Overloads multiplication operator for a scalar of type int32 and a tensor.
   function torch_tensor_premultiply_int32(scalar, tensor) result(output)
     use, intrinsic :: iso_c_binding, only : c_loc
+    use, intrinsic :: iso_fortran_env, only : int32
     integer(int32), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
     type(torch_tensor) :: wrk, output
@@ -2481,6 +2479,7 @@ contains
   !> Overloads multiplication operator for a scalar of type int64 and a tensor.
   function torch_tensor_premultiply_int64(scalar, tensor) result(output)
     use, intrinsic :: iso_c_binding, only : c_loc
+    use, intrinsic :: iso_fortran_env, only : int64
     integer(int64), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
     type(torch_tensor) :: wrk, output
@@ -2493,6 +2492,7 @@ contains
   !> Overloads multiplication operator for a scalar of type real32 and a tensor.
   function torch_tensor_premultiply_real32(scalar, tensor) result(output)
     use, intrinsic :: iso_c_binding, only : c_loc
+    use, intrinsic :: iso_fortran_env, only : real32
     real(real32), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
     type(torch_tensor) :: wrk, output
@@ -2505,6 +2505,7 @@ contains
   !> Overloads multiplication operator for a scalar of type real64 and a tensor.
   function torch_tensor_premultiply_real64(scalar, tensor) result(output)
     use, intrinsic :: iso_c_binding, only : c_loc
+    use, intrinsic :: iso_fortran_env, only : real64
     real(real64), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
     type(torch_tensor) :: wrk, output
@@ -2518,6 +2519,7 @@ contains
   !> Overloads multiplication operator for a tensor and a scalar of type int8.
   function torch_tensor_postmultiply_int8(tensor, scalar) result(output)
     use, intrinsic :: iso_c_binding, only : c_loc
+    use, intrinsic :: iso_fortran_env, only : int8
[…]
