### 2025-01-13 d5046a3
Unit tests for operator overloading; fix scalar bug (#225)
* Unit tests for operator overloads
* Drop pre/post multiply/divide, just pass trivial tensors
* Reuse multiply and divide interfaces
