+++ src/ftorch.F90
@@ -200,6 +200,17 @@ module ftorch
     module procedure torch_tensor_postmultiply_real64
   end interface
 
+  interface
+    function torch_tensor_multiply_c(tensor1_c, tensor2_c) result(output_c)  &
+        bind(c, name = 'torch_tensor_multiply')
+      use, intrinsic :: iso_c_binding, only : c_ptr
+      implicit none
+      type(c_ptr), value, intent(in) :: tensor1_c
+      type(c_ptr), value, intent(in) :: tensor2_c
+      type(c_ptr) :: output_c
+    end function torch_tensor_multiply_c
+  end interface
+
   interface operator (/)
     module procedure torch_tensor_divide
     module procedure torch_tensor_postdivide_int8
@@ -208,6 +219,17 @@ module ftorch
     module procedure torch_tensor_postdivide_int64
     module procedure torch_tensor_postdivide_real32
     module procedure torch_tensor_postdivide_real64
+  end interface
+
+  interface
+    function torch_tensor_divide_c(tensor1_c, tensor2_c) result(output_c)  &
+        bind(c, name = 'torch_tensor_divide')
+      use, intrinsic :: iso_c_binding, only : c_ptr
+      implicit none
+      type(c_ptr), value, intent(in) :: tensor1_c
+      type(c_ptr), value, intent(in) :: tensor2_c
+      type(c_ptr) :: output_c
+    end function torch_tensor_divide_c
   end interface
 
   interface operator (**)
@@ -2412,271 +2434,164 @@ contains
     type(torch_tensor), intent(in) :: tensor2
     type(torch_tensor) :: output
 
-    interface
-      function torch_tensor_multiply_c(tensor1_c, tensor2_c) result(output_c)  &
-          bind(c, name = 'torch_tensor_multiply')
-        use, intrinsic :: iso_c_binding, only : c_ptr
-        implicit none
-        type(c_ptr), value, intent(in) :: tensor1_c
-        type(c_ptr), value, intent(in) :: tensor2_c
-        type(c_ptr) :: output_c
-      end function torch_tensor_multiply_c
-    end interface
-
     output%p = torch_tensor_multiply_c(tensor1%p, tensor2%p)
   end function torch_tensor_multiply
 
   !> Overloads multiplication operator for a scalar of type int8 and a tensor.
   function torch_tensor_premultiply_int8(scalar, tensor) result(output)
+    use, intrinsic :: iso_c_binding, only : c_loc
     use, intrinsic :: iso_fortran_env, only : int8
-    integer(int8), intent(in) :: scalar
+    integer(int8), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
-    type(torch_tensor) :: output
+    type(torch_tensor) :: wrk, output
 
-    interface
-      function torch_tensor_premultiply_c(scalar_c, tensor_c) result(output_c) &
-          bind(c, name = 'torch_tensor_premultiply')
-        use, intrinsic :: iso_c_binding, only : c_ptr, c_int8_t
-        implicit none
-        integer(kind=c_int8_t), value, intent(in) :: scalar_c
-        type(c_ptr), value, intent(in) :: tensor_c
-        type(c_ptr) :: output_c
-      end function torch_tensor_premultiply_c
-    end interface
-
-    output%p = torch_tensor_premultiply_c(scalar, tensor%p)
+    ! Create a tensor with a single entry, the scalar pre-multiplier
+    call torch_tensor_from_array(wrk, [scalar], [1], torch_kCPU)
+    output%p = torch_tensor_multiply_c(wrk%p, tensor%p)
   end function torch_tensor_premultiply_int8
 
[…]
