### 2025-01-13 0b1c89d
Don't drop clean_up subroutine
