### 2025-01-13 b63f828
Don't drop clean_up subroutine
