### 2025-01-13 5f4418a
Put model on CUDA device in simplenet
