### 2025-01-13 e16c17d
Run example 3 if it's been built
