+++ run_test_suite.sh
@@ -82,7 +82,11 @@ fi
 
 # Run integration tests
 if [ "${RUN_INTEGRATION}" = true ]; then
-  EXAMPLES="1_SimpleNet 2_ResNet18 4_MultiIO 6_Autograd"
+  if [ -e "${BUILD_DIR}/test/examples/3_MultiGPU" ]; then
+    EXAMPLES="1_SimpleNet 2_ResNet18 3_MultiGPU 4_MultiIO 6_Autograd"
+  else
+    EXAMPLES="1_SimpleNet 2_ResNet18 4_MultiIO 6_Autograd"
+  fi
   for EXAMPLE in ${EXAMPLES}; do
     pip -q install -r examples/"${EXAMPLE}"/requirements.txt
     cd "${BUILD_DIR}"/test/examples/"${EXAMPLE}"
