+++ examples/3_MultiGPU/pt2ts.py
@@ -1,5 +1,7 @@
 """Load a PyTorch model and convert it to TorchScript."""
 
+import os
+import sys
 from typing import Optional
 
 # FPTLIB-TODO
