### 2025-01-13 b4246c4
Fix numbering in CMakeLists for examples
