### 2025-01-13 28e3778
Renaming in MultiGPU example; set up unit testing
