### 2025-01-13 6a532b5
Renaming in MultiGPU example; set up unit testing
