### 2025-01-13 5fa7801
Raise error if no CUDA in example 3
