### 2025-01-13 bec0734
Fix model filename passed to fortran
