### 2025-01-13 53ab151
Note on additional dependencies for test suite
