### 2025-01-13 Cambridge-ICCS/FTorch#214 @jatkinson1000
As an update here I have been looking at building using conda.
To start off I am looking at CPU only and building and running FTorch with tests.

A conda environment genreated from the following yaml works for me:
```yaml
name: FTorch-conda-cpu
channels:
  - conda-forge
dependencies:
  - fortran-compiler
  - cxx-compiler
  - cmake
  - pytorch
  - torchaudio # Not sure this is actually needed...
  - torchvision # For example 2 only?
  - openmpi # For pFUnit
```

and I then build and run tests (note intermediate building of pFUnit which takes a while) with:
```bash
git clone --recursive git@github.com:Goddard-Fortran-Ecosystem/pFUnit.git
cd pFUnit
mkdir build
cd build
cmake \
  -DCMAKE_INSTALL_PREFIX=$CONDA_PREFIX \
  ..
make tests
make install

cd ../../

git clone git@github.com:Cambridge-ICCS/FTorch.git

cd FTorch
mkdir build
cd build

cmake \
    -DCMAKE_INSTALL_PREFIX=$CONDA_PREFIX \
    -DCMAKE_PREFIX_PATH=$CONDA_PREFIX \
    -DCMAKE_BUILD_TYPE=Release \
    -DENABLE_CUDA=FALSE \
    -DCMAKE_BUILD_TESTS=TRUE \
    ..
cmake --build . --target install

cd ../../
./run_test_suite.sh
```

@matthewfeickert This did not raise an issiue of a missing zlib for me, though perhaps this will occur in the CUDA build which I plan to look at now I have this working.

As a further question, since I am not overly familiar with conda, how would one usually supply this information in a repository @timothyas? Would providing a set of `environment.yml` files (one for CPU-only, one for cuda etc) be sufficient, perhaps with accompanying scripts that would install FTorch into `$CONDA_PREFIX`?
