### 2025-01-13 Cambridge-ICCS/FTorch#214 @matthewfeickert
> @matthewfeickert This did not raise an issue of a missing zlib for me, though perhaps this will occur in the CUDA build which I plan to look at now I have this working.

```
Could NOT find ZLIB (missing: ZLIB_LIBRARY)
```

doesn't raise as an error, but I noticed it in the logs. You might want to check what `ZLIB_LIBRARY` is being set to.

> As a further question, since I am not overly familiar with conda, how would one usually supply this information in a repository @timothyas? Would providing a set of `environment.yml` files (one for CPU-only, one for cuda etc) be sufficient, perhaps with accompanying scripts that would install FTorch into `$CONDA_PREFIX`?

You could do that and/or define a `pixi.toml` at the repo top level that [defines CPU and GPU environments](https://pixi.sh/latest/features/multi_environment/#ambiguous-environment-selection) and the equivalent [tasks](https://pixi.sh/latest/features/advanced_tasks/) that the scripts would cover.
