### 2025-01-13 Cambridge-ICCS/FTorch#214 @jatkinson1000
### on zlib:
Ahhhhh, I think I understand - conda operates like a valve to the wider system, similar to pip.
We do not make use of zlib anywhere in our library, and `ZLIB_LIBRARY` is not in our CMakeLists. As data compression my guess would be that it is used somewhere as part of PyTorch, likely Torchvision, in which case I would have thought it ought to be pulled in as a dependency of that? Though perhaps I am missing something as I am coming new to conda.

I did a bit of digging and found a couple of issues referencing this on the pytorch organisation e.g. https://github.com/pytorch/vision/issues/4492 where they seem to be relying on the assumption that zlib will be present on a unix based system, and that Windows users should explicitly install it themselves.

Do you have any advice here, as I feel this is not a missing explicit dependency on our part so am hesitant about whether we should be specifying it in our list or if this would be misleading. I'm not really sure what to do regarding missing dependencies of dependencies.
I can see that it would make sense to add it to a conda environment file to ensure that we are covered.

### on providing build information:

I had a look at pixi, but my reading was that a `pixi.toml` would be specific to pixi and not generaliseable to conda users, unless there is a way of generating a conda environment from a `pixi.toml`?
Whilst it looks like a nice way to manage projects, with a focus on project rather than environment, I am wary of getting caught up in supporting a myriad of different build systems, especially if we do not have expertise in them ourselves (PRs always welcome though 😉).

I would like to read more on it if I had some time, however.
