### 2025-01-13 Cambridge-ICCS/FTorch#214 @matthewfeickert
> Do you have any advice here, as I feel this is not a missing explicit dependency on our part so am hesitant about whether we should be specifying it in our list or if this would be misleading. I'm not really sure what to do regarding missing dependencies of dependencies.

I will defer to you here. `libzlib` is a dependency of dependencies (this is coming from observing the structure of `pixi tree --invert libzlib`) but I agree that I don't think it would be wrong to not mention it as a build dependency, as you note that you do not directly rely on `ZLIB_LIBRARY` anywhere. So it is fully valid to say that this is something that might need to be explained if people have bugs during builds but that you don't need to document as a FTorch requirement anywhere. :+1:

> I had a look at pixi, but my reading was that a pixi.toml would be specific to pixi and not generaliseable to conda users

You're correct that `pixi`'s manifest and lock files are currently used only by `pixi`. The same could be said for `environment.yaml` being exclusive to `conda`/`mamaba`.

> unless there is a way of generating a conda environment from a pixi.toml?

[There is](https://pixi.sh/latest/reference/cli/#project-export-conda-environment):

```
pixi project export conda-environment environment.yml
```

I may have a different and more utilitarian view, but I've never thought that `conda`/`mamba` users loved the tool or the environment management but rather the "I can get binaries on my machine that are known to work fast" component of it. I've always viewed the thing that people are most interested in is the packages on conda-forge and being able to get them working on their machines. To me, `pixi` enables this better on the user and maintainer side than any other tool.

All that being said, you should do whatever you think is the most useful to your project. That also very well might be not spending maintainer time supporting any tools that talk to conda-forge, and that would be perfectly reasonable. :+1: You know your users far better than anyone else, and supporting yourselves as maintainers and developers and your users comes far ahead of any suggestions that might come up.
