### 2025-01-13 Cambridge-ICCS/FTorch#214 @timothyas
Hey folks, just wanted to pop in and say sorry for the radio silence on my end - deadlines last week, sick toddler this week. I'll try @matthewfeickert's recommendations when I get a chance just to try a gpu setup and report back (also with some what worked/didn't work). Thanks for the good discussion
