### 2025-01-14 90a51ff
strides should be initialised to zero; tensor_shape and layout can be assumed length
