### 2025-01-15 e2eab96
Give run_test_suite ability to only run unit tests
