### 2025-01-17 8f829f2
Add Spack configuration file
