### 2025-01-17 c4c0888
Add Spack configuration file
