+++ spack.yaml
@@ -1,0 +1,12 @@
+spack:
+  packages:
+    all:
+      compiler: [gcc@11.4.0]
+  specs:
+  - cmake
+  - py-torch
+  - pfunit
+  concretizer:
+    unify: true
+  config:
+    install_missing_compilers: true
