### 2025-01-17 9810b4e
Don't install MPI or OpenMP variants
