### 2025-01-20 ef8d8d5
Drop mkldnn; add py-torchvision
