### 2025-01-20 243ff8a
fix broken url to compiled version in readme for developer docs + nit
