### 2025-01-20 5522fca
Move input array checks out of autograd example (#231)
* Add missing summary of autograd to examples README
* Move cleanup subroutine and checks that input arrays aren't changed by arithmetic
* Call cleanup if assertion fails in operator overload tests
* Call cleanup if assertion fails in constructor tests
