### 2025-01-20 2135968
Implement torch_tensor_backward
