### 2025-01-20 2966702
Implement torch_tensor_backward
