### 2025-01-20 24559fe
Simplify autograd example
