### 2025-01-20 49da545
Simplify autograd example
