### 2025-01-20 8b05805
Setup requires_grad properly; Use TensorOptions in tensor constructors
