### 2025-01-20 be86846
Setup requires_grad properly; Use TensorOptions in tensor constructors
