### 2025-01-20 45cafe1
DO NOT MERGE - restrict to autograd example
