### 2025-01-20 c5ca8df
Implement get_gradient
