+++ src/ctorch.cpp
@@ -359,6 +359,14 @@ void torch_tensor_backward(const torch_tensor_t tensor,
   t->backward(*g);
 }
 
+EXPORT_C torch_tensor_t get_gradient(const torch_tensor_t tensor) {
+  auto t = reinterpret_cast<torch::Tensor *const>(tensor);
+  torch::Tensor *output = nullptr;
+  output = new torch::Tensor;
+  *output = t->grad();
+  return output;
+}
+
 // =============================================================================
 // --- Torch model API
 // =============================================================================
