### 2025-01-20 10d2499
Implement get_gradient
