+++ src/ctorch.cpp
@@ -268,7 +268,7 @@ torch_tensor_t torch_tensor_add(const torch_tensor_t tensor1,
   return output;
 }
 
-torch_tensor_t torch_tensor_minus(const torch_tensor_t tensor) {
+torch_tensor_t torch_tensor_negative(const torch_tensor_t tensor) {
   auto t = reinterpret_cast<torch::Tensor *const>(tensor);
   torch::Tensor *output = nullptr;
   output = new torch::Tensor;
