+++ src/ctorch.h
@@ -176,6 +176,13 @@ EXPORT_C torch_tensor_t torch_tensor_add(const torch_tensor_t tensor1,
                                          const torch_tensor_t tensor2);
 
 /**
+ * Overloads the minus operator for a single Torch Tensor
+ * @param Tensor to take the negative of
+ * @return the negative Tensor
+ */
+EXPORT_C torch_tensor_t torch_tensor_negative(const torch_tensor_t tensor);
+
+/**
  * Overloads the subtraction operator for two Torch Tensors
  * @param first Tensor to be subtracted
  * @param second Tensor to be subtracted
