### 2025-01-23 b21b15e
Option for running integration tests only
