### 2025-01-23 c8763c0
Spell out command line options in CI
Co-authored-by: Jack Atkinson <109271713+jatkinson1000@users.noreply.github.com>
