### 2025-01-23 8062108
Add missing items from help text
