### 2025-01-24 5b095f7
Add unit tests for tensor interrogation (#239)
* Give run_test_suite ability to only run unit tests
* Add unit tests for get_rank method
* Add unit tests for get_shape method
* Add unit test for torch_tensor_get_device_index on CPU
* Add unit test for torch_tensor_get_device_index on CUDA device
* Remove rank and shape checks from autograd example
* Option for running integration tests only
* Separate out unit test and integration test jobs in Ubuntu CI
* Spell out command line options in CI
* Add missing items from help text
* Improve docs on testing

---------

Co-authored-by: Jack Atkinson 
[…]
