+++ run_test_suite.sh
@@ -14,22 +14,35 @@ set -eu
 
 # Function to display help text
 show_help() {
-  echo "Usage: $0 [BUILD_DIR=<build_dir>] [--verbose | -V]"
+  echo "Usage: $0 [BUILD_DIR=<build_dir>] [--integration-only | i] [--unit-only | -u]"
+  echo "          [--verbose | -V] [--help | h]"
   echo
   echo "Options:"
   echo "  BUILD_DIR=<build_dir> Specify the build directory (default: src/build)."
-  echo "  --verbose | -V        Run with verbose ctest output."
-  echo "  --help | -h           Show this help message and exit."
+  echo "  --integration-only | -i        Run integration tests only."
+  echo "  --unit-only        | -u        Run unit tests only."
+  echo "  --verbose          | -V        Run with verbose ctest output."
+  echo "  --help             | -h        Show this help message and exit."
 }
 
 # Parse command line arguments
 BUILD_DIR="src/build"
+INTEGRATION_ONLY=false
+UNIT_ONLY=false
 VERBOSE=false
 HELP=false
 for ARG in "$@"; do
   case ${ARG} in
   BUILD_DIR=*)
     BUILD_DIR="${ARG#*=}"
+    ;;
+  --integration-only | -i)
+    INTEGRATION_ONLY=true
+    shift
+    ;;
+  --unit-only | -u)
+    UNIT_ONLY=true
+    shift
     ;;
   --verbose | -V)
     VERBOSE=true
@@ -61,15 +74,19 @@ else
 fi
 
 # Run unit tests
-cd "${BUILD_DIR}/test/unit"
-ctest "${CTEST_ARGS}"
-cd -
-
-# Run integration tests
-EXAMPLES="1_SimpleNet 2_ResNet18 4_MultiIO 6_Autograd"
-for EXAMPLE in ${EXAMPLES}; do
-  pip -q install -r examples/"${EXAMPLE}"/requirements.txt
-  cd "${BUILD_DIR}"/test/examples/"${EXAMPLE}"
+if [ "${INTEGRATION_ONLY}" = false ]; then
+  cd "${BUILD_DIR}/test/unit"
   ctest "${CTEST_ARGS}"
   cd -
-done
+fi
+
+# Run integration tests
+if [ "${UNIT_ONLY}" = false ]; then
+  EXAMPLES="1_SimpleNet 2_ResNet18 4_MultiIO 6_Autograd"
+  for EXAMPLE in ${EXAMPLES}; do
+    pip -q install -r examples/"${EXAMPLE}"/requirements.txt
+    cd "${BUILD_DIR}"/test/examples/"${EXAMPLE}"
+    ctest "${CTEST_ARGS}"
+    cd -
+  done
+fi
