+++ .github/workflows/test_suite_ubuntu.yml
@@ -99,7 +99,12 @@ jobs:
           cmake --build .
           cmake --install .
 
-      - name: Test suite
+      - name: Run unit tests
         run: |
           . ftorch/bin/activate
-          ./run_test_suite.sh -V
+          ./run_test_suite.sh --verbose --unit-only
+
+      - name: Run integration tests
+        run: |
+          . ftorch/bin/activate
+          ./run_test_suite.sh --verbose --integration-only
