### 2025-01-24 6c5cef8
Add dtype and device_type attrs for torch_tensor; implement getters
