### 2025-01-24 0d8ffd3
Rename get_<rank/shape> as torch_tensor_get_<rank/shape> for consistency
