+++ src/ftorch.F90
@@ -31,8 +31,8 @@ module ftorch
     integer, private :: dtype
     integer, private :: device_type
   contains
-    procedure :: get_rank
-    procedure :: get_shape
+    procedure :: torch_tensor_get_rank
+    procedure :: torch_tensor_get_shape
     procedure :: torch_tensor_get_dtype
     procedure :: torch_tensor_get_device_type
   end type torch_tensor
@@ -1381,7 +1381,7 @@ contains
     integer(c_int), parameter :: dtype = torch_kInt8  !! Data type
     type(c_ptr) :: cptr
 
-    my_shape = tensor%get_shape()
+    my_shape = tensor%torch_tensor_get_shape()
 
     if (present(sizes)) then
       if (.not. all(my_shape == sizes)) then
@@ -1411,7 +1411,7 @@ contains
     integer(c_int), parameter :: dtype = torch_kInt8  !! Data type
     type(c_ptr) :: cptr
 
-    my_shape = tensor%get_shape()
+    my_shape = tensor%torch_tensor_get_shape()
 
     if (present(sizes)) then
       if (.not. all(my_shape == sizes)) then
@@ -1441,7 +1441,7 @@ contains
     integer(c_int), parameter :: dtype = torch_kInt8  !! Data type
     type(c_ptr) :: cptr
 
-    my_shape = tensor%get_shape()
+    my_shape = tensor%torch_tensor_get_shape()
 
     if (present(sizes)) then
       if (.not. all(my_shape == sizes)) then
@@ -1471,7 +1471,7 @@ contains
     integer(c_int), parameter :: dtype = torch_kInt8  !! Data type
     type(c_ptr) :: cptr
 
-    my_shape = tensor%get_shape()
+    my_shape = tensor%torch_tensor_get_shape()
 
     if (present(sizes)) then
       if (.not. all(my_shape == sizes)) then
@@ -1501,7 +1501,7 @@ contains
     integer(c_int), parameter :: dtype = torch_kInt8  !! Data type
     type(c_ptr) :: cptr
 
-    my_shape = tensor%get_shape()
+    my_shape = tensor%torch_tensor_get_shape()
 
     if (present(sizes)) then
       if (.not. all(my_shape == sizes)) then
@@ -1531,7 +1531,7 @@ contains
     integer(c_int), parameter :: dtype = torch_kInt16  !! Data type
     type(c_ptr) :: cptr
 
-    my_shape = tensor%get_shape()
+    my_shape = tensor%torch_tensor_get_shape()
 
     if (present(sizes)) then
       if (.not. all(my_shape == sizes)) then
@@ -1561,7 +1561,7 @@ contains
     integer(c_int), parameter :: dtype = torch_kInt16  !! Data type
     type(c_ptr) :: cptr
 
-    my_shape = tensor%get_shape()
+    my_shape = tensor%torch_tensor_get_shape()
 
     if (present(sizes)) then
       if (.not. all(my_shape == sizes)) then
@@ -1591,7 +1591,7 @@ contains
     integer(c_int), parameter :: dtype = torch_kInt16  !! Data type
     type(c_ptr) :: cptr
 
-    my_shape = tensor%get_shape()
+    my_shape = tensor%torch_tensor_get_shape()
 
     if (present(sizes)) then
       if (.not. all(my_shape == sizes)) then
@@ -1621,7 +1621,7 @@ contains
     integer(c_int), parameter :: dtype = torch_kInt16  !! Data type
     type(c_ptr) :: cptr
 
-    my_shape = tensor%get_shape()
+    my_shape = tensor%torch_tensor_get_shape()
 
     if (present(sizes)) then
[…]
