### 2025-01-24 c4974a4
Make torch_tensor_get_device_index a class method
