### 2025-01-24 e046c9e
Make use of getters for device type and index
