+++ src/ftorch.F90
@@ -2493,9 +2493,13 @@ contains
     integer(int8), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
     type(torch_tensor) :: wrk, output
+    integer :: device_type
+    integer :: device_index
 
     ! Create a tensor with a single entry, the scalar pre-multiplier
-    call torch_tensor_from_array(wrk, [scalar], [1], torch_kCPU)
+    device_type = tensor%torch_tensor_get_device_type()
+    device_index = tensor%torch_tensor_get_device_index()
+    call torch_tensor_from_array(wrk, [scalar], [1], device_type, device_index)
     output%p = torch_tensor_multiply_c(wrk%p, tensor%p)
   end function torch_tensor_premultiply_int8
 
@@ -2506,9 +2510,13 @@ contains
     integer(int16), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
     type(torch_tensor) :: wrk, output
+    integer :: device_type
+    integer :: device_index
 
     ! Create a tensor with a single entry, the scalar pre-multiplier
-    call torch_tensor_from_array(wrk, [scalar], [1], torch_kCPU)
+    device_type = tensor%torch_tensor_get_device_type()
+    device_index = tensor%torch_tensor_get_device_index()
+    call torch_tensor_from_array(wrk, [scalar], [1], device_type, device_index)
     output%p = torch_tensor_multiply_c(wrk%p, tensor%p)
   end function torch_tensor_premultiply_int16
 
@@ -2519,9 +2527,13 @@ contains
     integer(int32), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
     type(torch_tensor) :: wrk, output
+    integer :: device_type
+    integer :: device_index
 
     ! Create a tensor with a single entry, the scalar pre-multiplier
-    call torch_tensor_from_array(wrk, [scalar], [1], torch_kCPU)
+    device_type = tensor%torch_tensor_get_device_type()
+    device_index = tensor%torch_tensor_get_device_index()
+    call torch_tensor_from_array(wrk, [scalar], [1], device_type, device_index)
     output%p = torch_tensor_multiply_c(wrk%p, tensor%p)
   end function torch_tensor_premultiply_int32
 
@@ -2532,9 +2544,13 @@ contains
     integer(int64), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
     type(torch_tensor) :: wrk, output
+    integer :: device_type
+    integer :: device_index
 
     ! Create a tensor with a single entry, the scalar pre-multiplier
-    call torch_tensor_from_array(wrk, [scalar], [1], torch_kCPU)
+    device_type = tensor%torch_tensor_get_device_type()
+    device_index = tensor%torch_tensor_get_device_index()
+    call torch_tensor_from_array(wrk, [scalar], [1], device_type, device_index)
     output%p = torch_tensor_multiply_c(wrk%p, tensor%p)
   end function torch_tensor_premultiply_int64
 
@@ -2545,9 +2561,13 @@ contains
     real(real32), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
     type(torch_tensor) :: wrk, output
+    integer :: device_type
+    integer :: device_index
 
     ! Create a tensor with a single entry, the scalar pre-multiplier
-    call torch_tensor_from_array(wrk, [scalar], [1], torch_kCPU)
+    device_type = tensor%torch_tensor_get_device_type()
+    device_index = tensor%torch_tensor_get_device_index()
+    call torch_tensor_from_array(wrk, [scalar], [1], device_type, device_index)
[…]
