### 2025-01-24 6e4b218
Alias methods to be less verbose
