+++ src/ftorch.F90
@@ -27,15 +27,15 @@ module ftorch
 
   !> Type for holding a Torch tensor.
   type torch_tensor
-    type(c_ptr) :: p = c_null_ptr  !! pointer to the tensor in memory
-    integer, private :: dtype
-    integer, private :: device_type
+    type(c_ptr) :: p = c_null_ptr    !! pointer to the tensor in memory
+    integer, private :: dtype        !! integer representing the tensor's data type
+    integer, private :: device_type  !! integer representing the tensor's device type
   contains
-    procedure :: torch_tensor_get_rank
-    procedure :: torch_tensor_get_shape
-    procedure :: torch_tensor_get_dtype
-    procedure :: torch_tensor_get_device_type
-    procedure :: torch_tensor_get_device_index
+    procedure :: get_rank => torch_tensor_get_rank
+    procedure :: get_shape => torch_tensor_get_shape
+    procedure :: get_dtype => torch_tensor_get_dtype
+    procedure :: get_device_type => torch_tensor_get_device_type
+    procedure :: get_device_index => torch_tensor_get_device_index
   end type torch_tensor
 
   !| Enumerator for Torch data types
@@ -1382,7 +1382,7 @@ contains
     integer(c_int), parameter :: dtype = torch_kInt8  !! Data type
     type(c_ptr) :: cptr
 
-    my_shape = tensor%torch_tensor_get_shape()
+    my_shape = tensor%get_shape()
 
     if (present(sizes)) then
       if (.not. all(my_shape == sizes)) then
@@ -1412,7 +1412,7 @@ contains
     integer(c_int), parameter :: dtype = torch_kInt8  !! Data type
     type(c_ptr) :: cptr
 
-    my_shape = tensor%torch_tensor_get_shape()
+    my_shape = tensor%get_shape()
 
     if (present(sizes)) then
       if (.not. all(my_shape == sizes)) then
@@ -1442,7 +1442,7 @@ contains
     integer(c_int), parameter :: dtype = torch_kInt8  !! Data type
     type(c_ptr) :: cptr
 
-    my_shape = tensor%torch_tensor_get_shape()
+    my_shape = tensor%get_shape()
 
     if (present(sizes)) then
       if (.not. all(my_shape == sizes)) then
@@ -1472,7 +1472,7 @@ contains
     integer(c_int), parameter :: dtype = torch_kInt8  !! Data type
     type(c_ptr) :: cptr
 
-    my_shape = tensor%torch_tensor_get_shape()
+    my_shape = tensor%get_shape()
 
     if (present(sizes)) then
       if (.not. all(my_shape == sizes)) then
@@ -1502,7 +1502,7 @@ contains
     integer(c_int), parameter :: dtype = torch_kInt8  !! Data type
     type(c_ptr) :: cptr
 
-    my_shape = tensor%torch_tensor_get_shape()
+    my_shape = tensor%get_shape()
 
     if (present(sizes)) then
       if (.not. all(my_shape == sizes)) then
@@ -1532,7 +1532,7 @@ contains
     integer(c_int), parameter :: dtype = torch_kInt16  !! Data type
     type(c_ptr) :: cptr
 
-    my_shape = tensor%torch_tensor_get_shape()
+    my_shape = tensor%get_shape()
 
     if (present(sizes)) then
       if (.not. all(my_shape == sizes)) then
@@ -1562,7 +1562,7 @@ contains
     integer(c_int), parameter :: dtype = torch_kInt16  !! Data type
     type(c_ptr) :: cptr
 
-    my_shape = tensor%torch_tensor_get_shape()
+    my_shape = tensor%get_shape()
 
     if (present(sizes)) then
       if (.not. all(my_shape == sizes)) then
@@ -1592,7 +1592,7 @@ contains
[…]
