### 2025-01-24 9160652
Implement get_device_type on C++ side; introduce get_ftorch_device
