+++ src/ctorch.cpp
@@ -40,7 +40,8 @@ constexpr auto get_dtype(torch_data_t dtype) {
   }
 }
 
-const auto get_device(torch_device_t device_type, int device_index) {
+// Mapping from FTorch device_type_t to libtorch DeviceType
+const auto get_libtorch_device(torch_device_t device_type, int device_index) {
   switch (device_type) {
   case torch_kCPU:
     if (device_index != -1) {
@@ -65,6 +66,20 @@ const auto get_device(torch_device_t device_type, int device_index) {
   }
 }
 
+// Mapping from libtorch DeviceType to FTorch device_type_t
+const torch_device_t get_ftorch_device(torch::DeviceType device_type) {
+  switch (device_type) {
+  case torch::kCPU:
+    return torch_kCPU;
+  case torch::kCUDA:
+    return torch_kCUDA;
+  default:
+    std::cerr << "[ERROR]: device type " << device_type << " not implemented in FTorch"
+              << std::endl;
+    exit(EXIT_FAILURE);
+  }
+}
+
 // =============================================================================
 // --- Functions for constructing tensors
 // =============================================================================
@@ -79,7 +94,7 @@ torch_tensor_t torch_empty(int ndim, const int64_t *shape, torch_data_t dtype,
     c10::IntArrayRef vshape(shape, ndim);
     tensor = new torch::Tensor;
     *tensor = torch::empty(vshape, torch::dtype(get_dtype(dtype)))
-                  .to(get_device(device_type, device_index));
+                  .to(get_libtorch_device(device_type, device_index));
   } catch (const torch::Error &e) {
     std::cerr << "[ERROR]: " << e.msg() << std::endl;
     delete tensor;
@@ -102,7 +117,7 @@ torch_tensor_t torch_zeros(int ndim, const int64_t *shape, torch_data_t dtype,
     c10::IntArrayRef vshape(shape, ndim);
     tensor = new torch::Tensor;
     *tensor = torch::zeros(vshape, torch::dtype(get_dtype(dtype)))
-                  .to(get_device(device_type, device_index));
+                  .to(get_libtorch_device(device_type, device_index));
   } catch (const torch::Error &e) {
     std::cerr << "[ERROR]: " << e.msg() << std::endl;
     delete tensor;
@@ -125,7 +140,7 @@ torch_tensor_t torch_ones(int ndim, const int64_t *shape, torch_data_t dtype,
     c10::IntArrayRef vshape(shape, ndim);
     tensor = new torch::Tensor;
     *tensor = torch::ones(vshape, torch::dtype(get_dtype(dtype)))
-                  .to(get_device(device_type, device_index));
+                  .to(get_libtorch_device(device_type, device_index));
   } catch (const torch::Error &e) {
     std::cerr << "[ERROR]: " << e.msg() << std::endl;
     delete tensor;
@@ -153,7 +168,7 @@ torch_tensor_t torch_from_blob(void *data, int ndim, const int64_t *shape,
     c10::IntArrayRef vstrides(strides, ndim);
     tensor = new torch::Tensor;
     *tensor = torch::from_blob(data, vshape, vstrides, torch::dtype(get_dtype(dtype)))
-                  .to(get_device(device_type, device_index));
+                  .to(get_libtorch_device(device_type, device_index));
 
   } catch (const torch::Error &e) {
     std::cerr << "[ERROR]: " << e.msg() << std::endl;
@@ -214,11 +229,6 @@ void torch_tensor_print(const torch_tensor_t tensor) {
   std::cout << *t << std::endl;
 }
 
-int torch_tensor_get_device_index(const torch_tensor_t tensor) {
-  auto t = reinterpret_cast<torch::Tensor *>(tensor);
[…]
