### 2025-01-24 c574e41
Implement get_dtype on C++ side; introduce get_ftorch_dtype
