+++ src/CMakeLists.txt
@@ -111,11 +111,11 @@ if(CMAKE_BUILD_TESTS)
 
   enable_testing()
 
-  # Unit tests
-  # NOTE: We do not currently support unit testing on Windows
-  if(UNIX)
-    add_subdirectory(test/unit)
-  endif()
+  # # Unit tests
+  # # NOTE: We do not currently support unit testing on Windows
+  # if(UNIX)
+  #   add_subdirectory(test/unit)
+  # endif()
 
   # Integration tests
   file(MAKE_DIRECTORY test/examples)
