### 2025-01-24 5c0f007
DO NOT MERGE: remove annoying compiler warnings
