### 2025-01-24 3946312
chore: add device type to resnet example
