### 2025-01-24 6034cb7
Convert example 3 to XPU
