+++ examples/CMakeLists.txt
@@ -1,9 +1,9 @@
 if(CMAKE_BUILD_TESTS)
   add_subdirectory(1_SimpleNet)
   add_subdirectory(2_ResNet18)
-  if(ENABLE_CUDA AND ENABLE_MPI)
+  # if(ENABLE_CUDA AND ENABLE_MPI)
     add_subdirectory(3_MultiGPU)
-  endif()
+  # endif()
   add_subdirectory(4_MultiIO)
   # add_subdirectory(5_Looping)
   add_subdirectory(6_Autograd)
