+++ .github/workflows/test_suite_windows.yml
@@ -3,25 +3,25 @@ name: TestSuiteWindows
 
 # Controls when the workflow will run
 on:
-  # Triggers the workflow on pushes to the "main" branch, i.e., PR merges
-  push:
-    branches: [ "main" ]
+  # # Triggers the workflow on pushes to the "main" branch, i.e., PR merges
+  # push:
+  #   branches: [ "main" ]
 
-  # Triggers the workflow on pushes to open pull requests with code changes
-  pull_request:
-    paths:
-      - '.github/workflows/test_suite_windows.yml'
-      - '**.bat'
-      - '**.c'
-      - '**.cpp'
-      - '**.fypp'
-      - '**.f90'
-      - '**.F90'
-      - '**.pf'
-      - '**.py'
-      - '**CMakeLists.txt'
-      - '**requirements.txt'
-      - '**data/*'
+  # # Triggers the workflow on pushes to open pull requests with code changes
+  # pull_request:
+  #   paths:
+  #     - '.github/workflows/test_suite_windows.yml'
+  #     - '**.bat'
+  #     - '**.c'
+  #     - '**.cpp'
+  #     - '**.fypp'
+  #     - '**.f90'
+  #     - '**.F90'
+  #     - '**.pf'
+  #     - '**.py'
+  #     - '**CMakeLists.txt'
+  #     - '**requirements.txt'
+  #     - '**data/*'
 
   # Allows you to run this workflow manually from the Actions tab
   workflow_dispatch:
