+++ examples/1_SimpleNet/pt2ts.py
@@ -103,8 +103,8 @@ if __name__ == "__main__":
     # trained_model = trained_model.to(device)
     # trained_model.eval()
     # trained_model_dummy_input = trained_model_dummy_input.to(device)
-    
-    device = torch.device('xpu')
+
+    device = torch.device("xpu")
     trained_model = trained_model.to(device)
     trained_model.eval()
     trained_model_dummy_input = trained_model_dummy_input.to(device)
