### 2025-01-24 c488f20
Require device_type for torch_model_load
