### 2025-01-27 7e52efe
Drop dtype/device type attributes
