### 2025-01-27 6fce43b
Inlining; formatting
