+++ src/ftorch.F90
@@ -27,7 +27,7 @@ module ftorch
 
   !> Type for holding a Torch tensor.
   type torch_tensor
-    type(c_ptr) :: p = c_null_ptr    !! pointer to the tensor in memory
+    type(c_ptr) :: p = c_null_ptr  !! pointer to the tensor in memory
   contains
     procedure :: get_rank => torch_tensor_get_rank
     procedure :: get_shape => torch_tensor_get_shape
@@ -2505,13 +2505,10 @@ contains
     integer(int8), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
     type(torch_tensor) :: wrk, output
-    integer :: device_type
-    integer :: device_index
 
     ! Create a tensor with a single entry, the scalar pre-multiplier
-    device_type = tensor%get_device_type()
-    device_index = tensor%get_device_index()
-    call torch_tensor_from_array(wrk, [scalar], [1], device_type, device_index)
+    call torch_tensor_from_array(wrk, [scalar], [1], tensor%get_device_type(), &
+                                 tensor%get_device_index())
     output%p = torch_tensor_multiply_c(wrk%p, tensor%p)
   end function torch_tensor_premultiply_int8
 
@@ -2522,13 +2519,10 @@ contains
     integer(int16), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
     type(torch_tensor) :: wrk, output
-    integer :: device_type
-    integer :: device_index
 
     ! Create a tensor with a single entry, the scalar pre-multiplier
-    device_type = tensor%get_device_type()
-    device_index = tensor%get_device_index()
-    call torch_tensor_from_array(wrk, [scalar], [1], device_type, device_index)
+    call torch_tensor_from_array(wrk, [scalar], [1], tensor%get_device_type(), &
+                                 tensor%get_device_index())
     output%p = torch_tensor_multiply_c(wrk%p, tensor%p)
   end function torch_tensor_premultiply_int16
 
@@ -2539,13 +2533,10 @@ contains
     integer(int32), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
     type(torch_tensor) :: wrk, output
-    integer :: device_type
-    integer :: device_index
 
     ! Create a tensor with a single entry, the scalar pre-multiplier
-    device_type = tensor%get_device_type()
-    device_index = tensor%get_device_index()
-    call torch_tensor_from_array(wrk, [scalar], [1], device_type, device_index)
+    call torch_tensor_from_array(wrk, [scalar], [1], tensor%get_device_type(), &
+                                 tensor%get_device_index())
     output%p = torch_tensor_multiply_c(wrk%p, tensor%p)
   end function torch_tensor_premultiply_int32
 
@@ -2556,13 +2547,10 @@ contains
     integer(int64), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
     type(torch_tensor) :: wrk, output
-    integer :: device_type
-    integer :: device_index
 
     ! Create a tensor with a single entry, the scalar pre-multiplier
-    device_type = tensor%get_device_type()
-    device_index = tensor%get_device_index()
-    call torch_tensor_from_array(wrk, [scalar], [1], device_type, device_index)
+    call torch_tensor_from_array(wrk, [scalar], [1], tensor%get_device_type(), &
+                                 tensor%get_device_index())
     output%p = torch_tensor_multiply_c(wrk%p, tensor%p)
   end function torch_tensor_premultiply_int64
 
@@ -2573,13 +2561,10 @@ contains
     real(real32), target, intent(in) :: scalar
     type(torch_tensor), intent(in) :: tensor
     type(torch_tensor) :: wrk, output
-    integer :: device_type
[…]
