### 2025-01-27 f5f05cd
Add example 5 to test build
