+++ src/CMakeLists.txt
@@ -129,8 +129,8 @@ if(CMAKE_BUILD_TESTS)
   #      DESTINATION ${CMAKE_CURRENT_SOURCE_DIR}/test/examples)
   file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/../examples/4_MultiIO
        DESTINATION ${CMAKE_CURRENT_SOURCE_DIR}/test/examples)
-  # file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/../examples/5_Looping
-  #      DESTINATION ${CMAKE_CURRENT_SOURCE_DIR}/test/examples)
+  file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/../examples/5_Looping
+       DESTINATION ${CMAKE_CURRENT_SOURCE_DIR}/test/examples)
   file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/../examples/6_Autograd
        DESTINATION ${CMAKE_CURRENT_SOURCE_DIR}/test/examples)
   add_subdirectory(test/examples)
