### 2025-01-27 b2c6c74
Update previous API changes; fix typos
