### 2025-01-27 06763d6
Drop ENABLE_MPI CMake argument
