### 2025-01-27 e0d4e45
Drop ENABLE_MPI CMake argument
