+++ examples/CMakeLists.txt
@@ -1,7 +1,7 @@
 if(CMAKE_BUILD_TESTS)
   add_subdirectory(1_SimpleNet)
   add_subdirectory(2_ResNet18)
-  if(ENABLE_CUDA AND ENABLE_MPI)
+  if(ENABLE_CUDA)
     add_subdirectory(3_MultiGPU)
   endif()
   add_subdirectory(4_MultiIO)
