### 2025-01-27 Cambridge-ICCS/FTorch#214 @jatkinson1000
As an update here I have been bashing my head against conda environments for a while now.
Still don't have a definitively working environment but found a few interesting things.

- It is tricky to get a CUDA build using the conda-forge implementation of PyTorch

To resolve this I went back to the install matrix on the [pytorch homepage](https://pytorch.org/) which suggests:

```
conda install pytorch torchvision torchaudio cpuonly -c pytorch
conda install pytorch torchvision torchaudio pytorch-cuda=11.8 -c pytorch -c nvidia
```
for CPU and CUDA respectively - i.e. installing from the `pytorch` and `nvidia` channels rather than `conda-forge`.

This works fine if you are only installing the above dependencies and running from python - @matthewfeickert note that installing from the `pytorch` channel will bring in `zlib` as a dependency, so its lack in your case seems to be an issue with it being missing from the `conda-forge` implementation.

However, once I start adding the other dependencies required for FTorch from `conda-forge` (fortran and cxx compilers etc.) some of the `pytorch` packages get overwritten leading to issues when trying to run things. For example FTorch will build but I hit MKL issues if trying to run cpu-only nets from Fortran:
```
Intel MKL ERROR: Parameter 13 was incorrect on entry to SGEMM
```
This occurs with the following packages installed:
```
channels:
  - conda-forge
dependencies:
  - cmake
  - cxx-compiler
  - fortran-compiler
  - openmpi
  - pytorch::cpuonly
  - pytorch::pytorch
  - pytorch::torchvision
```

I'm still looking at getting this to work, but it's proving to be a challenge.
