### 2025-01-27 6e36cb9
Remove MPI from example 3
