+++ examples/3_MultiGPU/README.md
@@ -6,22 +6,20 @@ multiple GPU devices.
 
 ## Description
 
-The same Python file `simplenet.py` is used from the earlier example. Recall that it
-defines a very simple PyTorch network that takes an input of length 5 and applies a
-single `Linear` layer to multiply it by 2.
+The same Python file `simplenet.py` is used from the earlier example. Recall
+that it defines a very simple PyTorch network that takes an input of length 5
+and applies a single `Linear` layer to multiply it by 2.
 
 The same `pt2ts.py` tool is used to save the simple network to TorchScript.
 
-A series of files `simplenet_infer_<LANG>` then bind from other languages to run the
-TorchScript model in inference mode.
+A series of files `multigpu_infer_<LANG>` then bind from other languages to run
+the TorchScript model in inference mode.
 
 ## Dependencies
 
 To run this example requires:
 
 - CMake
-- An MPI installation.
-- mpif90
 - FTorch (installed as described in main package)
 - Python 3
 
@@ -35,39 +33,39 @@ source venv/bin/activate
 pip install -r requirements.txt
 ```
 
-You can check that everything is working by running `simplenet.py`:
+You can check that everything is working by running `multigpu.py`:
 ```
-python3 simplenet.py
+python3 multigpu.py
 ```
 As before, this defines the network and runs it with an input tensor
-[0.0, 1.0, 2.0, 3.0, 4.0] to produce the result:
+[0.0, 1.0, 2.0, 3.0, 4.0]. The difference is that the code will make use of the
+default CUDA device (index 0) to produce the result:
 ```
+SimpleNet forward pass on CUDA device 0
 tensor([[0, 2, 4, 6, 8]])
 ```
 
-To save the SimpleNet model to TorchScript run the modified version of the `pt2ts.py`
-tool:
+To save the `SimpleNet` model to TorchScript run the modified version of the
+`pt2ts.py` tool:
 ```
 python3 pt2ts.py
 ```
-which will generate `saved_simplenet_model_cuda.pt` - the TorchScript instance of the
-network. The only difference with the earlier example is that the model is built to
-be run using CUDA rather than on CPU.
+which will generate `saved_multigpu_model_cuda.pt` - the TorchScript instance
+of the network. The only difference with the earlier example is that the model
+is built to be run using CUDA rather than on CPU.
 
-You can check that everything is working by running the `simplenet_infer_python.py`
-script. It's set up with MPI such that a different GPU device is associated with each
-MPI rank. You should substitute `<NP>` with the number of GPUs you wish to run with:
+You can check that everything is working by running the
+`multigpu_infer_python.py` script. It's set up such that it loops over two GPU
+devices. Run with:
 ```
-mpiexec -np <NP> python3 simplenet_infer_python.py
+python3 multigpu_infer_python.py
 ```
 This reads the model in from the TorchScript file and runs it with an different input
 tensor on each GPU device: [0.0, 1.0, 2.0, 3.0, 4.0], plus the device index in each
-entry. The result should be (some permutation of):
+entry. The result should be:
 ```
-0: tensor([[0., 2., 4., 6., 8.]])
-1: tensor([[ 2., 4.,  6.,  8., 10.]])
-2: tensor([[ 4., 6.,  8., 10., 12.]])
-3: tensor([[ 6., 8., 10., 12., 14.]])
+Output on device 0: tensor([[0., 2., 4., 6., 8.]])
+Output on device 1: tensor([[ 2., 4.,  6.,  8., 10.]])
 ```
 
[…]

[…]
