### 2025-01-27 7c15ea2
Remove MPI from example 3
