+++ examples/3_MultiGPU/multigpu_infer_fortran.f90
@@ -8,9 +8,6 @@ program inference
                       torch_tensor_from_array, torch_model_load, torch_model_forward, &
                       torch_delete
 
-   ! Import MPI
-   use mpi, only : mpi_init, mpi_finalize, mpi_comm_world, mpi_comm_rank
-
    implicit none
 
    ! Set precision for reals
@@ -29,51 +26,51 @@ program inference
    type(torch_tensor), dimension(1) :: in_tensors
    type(torch_tensor), dimension(1) :: out_tensors
 
-   ! MPI configuration
-   integer :: rank, ierr, i
-
-   call mpi_init(ierr)
-   call mpi_comm_rank(mpi_comm_world, rank, ierr)
+   integer, parameter :: num_devices = 2
+   integer :: device_index, i
 
    ! Get TorchScript model file as a command line argument
    num_args = command_argument_count()
    allocate(args(num_args))
    do ix = 1, num_args
-       call get_command_argument(ix,args(ix))
+      call get_command_argument(ix,args(ix))
    end do
 
-   ! Initialise data and print the values used on each MPI rank.
-   in_data = [(rank + i, i = 0, 4)]
-   write (6, 100) rank, in_data(:)
-   100 format("input on rank ", i1,": [", 4(f5.1,","), f5.1,"]")
+   do device_index = 0, num_devices-1
 
-   ! Create Torch input tensor from the above array and assign it to the first (and only)
-   ! element in the array of input tensors.
-   ! We use the torch_kCUDA device type with device index corresponding to the MPI rank.
-   call torch_tensor_from_array(in_tensors(1), in_data, tensor_layout, &
-                                torch_kCUDA, device_index=rank)
+      ! Initialise data and print the values used
+      in_data = [(device_index + i, i = 0, 4)]
+      write (6, 100) device_index, in_data(:)
+      100 format("input on device ", i1,": [", 4(f5.1,","), f5.1,"]")
 
-   ! Create Torch output tensor from the above array.
-   ! Here we use the torch_kCPU device type since the tensor is for output only
-   ! i.e. to be subsequently used by Fortran on CPU.
-   call torch_tensor_from_array(out_tensors(1), out_data, tensor_layout, torch_kCPU)
+      ! Create Torch input tensor from the above array and assign it to the first (and only)
+      ! element in the array of input tensors.
+      ! We use the torch_kCUDA device type with the given device index
+      call torch_tensor_from_array(in_tensors(1), in_data, tensor_layout, torch_kCUDA, &
+                                   device_index=device_index)
 
-   ! Load ML model. Ensure that the same device type and device index are used
-   ! as for the input data.
-   call torch_model_load(model, args(1), torch_kCUDA, device_index=rank)
+      ! Create Torch output tensor from the above array.
+      ! Here we use the torch_kCPU device type since the tensor is for output only
+      ! i.e. to be subsequently used by Fortran on CPU.
+      call torch_tensor_from_array(out_tensors(1), out_data, tensor_layout, torch_kCPU)
 
-   ! Infer
-   call torch_model_forward(model, in_tensors, out_tensors)
+      ! Load ML model. Ensure that the same device type and device index are used
+      ! as for the input data.
+      call torch_model_load(model, args(1), torch_kCUDA, device_index=device_index)
 
-   ! Print the values computed on each MPI rank.
-   write (6, 200) rank, out_data(:)
-   200 format("output on rank ", i1,": [", 4(f5.1,","), f5.1,"]")
+      ! Infer
+      call torch_model_forward(model, in_tensors, out_tensors)
 
-   ! Cleanup
-   call torch_delete(model)
-   call torch_delete(in_tensors)
[…]
