+++ examples/3_MultiGPU/multigpu_infer_python.py
@@ -1,19 +1,19 @@
 """Load saved MultiGPUNet to TorchScript and run inference example."""
 
 import torch
-from mpi4py import MPI
 
 
 def deploy(saved_model: str, device: str, batch_size: int = 1) -> torch.Tensor:
     """
-    Load TorchScript SimpleNet and run inference with example Tensor.
+    Load TorchScript MultiGPUNet and run inference with example Tensor.
 
     Parameters
     ----------
     saved_model : str
-        location of SimpleNet model saved to Torchscript
+        location of MultiGPUNet model saved to Torchscript
     device : str
-        Torch device to run model on, 'cpu' or 'cuda'
+        Torch device to run model on, 'cpu' or 'cuda'. May be followed by a colon and
+        then a device index, e.g., 'cuda:0' for the 0th CUDA device.
     batch_size : int
         batch size to run (default 1)
 
@@ -24,9 +24,6 @@ def deploy(saved_model: str, device: str, batch_size: int = 1) -> torch.Tensor:
     """
     input_tensor = torch.tensor([0.0, 1.0, 2.0, 3.0, 4.0]).repeat(batch_size, 1)
 
-    # Add the rank (device index) to each tensor to make them differ
-    input_tensor += MPI.COMM_WORLD.rank
-
     if device == "cpu":
         # Load saved TorchScript model
         model = torch.jit.load(saved_model)
@@ -34,6 +31,10 @@ def deploy(saved_model: str, device: str, batch_size: int = 1) -> torch.Tensor:
         output = model.forward(input_tensor)
 
     elif device.startswith("cuda"):
+
+        # Add the device index to each tensor to make them differ
+        input_tensor += int(device.split(":")[-1] or 0)
+
         # All previously saved modules, no matter their device, are first
         # loaded onto CPU, and then are moved to the devices they were saved
         # from, so we don't need to manually transfer the model to the GPU
@@ -53,11 +54,12 @@ def deploy(saved_model: str, device: str, batch_size: int = 1) -> torch.Tensor:
 if __name__ == "__main__":
     saved_model_file = "saved_multigpu_model_cuda.pt"
 
-    device_to_run = f"cuda:{MPI.COMM_WORLD.rank}"
+    for device_index in range(2):
+        device_to_run = f"cuda:{device_index}"
 
-    batch_size_to_run = 1
+        batch_size_to_run = 1
 
-    with torch.no_grad():
-        result = deploy(saved_model_file, device_to_run, batch_size_to_run)
+        with torch.no_grad():
+            result = deploy(saved_model_file, device_to_run, batch_size_to_run)
 
-    print(f"Output on device {device_to_run}: {result}")
+        print(f"Output on device {device_to_run}: {result}")
