+++ examples/3_MultiGPU/requirements.txt
@@ -1,2 +1,1 @@
-mpi4py
 torch
