### 2025-01-27 a2e49e9
Add notes on example 5
