+++ examples/CMakeLists.txt
@@ -5,7 +5,7 @@ if(CMAKE_BUILD_TESTS)
   add_subdirectory(4_MultiIO)
   # NOTE: Example 5_Looping is built here but not run as part of the integration
   #       test suite. This is because it demonstrates 'good' versus 'bad'
-  #       practice, as opposed to testing functionality.
+  #       practice, as opposed to functionality.
   add_subdirectory(5_Looping)
   add_subdirectory(6_Autograd)
 endif()
