### 2025-01-27 5ed699c
Avoid double negatives in the run tests shell script to make things clearer.
