### 2025-01-27 9dab4f4
Require `device_type` argument for `torch_model_load` (#254)
* Require device_type for torch_model_load
* Increment Autograd example number
* Add example 5 to test build
* Add section to API changes page
* Update previous API changes; fix typos
* Update example page to use new API
* Update example code in readme to specify device when loading model.

---------

Co-authored-by: Matt Archer <ma595@cam.ac.uk>
Co-authored-by: Jack Atkinson <jwa34@cam.ac.uk>
