### 2025-01-27 0348b95
Copy simplenet as example 7_MPI
