+++ examples/CMakeLists.txt
@@ -10,4 +10,5 @@ if(CMAKE_BUILD_TESTS)
   #       practice, as opposed to functionality.
   add_subdirectory(5_Looping)
   add_subdirectory(6_Autograd)
+  add_subdirectory(7_MPI)
 endif()
