### 2025-01-27 1338bb9
Copy simplenet as example 7_MPI
