+++ examples/CMakeLists.txt
@@ -7,4 +7,5 @@ if(CMAKE_BUILD_TESTS)
   add_subdirectory(4_MultiIO)
   # add_subdirectory(5_Looping)
   add_subdirectory(6_Autograd)
+  add_subdirectory(7_MPI)
 endif()
