### 2025-01-27 1190abb
Copy simplenet as example 7_MPI
