+++ examples/7_MPI/requirements.txt
@@ -1,0 +1,3 @@
+torch
+mpi4py
+numpy
