+++ examples/7_MPI/README.md
@@ -1,0 +1,3 @@
+# Example 7 - MPI
+
+TODO
