### 2025-01-27 31a3429
MPI versions of inference
