+++ examples/7_MPI/mpi_infer_python.py
@@ -4,6 +4,7 @@ import os
 import sys
 
 import torch
+from mpi4py import MPI
 
 
 def deploy(saved_model: str, device: str, batch_size: int = 1) -> torch.Tensor:
@@ -25,6 +26,9 @@ def deploy(saved_model: str, device: str, batch_size: int = 1) -> torch.Tensor:
         result of running inference on model with example Tensor input
     """
     input_tensor = torch.tensor([0.0, 1.0, 2.0, 3.0, 4.0]).repeat(batch_size, 1)
+
+    # Add the rank (device index) to each tensor to make them differ
+    input_tensor += MPI.COMM_WORLD.rank
 
     if device == "cpu":
         # Load saved TorchScript model
@@ -52,6 +56,7 @@ if __name__ == "__main__":
     filepath = os.path.dirname(__file__) if len(sys.argv) == 1 else sys.argv[1]
     saved_model_file = os.path.join(filepath, "saved_simplenet_model_cpu.pt")
 
+    rank = MPI.COMM_WORLD.rank
     device_to_run = "cpu"
 
     batch_size_to_run = 1
@@ -59,11 +64,15 @@ if __name__ == "__main__":
     with torch.no_grad():
         result = deploy(saved_model_file, device_to_run, batch_size_to_run)
 
-    print(result)
+    print(f"rank {rank}: result:\n{result}")
 
-    if not torch.allclose(result, torch.Tensor([0.0, 2.0, 4.0, 6.0, 8.0])):
+    expected = torch.Tensor([2 * (i+rank) for i in range(5)])
+    if not torch.allclose(result, expected):
         result_error = (
-            f"result:\n{result}\ndoes not match expected value:\n"
-            f"{torch.Tensor([0.0, 2.0, 4.0, 6.0, 8.0])}"
+            f"rank {rank}: result:\n{result}\ndoes not match expected value:\n"
+            f"{expected}"
         )
         raise ValueError(result_error)
+
+    if rank == 0:
+        print("MPI Python example ran successfully")
