### 2025-01-27 3f3e43b
Update paper to reflect specifying a device for torch models.
