### 2025-01-27 Cambridge-ICCS/FTorch#214 @matthewfeickert
@jatkinson1000 Note there's already a fully defined working environment in https://github.com/Cambridge-ICCS/FTorch/issues/214#issuecomment-2578883698.

> note that installing from the `pytorch` channel will bring in `zlib` as a dependency, so its lack in your case seems to be an issue with it being missing from the `conda-forge` implementation.

The `pytorch` channel is no longer maintained by PyTorch and so should be viewed as dead. PyTorch's view is that the conda-forge channel should be viewed now.
