### 2025-01-27 2fea062
Check MPI properly for Python case
