+++ examples/7_MPI/mpi_infer_python.py
@@ -56,23 +56,30 @@ if __name__ == "__main__":
     filepath = os.path.dirname(__file__) if len(sys.argv) == 1 else sys.argv[1]
     saved_model_file = os.path.join(filepath, "saved_simplenet_model_cpu.pt")
 
-    rank = MPI.COMM_WORLD.rank
+    comm = MPI.COMM_WORLD
+    rank = comm.rank
     device_to_run = "cpu"
 
     batch_size_to_run = 1
 
+    # Run inference on each rank
     with torch.no_grad():
         result = deploy(saved_model_file, device_to_run, batch_size_to_run)
-
     print(f"rank {rank}: result:\n{result}")
 
-    expected = torch.Tensor([2 * (i + rank) for i in range(5)])
-    if not torch.allclose(result, expected):
-        result_error = (
-            f"rank {rank}: result:\n{result}\ndoes not match expected value:\n"
-            f"{expected}"
-        )
-        raise ValueError(result_error)
+    # Gather the outputs onto rank 0
+    recvbuf = torch.empty([comm.size, 5], dtype=torch.float32) if rank == 0 else None
+    comm.Gather(result, recvbuf, root=0)
 
+    # Check that the correct values were attained
     if rank == 0:
+        for rank_chk, result_chk in enumerate(recvbuf):
+            expected = torch.Tensor([2 * (i + rank_chk) for i in range(5)])
+            if not torch.allclose(expected, result_chk):
+                result_error = (
+                    f"rank {rank_chk}: result:\n{result_chk}\n"
+                    f"does not match expected value:\n{expected}"
+                )
+                raise ValueError(result_error)
+
         print("MPI Python example ran successfully")
