### 2025-01-27 Cambridge-ICCS/FTorch#214 @jatkinson1000
Ah, that's really useful, thanks.
Is there somewhere that notes that, as the PyTorch home page is still pointing people to the `pytorch` channel, even for latest releases!!

> [@jatkinson1000](https://github.com/jatkinson1000) Note there's already a fully defined working environment in [#214 (comment)](https://github.com/Cambridge-ICCS/FTorch/issues/214#issuecomment-2578883698).
>
Yes, I was working from this at one point, but was then looking into the `pytorch` channel as I thought that was their advised approach based on the homepage and conda docs. I'll go back to working with this.
Still interesting that the `pytorch` channel pulled in `zlib` as a dependency whilst `conda-forge` seemingly does not.
