### 2025-01-27 Cambridge-ICCS/FTorch#214 @jatkinson1000
Ah, found https://github.com/pytorch/pytorch/issues/138506
