### 2025-01-27 8cfa14b
Check MPI properly in Fortran case
