### 2025-01-27 f818b48
NumPy not required for example 7
