+++ examples/7_MPI/requirements.txt
@@ -1,3 +1,2 @@
 torch
 mpi4py
-numpy
