### 2025-01-27 85100ca
Add checks for comm size
