### 2025-01-27 542ef3d
Add checks for comm size
