### 2025-01-27 4a838b8
Add checks for comm size
