### 2025-01-28 Cambridge-ICCS/FTorch#214 @matthewfeickert
> Ah, found [pytorch/pytorch#138506](https://github.com/pytorch/pytorch/issues/138506)

Yeah. That and a few other places.

Sorry for not being responsive today &mdash; there have been some pressing issues that have kept my attention away from GitHub.

I'll reiterate that the FTorch team should not feel obliged to provide conda support, but I can spend some time later this week to update the example `pixi` environment I provided and give a working `environment.yml` from

```
pixi project export conda-environment environment.yml
```

(c.f. https://github.com/Cambridge-ICCS/FTorch/issues/214#issuecomment-2588371575). From scratch on https://github.com/Cambridge-ICCS/FTorch/commit/9dab4f49f7cb02eb4d94810ab890036569026507 it seems that the environment I had working previously has an issue with Caffe2's CMake files not picking up CUDA. I would just diff the environments lock files but I had made that under `/tmp` and have since restarted my laptop.
