### 2025-01-28 ef07fca
Bugfix: Include missing CheckLanguage module in the CMakeLists of example 3.
