### 2025-01-29 Cambridge-ICCS/FTorch#214 @matthewfeickert
@jatkinson1000 Given what I debugged in https://github.com/conda-forge/cuda-feedstock/issues/59#issuecomment-2620910028, can you please try a build with either the `pixi.toml` there or for `conda` the following `environment.yml`

```yaml
name: default
channels:
- conda-forge
- nodefaults
dependencies:
- fortran-compiler >=1.5.2,<2
- cxx-compiler >=1.5.2,<2
- cuda-compiler >=12.6.3,<13
- cuda-libraries-dev >=12.6.3,<13
- cuda-nvtx-dev >=12.6.77,<13
- cmake >=3.31.2,<4
- pytorch >=2.5.1,<3
- zlib >=1.3.1,<2
- git >=2.47.1,<3
- torchvision >=0.20.1,<0.21
- mpi4py >=4.0.1,<5
- openmpi-mpifort >=5.0.6,<6
```

and then please use the following Cmake command in the activated environment

```
cmake \
    $CMAKE_ARGS \
    -DCMAKE_INSTALL_PREFIX=$CONDA_PREFIX \
    -DCMAKE_PREFIX_PATH=$CONDA_PREFIX \
    -DCMAKE_BUILD_TYPE=Release \
    -DPython_EXECUTABLE=$CONDA_PREFIX/bin/python \
    -DENABLE_CUDA=TRUE \
    -DCUDA_TOOLKIT_ROOT_DIR=$CONDA_PREFIX/targets/x86_64-linux \
    -Dnvtx3_dir=$CONDA_PREFIX/targets/x86_64-linux/include/nvtx3
    -DCMAKE_VERBOSE_MAKEFILE=TRUE \
    -S FTorch/src/ \
    -B build && \
cmake -LH build && \
cmake --build build --clean-first --parallel=10 && \
cmake --install build
```

to do an out-of-source build of FTorch?
