### 2025-01-29 dc790ae
Bugfix: mismatch in Fortran array character array lengths in example 2 causing the wrong resnet category to be printed.
