### 2025-01-29 2cd3d09
Bugfix: mismatch in Fortran array character array lengths in example 2 (#263)
* Bugfix: mismatch in Fortran array character array lengths in example 2 causing the wrong resnet category to be printed.

* Update filename strings to len=128 in example 2 for consistency.
