### 2025-01-29 b5d5def
Add CPU-only conda environment file and associated instructions.
