### 2025-01-29 c21040d
Build example 3 if CUDA enabled (#222)
* Build example 3 if CUDA enabled
* Put model on CUDA device in simplenet
* Run example 3 if it's been built
* Add missing imports for pt2ts
* More helpful output for simplenet_infer_python
* Fix numbering in CMakeLists for examples
* Renaming in MultiGPU example; set up unit testing
* Raise error if no CUDA in example 3
* Fix model filename passed to fortran
* Bugfix: Update example 3 README with correct names of files simplenet -> multigpu.
* Bugfix: explicitly move loaded net to required device in multigpu python
in example 3.
* Bugfix: Include missing
[…]
