### 2025-01-29 4b0fec6
Update conda CUDA instructions to correctly locate CUDA components based on the advice from @matthewfeickert.
