+++ conda/environment_cuda.yaml
@@ -7,12 +7,13 @@ dependencies:
   - fortran-compiler
   - cxx-compiler
   - cmake
-  - openmpi  # For pFUnit to run the test suite and example 3
+  - openmpi-mpifort  # For pFUnit to run the test suite and MultiGPU example
   - cuda-version 12.6  # Set version of cuda here
   - cuda-compiler
   - cuda-libraries-dev
   - cuda-nvtx-dev
   - python <3.13  # Restrict as PyTorch not functional in 3.13 at present
   - pytorch
-  - torchvision  # For example 2 only
+  - torchvision  # For ResNet example only
+  - mpi4py  # For MultiGPU example only
   - zlib  # missing dependency of pytorch
