### 2025-01-29 #7602 @timothyas
Here's a summary of my review thus far:
- The paper is excellent, and good to go. The authors fixed a typo that I pointed out and there was only one remaining minor suggestion I had in [this issue](https://github.com/Cambridge-ICCS/FTorch/issues/212) which is that I think it's worthwhile to mention fortran-tf-lib as a similar library in TF/Keras (suggested by @jatkinson1000). This is minor though, and I'm open to what the authors want to do. Just let me know and we can close that issue and check the final paper box.
- The code and documentation is overall excellent as well. I ran into some really hairy environment issues and opened [this issue](https://github.com/Cambridge-ICCS/FTorch/issues/214) which has become really active (thanks everyone!). I think that I ran into problems largely because of the machine I was using, coupled with the fact that I was using conda. I want to re-emphasize that I do not expect the authors to get conda working for the package unless they want to. I view this issue as "good enough" because I was able to get a CPU installation working for all the provided example cases. My current status is that I wanted to just retry the GPU installation based on comments from @matthewfeickert (et al) in that thread, which I could provide to the authors, and I'm happy with whatever you want to take with that feedback (put it in docs or leave it as a searchable GH issue, either way). I've been swamped (toddler had RSV so lost a week, lots of deadlines) so I haven't gotten to this yet, but will hopefully get to it this week or early next week.

TL;DR once the minor paper issue is addressed (either with a no or integrated changes) and I can give the authors some feedback on the GPU installation, I'm very happy to recommend this package for publication.
