### 2025-01-29 Cambridge-ICCS/FTorch#214 @jatkinson1000
RE conda.

Thank you @matthewfeickert!

I was working on this yesterday and got close, with my efforts in PR #264
I ended up hitting the issue you note above which is resolved by providing the additional CMake flags.

I can also confirm that the environment you provide works flawlessley as well.
I could not have got to where I did without your suggestions, so please let me know if you would like to be added as a co-author of the commit when it gets merged.

If you happen to have a moment to look at #264 it would be greatly appreciated.
I had a few open questions about best practices in conda for pinning dependency versions - I see you set the minimum to what is in your environment, and < the next major version? I also see that you did not specify the `nvidia` channel in your environment, even though it still seems to fetch from there?

@timothyas if yopu fancy checking the environments in #264 and letting us know how you get on I can move towards closing this.
