### 2025-01-29 Cambridge-ICCS/FTorch#214 @matthewfeickert
> I can also confirm that the environment you provide works flawlessley as well. I could not have got to where I did without your suggestions, so please let me know if you would like to be added as a co-author of the commit when it gets merged.

Very glad to hear it, and kudos to you and the rest of the team for putting in the extra work to provide additional support options for users. That's really fantastic to see overall. 🎉

I'm happy to be added as a co-author (but no pressure or worries if you forget).

> If you happen to have a moment to look at [#264](https://github.com/Cambridge-ICCS/FTorch/pull/264) it would be greatly appreciated.

Sure, I can look at it later tonight US time.

Some quick responses:

> I had a few open questions about best practices in conda for pinning dependency versions - I see you set the minimum to what is in your environment, and < the next major version?

That's more an artifact of my exporting the environment from `pixi` and wanting to give you something to play with that would hopefully not break right away due to any new releases. `pixi` by default when you add a dependency to your project (with `pixi add <dependency name>`) will set the latest version it can find as a lower bound and then cap it at the next major version. So when I created the environment PyTorch `v2.5.1` was the current release (PyTorch `v2.6.0` just got released on GitHub today) and so

```
pixi add pytorch
```

resulted in

```
pytorch = ">=2.5.1,<3"
```

`pixi` generally assumes that you're working on a project that is an instantiation or application and so defaults to adding upper bounds, but as FTorch is a library I would in general recommend trying to follow the advice that @henryiii lays out in his blog posts:
* [App vs. Library](https://iscinumpy.dev/post/app-vs-library/)
* [Should You Use Upper Bound Version Constraints?](https://iscinumpy.dev/post/bound-version-constraints/)

Now, obviously these blog posts are focused on **Python**, and FTorch is a Fortran library, but I think the general ideas presented transcend languages and I would encourage presenting environment definition files with lower bounds to users and letting them determine if they need to place restrictions on that environment to suit their own application needs. That being said, there may be times where you know that not having upper bounds will lead to broken environments given ABI issues and these can't be fixed with a [quick repodata patch PR](https://github.com/conda-forge/conda-forge-repodata-patches-feedstock) &mdash; and I wouldn't expect people who aren't conda-forge package maintainers to ever feel responsible for doing so anyway &mdash; and placing an upper bound is the right way to go.

Hopefully that's helpful, and not just annoyingly vague.

> I also see that you did not specify the `nvidia` channel in your environment, even though it still seems to fetch from there?

Correct, all the packages installed in the environment come from conda-forge and nowhere else! The NVIDIA team has really done some amazing work to make the CUDA ecosystem work really nicely on conda-forge without the need to specify any additional channels. So you can can get your entire CUDA toolchain (from the compiler to the runtime libraries) from just conda-forge. The [CUDA metapackage guides](https://github.com/conda-forge/cuda-feedstock/blob/main/recipe/README.md) are the right place to get more information and to ask questions just open issues in the `cuda` feedstock: https://github.com/conda-forge/cuda-feedstock.
