### 2025-01-30 f2b2e2d
Add note on fortran-tf-lib to paper with references, plus some small textual changes.
