+++ paper/paper.bib
@@ -1,3 +1,12 @@
+@software{Abadi_TensorFlow_Large-scale_machine_2015,
+  author = {Abadi, Martín and Agarwal, Ashish and Barham, Paul and Brevdo, Eugene and Chen, Zhifeng and Citro, Craig and Corrado, Greg S. and Davis, Andy and Dean, Jeffrey and Devin, Matthieu and Ghemawat, Sanjay and Goodfellow, Ian and Harp, Andrew and Irving, Geoffrey and Isard, Michael and Jozefowicz, Rafal and Jia, Yangqing and Kaiser, Lukasz and Kudlur, Manjunath and Levenberg, Josh and Mané, Dan and Schuster, Mike and Monga, Rajat and Moore, Sherry and Murray, Derek and Olah, Chris and Shlens, Jonathon and Steiner, Benoit and Sutskever, Ilya and Talwar, Kunal and Tucker, Paul and Vanhoucke, Vincent and Vasudevan, Vijay and Viégas, Fernanda and Vinyals, Oriol and Warden, Pete and Wattenberg, Martin and Wicke, Martin and Yu, Yuan and Zheng, Xiaoqiang},
+  doi = {10.5281/zenodo.4724125},
+  license = {Apache-2.0},
+  month = nov,
+  title = {{TensorFlow, Large-scale machine learning on heterogeneous systems}},
+  year = {2015}
+}
+
 @article{bishara2023state,
   title={A state-of-the-art review on machine learning-based multiscale modeling, simulation, homogenization and design of materials},
   author={Bishara, Dana and Xie, Yuxi and Liu, Wing Kam and Li, Shaofan},
@@ -28,6 +37,14 @@
   title    = {Fortitude},
   url      = {https://github.com/PlasmaFAIR/fortitude},
   year={2024},
+}
+
+@Online{fortran-tf-lib,
+  accessed = {2025-01-30},
+  author   = {Cambridge-ICCS},
+  title    = {fortran-tf-lib},
+  url      = {https://github.com/Cambridge-ICCS/fortran-tf-lib},
+  year={2023},
 }
 
 @Online{fypp,
