### 2025-01-30 783f582
Enforce the use of a virtual environment in the integration testing script.
