### 2025-01-30 5819c7c
Enforce the use of a virtual environment in the integration testing script.
