+++ pages/testing.md
@@ -39,10 +39,10 @@ See the subsections below for instructions on how to run subsets of the full
 test suite.
 
 > <sup>1</sup> _If you built FTorch against LibTorch (rather than creating a
-virtual environment) then either
+virtual environment) then you will need to
 [create a virtual environment](https://docs.python.org/3/library/venv.html) for
-the purposes of testing, or note that this script may have your Python
-environment install some modules._
+the purposes of integration testing as this script will install packages into your
+Python environment and will check that a virtual environment is in use._
 
 #### Running unit tests on Unix
 
