### 2025-01-30 7e20d71
Apply suggestions from code review by @matthewfeickert
Co-authored-by: Matthew Feickert <matthew.feickert@cern.ch>
