### 2025-01-30 3742327
Attempt to break example 3
