### 2025-01-30 aa79743
Mention GPU/CUDA as dependency
