### 2025-01-30 08d6da2
Add `get_dtype` and `get_device_type` methods for `torch_tensor` (#251)
* Add dtype and device_type attrs for torch_tensor; implement getters
* Rename get_<rank/shape> as torch_tensor_get_<rank/shape> for consistency
* Make torch_tensor_get_device_index a class method
* Add unit test for torch_tensor_get_device_type on CPU
* Add unit test for torch_tensor_get_device_type on CUDA device
* Add unit test for torch_tensor_get_dtype
* Make use of getters for device type and index
* Alias methods to be less verbose
* Implement get_device_type on C++ side; introduce get_ftorch_device
* Implement get_
[…]
