+++ src/ftorch.F90
@@ -29,8 +29,11 @@ module ftorch
   type torch_tensor
     type(c_ptr) :: p = c_null_ptr  !! pointer to the tensor in memory
   contains
-    procedure :: get_rank
-    procedure :: get_shape
+    procedure :: get_rank => torch_tensor_get_rank
+    procedure :: get_shape => torch_tensor_get_shape
+    procedure :: get_dtype => torch_tensor_get_dtype
+    procedure :: get_device_type => torch_tensor_get_device_type
+    procedure :: get_device_index => torch_tensor_get_device_index
   end type torch_tensor
 
   !| Enumerator for Torch data types
@@ -2273,27 +2276,8 @@ contains
     call torch_tensor_print_c(tensor%p)
   end subroutine torch_tensor_print
 
-  !> Determines the device index of a tensor.
-  function torch_tensor_get_device_index(tensor) result(device_index)
-    use, intrinsic :: iso_c_binding, only : c_int
-    type(torch_tensor), intent(in) :: tensor  !! Input tensor
-    integer(c_int) :: device_index            !! Device index of tensor
-
-    interface
-      function torch_tensor_get_device_index_c(tensor) result(device_index) &
-          bind(c, name = 'torch_tensor_get_device_index')
-        use, intrinsic :: iso_c_binding, only : c_int, c_ptr
-        implicit none
-        type(c_ptr), value, intent(in) :: tensor
-        integer(c_int) :: device_index
-      end function torch_tensor_get_device_index_c
-    end interface
-
-    device_index = torch_tensor_get_device_index_c(tensor%p)
-  end function torch_tensor_get_device_index
-
   !> Determines the rank of a tensor.
-  function get_rank(self) result(rank)
+  function torch_tensor_get_rank(self) result(rank)
     class(torch_tensor), intent(in) :: self
     integer(kind=int32) :: rank  !! rank of tensor
 
@@ -2308,10 +2292,10 @@ contains
     end interface
 
     rank = torch_tensor_get_rank_c(self%p)
-  end function get_rank
+  end function torch_tensor_get_rank
 
   !> Determines the shape of a tensor.
-  function get_shape(self) result(sizes)
+  function torch_tensor_get_shape(self) result(sizes)
     use, intrinsic :: iso_c_binding, only : c_f_pointer, c_int, c_long, c_long_long, c_ptr
     class(torch_tensor), intent(in) :: self
 #ifdef UNIX
@@ -2335,7 +2319,64 @@ contains
     ndims(1) = self%get_rank()
     cptr = torch_tensor_get_sizes_c(self%p)
     call c_f_pointer(cptr, sizes, ndims)
-  end function get_shape
+  end function torch_tensor_get_shape
+
+  !> Returns the data type of a tensor.
+  function torch_tensor_get_dtype(tensor) result(dtype)
+    use, intrinsic :: iso_c_binding, only : c_int
+    class(torch_tensor), intent(in) :: tensor  !! Input tensor
+    integer(c_int) :: dtype                !! Data type of tensor
+
+    interface
+      function torch_tensor_get_dtype_c(tensor) result(dtype) &
+          bind(c, name = 'torch_tensor_get_dtype')
+        use, intrinsic :: iso_c_binding, only : c_int, c_ptr
+        implicit none
+        type(c_ptr), value, intent(in) :: tensor
+        integer(c_int) :: dtype
+      end function torch_tensor_get_dtype_c
+    end interface
+
+    dtype = torch_tensor_get_dtype_c(tensor%p)
+  end function torch_tensor_get_dtype
+
+  !> Returns the device type of a tensor.
+  function torch_tensor_get_device_type(tensor) result(device_type)
[…]
