+++ src/ctorch.cpp
@@ -12,7 +12,8 @@
 // --- Constant expressions
 // =============================================================================
 
-constexpr auto get_dtype(torch_data_t dtype) {
+// Mapping from FTorch device_data_t to libtorch Dtype
+constexpr auto get_libtorch_dtype(torch_data_t dtype) {
   switch (dtype) {
   case torch_kUInt8:
     std::cerr << "[WARNING]: uint8 not supported in Fortran" << std::endl;
@@ -40,7 +41,37 @@ constexpr auto get_dtype(torch_data_t dtype) {
   }
 }
 
-const auto get_device(torch_device_t device_type, int device_index) {
+// Mapping from libtorch Dtype to FTorch device_data_t
+torch_data_t get_ftorch_dtype(caffe2::TypeMeta dtype) {
+  if (dtype == torch::kUInt8) {
+    std::cerr << "[WARNING]: uint8 not supported in Fortran" << std::endl;
+    // See https://gcc.gnu.org/onlinedocs/gfortran/ISO_005fFORTRAN_005fENV.html
+    exit(EXIT_FAILURE);
+  } else if (dtype == torch::kInt8) {
+    return torch_kInt8;
+  } else if (dtype == torch::kInt16) {
+    return torch_kInt16;
+  } else if (dtype == torch::kInt32) {
+    return torch_kInt32;
+  } else if (dtype == torch::kInt64) {
+    return torch_kInt64;
+  } else if (dtype == torch::kFloat16) {
+    std::cerr << "[WARNING]: float16 not supported in Fortran" << std::endl;
+    // See https://gcc.gnu.org/onlinedocs/gfortran/ISO_005fFORTRAN_005fENV.html
+    exit(EXIT_FAILURE);
+  } else if (dtype == torch::kFloat32) {
+    return torch_kFloat32;
+  } else if (dtype == torch::kFloat64) {
+    return torch_kFloat64;
+  } else {
+    std::cerr << "[ERROR]: data type " << dtype << " not supported in Fortran"
+              << std::endl;
+    exit(EXIT_FAILURE);
+  }
+}
+
+// Mapping from FTorch device_type_t to libtorch DeviceType
+const auto get_libtorch_device(torch_device_t device_type, int device_index) {
   switch (device_type) {
   case torch_kCPU:
     if (device_index != -1) {
@@ -65,6 +96,20 @@ const auto get_device(torch_device_t device_type, int device_index) {
   }
 }
 
+// Mapping from libtorch DeviceType to FTorch device_type_t
+const torch_device_t get_ftorch_device(torch::DeviceType device_type) {
+  switch (device_type) {
+  case torch::kCPU:
+    return torch_kCPU;
+  case torch::kCUDA:
+    return torch_kCUDA;
+  default:
+    std::cerr << "[ERROR]: device type " << device_type << " not implemented in FTorch"
+              << std::endl;
+    exit(EXIT_FAILURE);
+  }
+}
+
 // =============================================================================
 // --- Functions for constructing tensors
 // =============================================================================
@@ -78,8 +123,8 @@ torch_tensor_t torch_empty(int ndim, const int64_t *shape, torch_data_t dtype,
     // This doesn't throw if shape and dimensions are incompatible
     c10::IntArrayRef vshape(shape, ndim);
     tensor = new torch::Tensor;
-    *tensor = torch::empty(vshape, torch::dtype(get_dtype(dtype)))
-                  .to(get_device(device_type, device_index));
+    *tensor = torch::empty(vshape, torch::dtype(get_libtorch_dtype(dtype)))
+                  .to(get_libtorch_device(device_type, device_index));
   } catch (const torch::Error &e) {
[…]
