### 2025-01-30 88e836e
Don't require MPI for example 3
