### 2025-01-30 a1fc0e7
Errors for operators applied to tensors on different devices
