### 2025-01-30 4216eb4
Assigned tensor needs to be inout
