### 2025-01-30 5ab7f64
Assigned tensor needs to be inout
