### 2025-01-30 4043a65
Do not interrogate tensors that haven't been constructed
