### 2025-01-30 50125ff
Do not interrogate tensors that haven't been constructed
