+++ spack.yaml
@@ -12,3 +12,5 @@ spack:
     unify: true
   config:
     install_missing_compilers: true
+    install_tree: /opt/software
+  view: /opt/views/view
