+++ .github/workflows/docker.yml
@@ -23,7 +23,6 @@ concurrency:
 jobs:
   # This workflow contains a single job called "docker"
   docker:
-    name: Build Docker container
     # The type of runner that the job will run on
     runs-on: ubuntu-latest
     permissions:
