### 2025-01-30 ab08b99
Use Docker container in CI Ubuntu test workflow
