### 2025-01-30 Cambridge-ICCS/FTorch#214 @timothyas
Hey all, I just commented over in #264 but I'll repeat here that those conda build instructions worked flawlessly for me. As far as I'm concerned you guys have gone above and beyond to check the installation and reproducibility boxes. I think that this can be closed, especially since it pertains to my review of the JOSS paper, but I don't want to kill any good conversations.

@matthewfeickert @jatkinson1000 let me know if you'd like me to keep this open, and/or where any remaining discussion might go if this needs to close purely for the JOSS review.

Thanks again!
