### 2025-01-31 Cambridge-ICCS/FTorch#214 @matthewfeickert
> As far as I'm concerned you guys have gone above and beyond to check the installation and reproducibility boxes. I think that this can be closed, especially since it pertains to my review of the JOSS paper, but I don't want to kill any good conversations.

I have the same view, @timothyas. The FTorch team is putting in some incredible work here!
