### 2025-01-31 d01c20a
Rework strategy/matrix syntax
