+++ .github/workflows/test_suite_ubuntu.yml
@@ -37,11 +37,11 @@ jobs:
   test-suite-ubuntu:
     # The type of runner that the job will run on
     runs-on: ubuntu-latest
-    fail-fast: false
-    matrix:
-      std: ["f2008", "f2018"]
+    strategy:
+      fail-fast: false
       matrix:
         container: ["ghcr.io/cambridge-iccs/ftorch-dev-env:latest"]
+        std: ["f2008", "f2018"]
 
     # The Docker container that the job will use
     container:
