### 2025-01-31 2ead403
Activate Spack environment in CI tests
