+++ .github/workflows/test_suite_ubuntu.yml
@@ -76,16 +76,19 @@ jobs:
         run: |
           . /opt/spack-environment/activate.sh
           . ftorch/bin/activate
-          VN=$(python -c "import sys; print('.'.join(sys.version.split('.')[:2]))")
-          export Torch_DIR=${VIRTUAL_ENV}/lib/python${VN}/site-packages
-          export BUILD_DIR=$(pwd)/src/build
+          VN="$(python -c 'import sys; print(".".join(sys.version.split(".")[:2]))')"
+          PLATFORM="$(ls /opt/software | grep linux)"
+          COMPILER="$(ls /opt/software/${PLATFORM} | grep gcc)"
+          PYTORCH="$(ls /opt/software/${PLATFORM}/${COMPILER} | grep py-torch-)"
+          export Torch_DIR="/opt/software/${PLATFORM}/${COMPILER}/${PYTORCH}/lib/python${VN}/site-packages"
           export PFUNIT_DIR="/opt/view/$(ls /opt/view | grep PFUNIT)"
-          mkdir ${BUILD_DIR}
-          cd ${BUILD_DIR}
+          export BUILD_DIR="$(pwd)/src/build"
+          mkdir "${BUILD_DIR}"
+          cd "${BUILD_DIR}"
           cmake .. \
             -DPython_EXECUTABLE="$(which python)" \
             -DCMAKE_BUILD_TYPE=Release \
-            -DCMAKE_INSTALL_PREFIX=${BUILD_DIR} \
+            -DCMAKE_INSTALL_PREFIX="${BUILD_DIR}" \
             -DCMAKE_BUILD_TESTS=TRUE \
             -DCMAKE_PREFIX_PATH="${PFUNIT_DIR};${Torch_DIR}" \
             -DCMAKE_Fortran_FLAGS="-std=${{ matrix.std }}"
