### 2025-01-31 #7602 @timothyas
Alright everyone, both of my (minor) issues are closed up and so I think this paper, code, and docs are good to go 🎉🎉 great work on this package, I'm excited to see how it gets used.

Also, many thanks to the team for going above and beyond by adding a very streamlined conda installation. I really appreciate all the hard work you put in and I think it just makes your package that much more accessible. Great work!
