### 2025-01-31 1f46234
Take Docker image out of matrix
