### 2025-02-03 4317bfc
Switch to openmpi; attempt to fix Fortran case
