+++ .github/workflows/test_suite_ubuntu.yml
@@ -66,7 +66,7 @@ jobs:
       - name: Install an MPI distribution
         run: |
           sudo apt update
-          sudo apt install mpich
+          sudo apt install openmpi
 
       - name: Install pFUnit
         run: |
