### 2025-02-03 19f0930
Switch to openmpi; attempt to fix Fortran case
