### 2025-02-03 a3b1aef
Install all openmpi packages
