### 2025-02-03 36c4d10
Drop mkldnn; only specify gcc11; drop unused arg
