### 2025-02-03 a2f8c5f
Simplify Torch_DIR
