### 2025-02-03 4232a82
Worked example for tensor manipulation
