### 2025-02-04 6019976
Add information to README and docs regarding building with multiple jobs.
