### 2025-02-04 614f841
Add information to README and docs regarding building with multiple jobs.
