### 2025-02-04 247db32
Remove note about using make to build for CMake < 3.15 as we have now set the minimum CMake to be 3.15.
