### 2025-02-04 08ca98d
Remove note about using make to build for CMake < 3.15 as we have now set the minimum CMake to be 3.15.
