### 2025-02-04 b5bbebb
add note to conda about using parallel build in CMake.
