### 2025-02-04 5079892
Set lower bounds on versions for some conda packages and add note in the conda README.
