### 2025-02-04 431224b
Add note about conda to the main README and build docs page.
