### 2025-02-04 #7602 @jatkinson1000
@parikshitbajpai and @anand-me please do let me know if you have any questions or are having issues I can help with.

@matthewfeickert I know you are the editor here, but given your (much appreciated) involvement with the repo is there any possibility of you acting as one of the reviewers to keep things moving? Perfectly understand if not, but we are keen to get this through so our collaborators can cite JOSS rather than our GitHub/website in upcoming papers.
