### 2025-02-05 #7602 @matthewfeickert
> is there any possibility of you acting as one of the reviewers to keep things moving? Perfectly understand if not, but we are keen to get this through so our collaborators can cite JOSS rather than our GitHub/website in upcoming papers.

@jatkinson1000 We prefer to have at minimum two independent reviews from external reviewers that are not editors, unless there are extreme circumstances (e.g. all reviewers pulling out and being unable to find new ones). I appreciate the desire to have the review completed quickly, so I will try to complete as much of the post-review editor actions before @parikshitbajpai finishes their review.

One of those post-review things is going to involve a request that you create a full archive with a DOI for the release of the software the the review process culminates in. There are multiple services that can allow this, but I would recommend using [Zenodo](https://zenodo.org/), which has [GitHub integration](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content#issuing-a-persistent-identifier-for-your-repository-with-zenodo). If you don't already have a Zenodo account I would recommend creating one so that making a release with a DOI is an easy operation with no additional overhead.
