### 2025-02-05 f7b01c2
Typographical fixes and catching of infelicities in conda documentation by @jwallwork23
Co-authored-by: Joe Wallwork <22053413+jwallwork23@users.noreply.github.com>
