+++ conda/environment_cpu.yaml
@@ -8,5 +8,5 @@ dependencies:
   - cmake >=3.15
   - openmpi >=5.0.6  # For pFUnit
   - pytorch-cpu >=2.5.1  # Force cpu-only version
-  - torchvision  # For example 2
+  - torchvision  # For the ResNet example
   - zlib  # missing dependency for pytorch
