### 2025-02-05 #7602 @parikshitbajpai
@jatkinson1000, is there any traction on adding MPS support? I see https://github.com/Cambridge-ICCS/FTorch/issues/127 and has a linked PR (FTorch#125). Just wondering if there was any update. This doesn't affect the review but would be nice to have working. In the meantime, I will try the GPU features on one of our clusters.
Thanks!
