### 2025-02-06 #7602 @jatkinson1000
Hi @parikshitbajpai I believe MPS was working in [#125](https://github.com/Cambridge-ICCS/FTorch/pull/125), but was part of a larger issue to add multiple additional devices (intel XPU).
We discussed this last week in the team and decided, since XPU took a bit more work, we should probably separate these out.

However, we're currently experiencing issues with linking to libtorch binaries on Apple Silicon Macs (see open issue on the pytorch repo here: https://github.com/pytorch/pytorch/issues/143571).
Once these are resolved we'll look to re-test the MPS acceleration and merge.
