### 2025-02-06 ce892c3
Add conda environment files and instructions (#264)
* Add CPU-only conda environment file and associated instructions.
* Add cuda conda environment that allowed FTorch to be built and run, with tests passing as part of the integration suite script.
* Update conda CUDA instructions to correctly locate CUDA components based on the advice from @matthewfeickert.
* Set lower bounds on versions for some conda packages and add note in the conda README.
* Add note about conda to the main README and build docs page.

---------

Co-authored-by: Matthew Feickert <matthew.feickert@cern.ch>
Co-authored-by:
[…]
