### 2025-02-06 f58fa92
Make device enums consistent with PyTorch
