+++ src/ftorch.F90
@@ -52,11 +52,13 @@ module ftorch
 
   !| Enumerator for Torch devices
   !  From c_torch.h (torch_device_t)
+  !   NOTE: Numbered consistently with
+  !         https://github.com/pytorch/pytorch/blob/main/c10/core/DeviceType.h
   enum, bind(c)
     enumerator :: torch_kCPU = 0
     enumerator :: torch_kCUDA = 1
-    enumerator :: torch_kMPS = 2
-    enumerator :: torch_kXPU = 3
+    enumerator :: torch_kXPU = 12
+    enumerator :: torch_kMPS = 13
   end enum
 
   ! ============================================================================
