### 2025-02-06 #7602 @anand-me
Hi @jatkinson1000, I've extensively reviewed the code (past few days) and tested all examples - they are well-written and nicely documented! Everything works well, though I'd suggest adding '**requests**' as a dependency for Example 2 (ResNet-18) in requirements.txt. Overall, excellent work on making PyTorch-Fortran interop so seamless.
Thank you!
