### 2025-02-07 #7602 @matthewfeickert
Thanks for your review, @anand-me. (I know that you're busy finishing your Ph.D. and so I appreciate you finding time to squeeze this in too.)

> 5\. Suggested Documentation Additions:
>
> * Module load requirements for different systems
> * PATH and LD_LIBRARY_PATH handling
> * Best practices for HPC filesystem usage (project vs home space)
> * MPI compiler configuration guidance

Can you please open up an Issue on https://github.com/Cambridge-ICCS/FTorch/issues/ with `[JOSS Review]: <and then whatever the relevant title would be>` as the title with some more detailed thoughts on what the FTorch team should add? Please also link back here in the Issue (so add `https://github.com/openjournals/joss-reviews/issues/7602`).
