### 2025-02-07 Cambridge-ICCS/FTorch#277 @anand-me
HPC Deployment Documentation Suggestions
During JOSS review ([openjournals/joss-reviews#7602](https://github.com/openjournals/joss-reviews/issues/7602)), identified some documentation gaps for HPC deployment.

Suggested additions to documentation:
- Module load requirements for different (HPC) systems
- PATH and LD_LIBRARY_PATH handling
- Best practices for HPC filesystem usage (project vs home space)
- MPI compiler configuration guidance

Environment tested:
- CUDA 11.8
- OpenMPI
- GNU compiler (gfortran 8.5.0)
- Python 3.8

All examples were tested successfully on a ` HPC system` environment and I was particularly impressed by `Example 3 (MultiGPU)` which demonstrated seamless multi-device operation with correct rank-wise tensor operations across GPUs. These documentation additions would help streamline the `HPC deployment` process for future users.
