### 2025-02-09 #7602 @parikshitbajpai
Just finished my review. This is a great library and enables several applications especially in my field where a lot of the codes are in Fortran. I referenced a pending MPS pull request (thanks @jatkinson1000 for clarifying the status) in a previous comment but, apart from that, didn't have any concerns. I was able to follow the Conda instructions to install the library and the doco on executing tests and examples is well detailed. Overall, this is a great library and good to go. I would be using it for a couple of my applications for sure.

Thanks @jatkinson1000 and colleagues for the great work.

P.S. Just as I was writing about the detailed instructions, I noticed that I didn't see a code coverage badge. Having one would be great so people can glance quickly what percentage of code is tested.
