### 2025-02-10 #7602 @matthewfeickert
Thanks @timothyas, @parikshitbajpai, and @anand-me for your reviews!

@anand-me, given your review and Issue https://github.com/Cambridge-ICCS/FTorch/issues/277, as FTorch does make multiple references to HPC environments in their paper, do you view the things you've highlighted in the issue as important enough that they should be resolved prior to publication?
