### 2025-02-10 Cambridge-ICCS/FTorch#277 @jatkinson1000
Thanks @anand-me

Note that we have some information about setting compiler/linking flags here: https://cambridge-iccs.github.io/FTorch/page/cmake.html#building-other-projects-with-make

However, I have opened #278 move this to a specific HPC page with a bit more info.
Given this is our main userbase some extra info around this is a good suggestion, thank you.
