### 2025-02-10 #7602 @jatkinson1000
> P.S. Just as I was writing about the detailed instructions, I noticed that I didn't see a code coverage badge. Having one would be great so people can glance quickly what percentage of code is tested.

Thanks @parikshitbajpai for your kind words.

I have done this with PyTest and Codecov before, but not Fortran/C++
We have had a quick look, but if you have any suggestions for a workflow that can interact with Fortran/pFUnit to generate a coverage report that would be really useful.
