### 2025-02-10 #7602 @matthewfeickert
@jatkinson1000 I'll aim to give feedback on the paper in the next 48 hours and will open up an Issue for it. As a heads up, [the paper is too long](https://joss.readthedocs.io/en/latest/paper.html) (something I doubt you're used to hearing in a review. 🙂) but I'll give recommendations on how to shorten it.
