+++ paper/paper.md
@@ -48,16 +48,16 @@ bibliography: paper.bib
 
 In the last decade, machine learning (ML) and deep learning (DL) techniques have
 revolutionised many fields within science, industry, and beyond.
-Researchers across many domains are increasingly seeking to combine ML approaches
-with traditional numerical modelling to advance research.
-This typically brings about a computational challenge of _programming
+Researchers across domains are increasingly seeking to combine ML
+with numerical modelling to advance research.
+This typically brings about the challenge of _programming
 language interoperation_. PyTorch is a popular framework for designing and
 training ML/DL models whilst Fortran remains a language of choice for many
-scientific models running on high-performance computing (HPC) systems.
+high-performance computing (HPC) scientific models.
 The `FTorch` library provides an easy-to-use, performant method for coupling
 the two, allowing users to call PyTorch models from Fortran.
 
-`FTorch` is open-source, open-development, and well documented with minimal dependencies.
+`FTorch` is open-source, open-development, and well-documented with minimal dependencies.
 A central tenet of its design, in contrast to other approaches, is
 that FTorch removes dependence on the Python runtime (and virtual environments).
 By building on the `LibTorch` backend it allows users to run ML models on both
@@ -102,35 +102,35 @@ It removes the need for detailed knowledge about language interoperation
 and the need to develop bespoke coupling code, instead providing a Fortran interface
 designed to be familiar to both PyTorch and Fortran users.
 
-Further, `FTorch` has no dependence on the Python runtime.
-This is appealing for deployment on HPC systems where the management of
-Python environments can be challenging, and it avoids the
-incurred overhead of object representations in Python's runtime.
+Having no dependence on the Python runtime, `FTorch` avoids the
+incurred overhead of object representations in Python's runtime and
+is appealing for HPC systems where the management of
+Python environments can be challenging.
 
 
 # Software description
 
-PyTorch is itself built on an underlying framework `LibTorch` written in `C++` which can be obtained
-as a separate library and is accessible through a `C++` API.
+PyTorch itself builds on an underlying `C++` framework `LibTorch` which can be obtained
+as a separate library accessible through a `C++` API.
 By accessing this directly (rather than via PyTorch), `FTorch` avoids the use of Python at run-time.
 
 Using the `iso_c_binding` module, intrinsic to Fortran since the 2003 standard,
 we provide a Fortran wrapper to `LibTorch`.
 This enables shared memory use (where possible) to
-further improve efficiencies by reducing data-transfer during coupling.^[i.e. the same
+maximise efficiency by reducing data-transfer during coupling.^[i.e. the same
 data in memory is used by both `LibTorch` and `Fortran` without creating a copy.]
 
-`FTorch` is [open source on GitHub](https://github.com/Cambridge-ICCS/FTorch).
+`FTorch` is [open source](https://github.com/Cambridge-ICCS/FTorch).
 It can be built from source using CMake.
-Minimum dependencies are an installation of `LibTorch`, CMake,
+Minimum dependencies are `LibTorch`, CMake,
 and Fortran (2008 standard), `C`, and `C++` (`C++17` standard) compilers.^[To utilise GPU devices, users require the appropriate `LibTorch` binary plus any relevant dependencies, e.g. CUDA for NVIDIA devices.]
-The library is primarily developed in Linux, but also runs on macOS (arm64) and Windows.
+The library is primarily developed in Linux, but also runs on macOS and Windows.
 
 ## Key components and workflow leveraging FTorch
 
 #. Build, train, and validate a model in PyTorch.
-#. Save the model in TorchScript form, a strongly-typed subset of Python.
[…]
