### 2025-02-11 #7602 @jatkinson1000
Thanks @matthewfeickert
I was aware we were on the longer side and we'd already had a couple of passes to cut down as much as we (felt we) could before submission 😅.
Your advice as a separate set of eyes would be greatly appreciated

Hopefully there is some flexibility as I am aware of other papers who have exceeded the 1000 limit (e.g. #6079 (1569) and #7571 (1477))

I have just made [a commit](https://github.com/Cambridge-ICCS/FTorch/pull/200/commits/09ac01e48844c3a3f8f50f367b83454ce41c2643) that reduces the count by 110 to bring us down towards this level.
