### 2025-02-20 Cambridge-ICCS/FTorch#290 @matthewfeickert
[JOSS Review] Paper revision suggestions
These are questions and revisions for JOSS submission https://github.com/openjournals/joss-reviews/issues/7602.

[As mentioned](https://github.com/openjournals/joss-reviews/issues/7602#issuecomment-2649162987), these suggested revisions remove large sections to conform to [the JOSS paper format](https://joss.readthedocs.io/en/latest/paper.html#joss-paper-format). I appreciate that a lot of work and time went into writing this paper, so I want to be clear that the revisions are not supposed to be meant as dismissive or invalidating of the relevance of the information. Any material that is recommended to be removed should be migrated to the documentation (though migrating it to the docs not need to happen prior to acceptance.)

* https://github.com/Cambridge-ICCS/FTorch/blob/96a419ec7480f44c6d748406351a28306f581b65/paper/paper.md#L13-L14

Just to double check, Athena Elafrou doesn't have an ORCID, correct?

* https://github.com/Cambridge-ICCS/FTorch/blob/96a419ec7480f44c6d748406351a28306f581b65/paper/paper.md?plain=1#L54

Can you cite PyTorch when it is first mentioned in the Summary instead of later in the Statement of need?

* https://github.com/Cambridge-ICCS/FTorch/blob/96a419ec7480f44c6d748406351a28306f581b65/paper/paper.md?plain=1#L79

An emdash (&mdash;) should be used here, not a dash (-). In Markdown an emdash is represented by `&mdash;`.

* While I believe you RE:

https://github.com/Cambridge-ICCS/FTorch/blob/96a419ec7480f44c6d748406351a28306f581b65/paper/paper.md#L78-L79

is there a citation you can provide here? While I appreciate rigor, I also don't think that for a JOSS paper you need to motivate the importance of ML, so if you instead wanted to revise down this paragraph significantly that would be okay too.

* https://github.com/Cambridge-ICCS/FTorch/blob/96a419ec7480f44c6d748406351a28306f581b65/paper/paper.md?plain=1#L88

"Ideally users" should be "Ideally, users"

* https://github.com/Cambridge-ICCS/FTorch/blob/96a419ec7480f44c6d748406351a28306f581b65/paper/paper.md?plain=1#L92

Should be an emdash `&mdash;` not an endash `&ndash;` (assuming you were going for `LaTeX's `--` endash syntax.)

* Most of the "Software description" section [is not needed](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain), and should be moved to the documentation, if it is not there already. I would recommend that all of the "Software description" section be removed except for the following segments (revised as you see fit):

   - https://github.com/Cambridge-ICCS/FTorch/blob/96a419ec7480f44c6d748406351a28306f581b65/paper/paper.md#L116-L120

   - https://github.com/Cambridge-ICCS/FTorch/blob/96a419ec7480f44c6d748406351a28306f581b65/paper/paper.md#L145-L148

* All of "Examples and Tooling"

https://github.com/Cambridge-ICCS/FTorch/blob/96a419ec7480f44c6d748406351a28306f581b65/paper/paper.md?plain=1#L185

should be removed, with the exeception of

https://github.com/Cambridge-ICCS/FTorch/blob/96a419ec7480f44c6d748406351a28306f581b65/paper/paper.md#L201-L202

which can be integrated into the previous section.

* https://github.com/Cambridge-ICCS/FTorch/blob/96a419ec7480f44c6d748406351a28306f581b65/paper/paper.md?plain=1#L220

Does [Fiats](https://berkeleylab.github.io/fiats) have a citation you can use?

* Please condense down

https://github.com/Cambridge-ICCS/FTorch/blob/96a419ec7480f44c6d748406351a28306f581b65/paper/paper.md#L256-L259

and

https://github.com/Cambridge-ICCS/FTorch/blob/96a419ec7480f44c6d748406351a28306f581b65/paper/paper.md#L265-L267

into a single entry. While the uses are different, this section is already too long as ongoing research projects using the software should be cited briefly and do not need full descriptions.

* Do ICON

https://github.com/Cambridge-ICCS/FTorch/blob/96a419ec7480f44c6d748406351a28306f581b65/paper/paper.md?plain=1#L261

and CESM

https://github.com/Cambridge-ICCS/FTorch/blob/96a419ec7480f44c6d748406351a28306f581b65/paper/paper.md?plain=1#L269

have citable references? If not, please cite their websites.
