### 2025-02-20 Cambridge-ICCS/FTorch#290 @jatkinson1000
> Just to double check, Athena Elafrou doesn't have an ORCID, correct?

Correct

> Can you cite PyTorch when it is first mentioned in the Summary instead of later in the Statement of need?

- [ ] Done in commit https://github.com/Cambridge-ICCS/FTorch/pull/200/commits/0f7e1152fa9239183104f72a154a845237249bb9

> An emdash (—) should be used here, not a dash (-). In Markdown an emdash is represented by `&mdash;`.
> Should be an emdash `&mdash;` not an endash `&ndash;` (assuming you were going for `LaTeX's `--` endash syntax.)

- [ ] Done in commit https://github.com/Cambridge-ICCS/FTorch/pull/200/commits/0f7e1152fa9239183104f72a154a845237249bb9

> "Ideally users" should be "Ideally, users"

- [ ] Done in commit https://github.com/Cambridge-ICCS/FTorch/pull/200/commits/0f7e1152fa9239183104f72a154a845237249bb9

> is there a citation you can provide here? While I appreciate rigor, I also don't think that for a JOSS paper you need to motivate the importance of ML, so if you instead wanted to revise down this paragraph significantly that would be okay too.

- [ ] TODO

>     * Most of the "Software description" section [is not needed](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain), and should be moved to the documentation, if it is not there already. I would recommend that all of the "Software description" section be removed except for the following segments (revised as you see fit):

- [ ] TODO

>  Examples and Tooling should be removed, with the exeception of `pt2ts` which can be integrated into the previous section.

- [ ] TODO

> Does [Fiats](https://berkeleylab.github.io/fiats) have a citation you can use?

- [ ] I will look. TODO

>     * Please condense down DataWave

- [ ] TODO

>     * Do ICON and CESM have citable references? If not, please cite their websites.

They are large operational models (German weather service, and contributor to CMIP used globally by researchers) so whilst there are many papers there is no single citation for the 'model' as a whole as far as I am aware. I will add websites.

- [ ] TODO
