### 2025-02-21 Cambridge-ICCS/FTorch#290 @matthewfeickert
Thanks @jatkinson1000 (and @jwallwork23 from the commit record). We really want the paper to [not include additional material that can live in the documentation](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) though

> Given this format, a "full length" paper is not permitted, and software documentation such as API (Application Programming Interface) functionality should not be in the paper and instead should be outlined in the software documentation.

The paper is supposed to serve as a minimal citable reference for the software. So

https://github.com/Cambridge-ICCS/FTorch/blob/aeadefd6a7c282d9ef628581f1adf66c52961b87/paper/paper.md#L112-L167

should get revised down to something like

```markdown
# Software description

`FTorch` is a Fortran wrapper to the `LibTorch` C++ framework using the `iso_c_binding`
module, intrinsic to Fortran since the 2003 standard
This enables shared memory use (where possible) to
maximise efficiency by reducing data-transfer during coupling^[i.e. the same
data in memory is used by both `LibTorch` and Fortran without creating a copy.]
and avoids any use of Python at runtime.
PyTorch types are represented through derived types in `FTorch`, with Tensors supported
across a range of data types and ranks using the `fypp` preprocessor [@fypp].

We utilise the existing support in `LibTorch` for
GPU acceleration without additional device-specific code.
`torch_tensor`s are targeted to a device through a
`device_type` enum, currently supporting CPU, CUDA, XPU, and MPS.
Multiple GPUs may be targeted through the optional `device_index` argument.

Typically, users train a model in PyTorch and save it as TorchScript, a strongly-typed
subset of Python.
`FTorch` provides a utility script (`pt2ts.py`) to assist users with this process.
The Torchscript model is then loaded by `FTorch` and run using `LibTorch`.

Full details and additional information and resources are provided in the documentation available at
[https://cambridge-iccs.github.io/FTorch](https://cambridge-iccs.github.io/FTorch).
```
