+++ paper/paper.bib
@@ -31,14 +31,6 @@
   doi={10.1029/2022GL098174}
 }
 
-@Online{fortitude,
-  accessed = {2024-11-13},
-  author   = {Pattinson, Liam and Hill, Peter},
-  title    = {Fortitude},
-  url      = {https://github.com/PlasmaFAIR/fortitude},
-  year={2024},
-}
-
 @Online{fiats,
   accessed = {2024-11-13},
   author   = {Rouson, Damien and Rasmussen, Katherine},
@@ -149,13 +141,6 @@
   year={2015},
   publisher={Nature Publishing Group UK London},
   doi={10.1038/ngeo2398}
-}
-
-@Online{CAMML,
-  accessed = {2024-03-25},
-  author   = {{M2LInES}},
-  title    = {CAM-ML},
-  url      = {https://github.com/m2lines/CAM-ML},
 }
 
 @Online{CAMGW,
