### 2025-02-25 Cambridge-ICCS/FTorch#290 @jatkinson1000
Hi @matthewfeickert thank you for your guidance.
I have removed the listing and references to linting tools and test suites in https://github.com/Cambridge-ICCS/FTorch/pull/200/commits/3603437af450fd37a02559741ac1d27e42d63ade - let me know if this reads better.

As a side note I often use papers (including JOSS) to point me to interesting things and see what others are doing in the field.
As such I was keen to highlight what I see as underutilised or emerging pieces of software in our paper, in particular fypp and the recent [fortitude](https://github.com/PlasmaFAIR/fortitude) linter. I accept that they can be found in our source/documentation, but (opinion) it would be good if the scope of JOSS expanded slightly to cite software upon which we build. We cite PyTorch and the 'other approaches' in our paper, but it would be nice to see developers of key tooling we use also able to recieve recognition for their work in "academic currency".

The code listing I accept we should not have included 😉 apologies for that.
