### 2025-02-26 Cambridge-ICCS/FTorch#290 @matthewfeickert
> I have removed the listing and references to linting tools and test suites in [3603437](https://github.com/Cambridge-ICCS/FTorch/commit/3603437af450fd37a02559741ac1d27e42d63ade) - let me know if this reads better.

I think that looks good @jatkinson1000. 👍

> As a side note I often use papers (including JOSS) to point me to interesting things and see what others are doing in the field.
> As such I was keen to highlight what I see as underutilised or emerging pieces of software in our paper, in particular fypp and the recent [fortitude](https://github.com/PlasmaFAIR/fortitude) linter. I accept that they can be found in our source/documentation, but (opinion) it would be good if the scope of JOSS expanded slightly to cite software upon which we build. We cite PyTorch and the 'other approaches' in our paper, but it would be nice to see developers of key tooling we use also able to receive recognition for their work in "academic currency".

Thanks very much for this comment. I've raised it with the JOSS Editorial Board and it has resulted in enough discussion that it will be discussed further at our next JOSS Editors meeting.

For this paper though, we'd still suggest removing the explicit suggestion of testing and code quality tools (as you already did in https://github.com/Cambridge-ICCS/FTorch/commit/3603437af450fd37a02559741ac1d27e42d63ade), but if it is important to you all as authors to be able to provide citations for them you're welcome to add a generic sentence that makes reference to them as a collective and then add all the citations you'd like at the end of it. (Something like

```
`FTorch` uses testing and code quality validation during development [@insert all your citations here].
```
)

This is of course totally optional. I'll wait for your decision and then move forward with the final approval recommendation steps in https://github.com/openjournals/joss-reviews/issues/7602#issuecomment-2649148914.
