### 2025-02-26 Cambridge-ICCS/FTorch#290 @danielskatz
As track editor, I support @matthewfeickert's recommendation for this paper. The current paper is a reasonable length with reasonable content, and adding much more content wouldn't be in-line with JOSS's current paper length guidance, in my opinion, but adding one more sentence with references to tools would be fine.
