### 2025-02-26 Cambridge-ICCS/FTorch#290 @jatkinson1000
Thanks both. On your advice I will leave as it is, but thank you @matthewfeickert for taking this and raising it.
I hope that it generates interesting discussion 😃
