### 2025-02-26 #7602 @matthewfeickert
@jatkinson1000 all that's left to do now before recommending publication is to ensure that there is a long term public archive of the code that was reviewed. In this case there should also be a tag that corresponds to the version of the code reviewed. Before you do that, read on RE: archiving:

We'd suggest depositing the code either with [Zenodo](https://zenodo.org/) or with [figshare](https://figshare.com/) to get an archive with a DOI. If you use Zenodo there is [an (optional) GitHub integration](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) that can create a Zenodo archive for you anytime you make a GitHub release of your code.

[![Zenodo_DOI_guide](https://user-images.githubusercontent.com/5142394/234173148-2226325d-0a42-4526-b546-75b469d26c4c.png)](https://zenodo.org/account/settings/github/)

Once you make a release of `FTorch` with a DOI for the archive, and have verified all related author tasks listed in https://github.com/openjournals/joss-reviews/issues/7602#issuecomment-2649148914, please report the version tag and the DOI here and I'll have @editorialbot update the version listed in the review and add the archive.

Let me know if you have any questions. :+1:
