### 2025-02-28 #7602 @matthewfeickert
> Once you make a release of `FTorch` with a DOI for the archive,

@jatkinson1000 to be clear, it doesn't have to be the "v1.0.0" release if you're not ready for that yet. It can be any release &mdash; there just needs to be a release and archive associated with the reviewed code.
