### 2025-03-04 #7602 @jatkinson1000
Hi @matthewfeickert I had some leave at the end of last week but back now.
I have created an archive at https://doi.org/10.5281/zenodo.14968153 with DOI 10.5281/zenodo.14968153 and ensured that the author list is consistent.
The version is 1.0.0

It seems only you can set these with editorialbot but hopefully all the details are now there.

Let me know if you need me to do anything else!

Thank you for your efforts as editor and to all our reviewers for their input.
