### 2025-03-05 #7602 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
