### 2025-03-05 #7602 @jatkinson1000
> can you please update the Zenodo metadata so that the title for https://zenodo.org/records/14968153 is "FTorch: a library for coupling PyTorch models to Fortran" instead of "FTorch: v1.0.0"? We just want to make sure these match for the release that corresponds to the paper reviewed code. After that using the default SemVer is fine. 👍

@matthewfeickert Done! - https://zenodo.org/records/14968153 🚀
