### 2025-03-05 #7602 @danielskatz
@jatkinson1000 - the deposit and metadata look good to me. As track editor, I'll proofread this submission next and let you know if you need to do anything else.
