### 2025-03-05 #7602 @danielskatz
@jatkinson1000 - I've suggested a few minor changes in the paper in https://github.com/Cambridge-ICCS/FTorch/pull/308. Please merge this, or let me know what you disagree with, then we can proceed.
