### 2025-03-05 #7602 @jatkinson1000
Thanks @danielskatz I have merged that to `main`, and cherry-picked the merge to the `JOSS` branch to re-render the paper here.

The typographical changes won't appear in the zenodo archive, but hopefully this is not an issue as they are minor and not affecting the software itself.
