### 2025-03-05 #7602 @danielskatz
Congratulations to @jatkinson1000 (Jack Atkinson) and co-authors on your publication!!
