### 2025-03-05 #7602 @danielskatz
> Thanks [@danielskatz](https://github.com/danielskatz) and everyone else who was involved in review.
>
> On the JOSS website I can see we appear as 'in' C++: https://joss.theoj.org/papers/10.21105/joss.07602 I'm not sure how this is set, but could we instead appear primarily 'in' Fortran (and then Python and C++)?

@arfon & @xuanxu - do you know how this works? Can we change it?
