### 2025-03-06 #7602 @matthewfeickert
> We're enjoying paper day

Very well deserved! This was a great JOSS submission, so thank you for it.
